Module cargo::core::compiler

Modules

rustdoc
Utilities for building with rustdoc.

standard_lib
Code for building the standard library.

unit_dependencies
Constructs the dependency graph for compilation.

unit_graph

Structs

BuildConfig
Configuration information for a rustc build.

BuildContext
The build context, containing all information about a build task.

BuildOutput
Contains the parsed output of a custom build script.

BuildScriptOutputs
Map of packages to build script output.

BuildScripts
Linking information for a Unit.

Compilation
A structure returning the result of a compilation.

CompileTarget
Abstraction for the representation of a compilation target that Cargo has.

Context
Collection of all the stuff that is needed to perform a build.

DefaultExecutor
A DefaultExecutor calls rustc without doing anything else. It is Cargo's default behaviour.

Doctest
Structure with enough information to run rustdoc --test.

FileType
Type of each file generated by a Unit.

Metadata
The Metadata is a hash used to make unique file names for each unit in a build. It is also use for symbol mangling.

RustcTargetData
Collection of information about rustc and the host and target.

TargetInfo
Information about the platform target gleaned from querying rustc.

Unit
All information needed to define a unit.

UnitInterner
A small structure used to "intern" Unit values.

Enums

CompileKind
Indicator for how a unit is being compiled.

CompileMode
The general "mode" for what to do. This is used for two purposes. The commands themselves pass this in to compile_ws to tell it the general execution strategy. This influences the default targets selected. The other use is in the Unit struct to indicate what is being done with a specific target.

CrateType

FileFlavor
Kind of each file generated by a Unit, part of FileType.

Freshness
Indication of the freshness of a package.

MessageFormat

Traits

Executor
A glorified callback for executing calls to rustc. Rather than calling rustc directly, we'll use an Executor, giving clients an opportunity to intercept the build calls.

Functions

extern_args
Generates a list of --extern arguments.