Module cargo::core::compiler::artifact

Generate artifact information from unit dependencies for configuring the compiler environment.

Functions

get_env

Return all environment variables for the given unit-dependencies if artifacts are present.

match_artifacts_kind_with_targets 🔒

Given a dependency with an artifact artifact_dep and a set of available targets of its package, find a target for each kind of artifacts that are to be built.

unit_artifact_type_name_upper 🔒