Module cargo::core::compiler::compilation

Structs

Compilation

A structure returning the result of a compilation.

Doctest

Structure with enough information to run rustdoc --test.

UnitOutput

Information about the output of a unit.

Functions

fill_rustc_tool_env 🔒

Prepares a rustc_tool process with additional environment variables that are only relevant in a context that has a unit

target_runner 🔒