Module cargo::core::compiler::context::compilation_files

Structs

CompilationFiles

Collection of information about the files emitted by the compiler, and the output directory structure.

MetaInfo 🔒

Information about the metadata hashes used for a Unit.

Metadata

The Metadata is a hash used to make unique file names for each unit in a build. It is also use for symbol mangling.

OutputFile

Info about a single file emitted by the compiler.

Constants

METADATA_VERSION 🔒

This is a generic version number that can be changed to make backwards-incompatible changes to any file structures in the output directory. For example, the fingerprint files or the build-script output files. Normally cargo updates ship with rustc updates which will cause a new hash due to the rustc version changing, but this allows cargo to be extra careful to deal with different versions of cargo that use the same rustc version.

Functions

compute_metadata 🔒

hash_rustc_version 🔒

metadata_of 🔒

should_use_metadata 🔒

Returns whether or not this unit should use a metadata hash.