Struct cargo::core::compiler::context::compilation_files::CompilationFiles

pub struct CompilationFiles<'a, 'cfg> {
    pub(super) host: Layout,
    pub(super) target: HashMap<CompileTarget, Layout>,
    export_dir: Option<PathBuf>,
    roots: Vec<Unit>,
    ws: &'a Workspace<'cfg>,
    metas: HashMap<Unit, MetaInfo>,
    outputs: HashMap<Unit, LazyCell<Arc<Vec<OutputFile>>>>,
}

Collection of information about the files emitted by the compiler, and the output directory structure.

Fields
host: Layout

The target directory layout for the host (and target if it is the same as host).

target: HashMap<CompileTarget, Layout>

The target directory layout for the target (if different from then host).

export_dir: Option<PathBuf>

Additional directory to include a copy of the outputs.

roots: Vec<Unit>

The root targets requested by the user on the command line (does not include dependencies).

ws: &'a Workspace<'cfg>metas: HashMap<Unit, MetaInfo>

Metadata hash to use for each unit.

outputs: HashMap<Unit, LazyCell<Arc<Vec<OutputFile>>>>

For each Unit, a list all files produced.

Implementations

impl<'a, 'cfg: 'a> CompilationFiles<'a, 'cfg>

pub(super) fn new(
cx: &Context<'a, 'cfg>,
host: Layout,
target: HashMap<CompileTarget, Layout>
) -> CompilationFiles<'a, 'cfg>

pub fn layout(&self, kind: CompileKind) -> &Layout

Returns the appropriate directory layout for either a plugin or not.

pub fn metadata(&self, unit: &Unit) -> Metadata

Gets the metadata for the given unit.

See module docs for more details.

pub fn use_extra_filename(&self, unit: &Unit) -> bool

Returns whether or not -C extra-filename is used to extend the output filenames to make them unique.

pub fn target_short_hash(&self, unit: &Unit) -> String

Gets the short hash based only on the PackageId. Used for the metadata when metadata returns None.

pub fn out_dir(&self, unit: &Unit) -> PathBuf

Returns the directory where the artifacts for the given unit are initially created.

pub fn export_dir(&self) -> Option<PathBuf>

Additional export directory from --out-dir.

fn pkg_dir(&self, unit: &Unit) -> String

Directory name to use for a package in the form NAME-HASH.

Note that some units may share the same directory, so care should be taken in those cases!

pub fn host_dest(&self) -> &Path

Returns the final artifact path for the host (/…/target/debug)

pub fn host_root(&self) -> &Path

Returns the root of the build output tree for the host (/…/target)

pub fn host_deps(&self) -> &Path

Returns the host deps directory path.

pub fn deps_dir(&self, unit: &Unit) -> &Path

Returns the directories where Rust crate dependencies are found for the specified unit.

pub fn fingerprint_dir(&self, unit: &Unit) -> PathBuf

Directory where the fingerprint for the given unit should go.

pub fn fingerprint_file_path(&self, unit: &Unit, prefix: &str) -> PathBuf

Returns the path for a file in the fingerprint directory.

The “prefix” should be something to distinguish the file from other files in the fingerprint directory.

pub fn message_cache_path(&self, unit: &Unit) -> PathBuf

Path where compiler output is cached.

pub fn build_script_dir(&self, unit: &Unit) -> PathBuf

Returns the directory where a compiled build script is stored. /path/to/target/{debug,release}/build/PKG-HASH

fn artifact_dir(&self, unit: &Unit) -> PathBuf

Returns the directory for compiled artifacts files. /path/to/target/{debug,release}/deps/artifact/KIND/PKG-HASH

pub fn build_script_run_dir(&self, unit: &Unit) -> PathBuf

Returns the directory where information about running a build script is stored. /path/to/target/{debug,release}/build/PKG-HASH

pub fn build_script_out_dir(&self, unit: &Unit) -> PathBuf

Returns the “OUT_DIR” directory for running a build script. /path/to/target/{debug,release}/build/PKG-HASH/out

pub fn bin_link_for_target(
&self,
target: &Target,
kind: CompileKind,
bcx: &BuildContext<'_, '_>
) -> CargoResult<PathBuf>

Returns the path to the executable binary for the given bin target.

This should only to be used when a Unit is not available.

pub(super) fn outputs(
&self,
unit: &Unit,
bcx: &BuildContext<'a, 'cfg>
) -> CargoResult<Arc<Vec<OutputFile>>>

Returns the filenames that the given unit will generate.

Note: It is not guaranteed that all of the files will be generated.

fn uplift_to(
&self,
unit: &Unit,
file_type: &FileType,
from_path: &Path
) -> Option<PathBuf>

Returns the path where the output for the given unit and FileType should be uplifted to.

Returns None if the unit shouldn’t be uplifted (for example, a dependent rlib).

fn calc_outputs(
&self,
unit: &Unit,
bcx: &BuildContext<'a, 'cfg>
) -> CargoResult<Arc<Vec<OutputFile>>>

fn calc_outputs_rustc(
&self,
unit: &Unit,
bcx: &BuildContext<'a, 'cfg>
) -> CargoResult<Vec<OutputFile>>

Computes the actual, full pathnames for all the files generated by rustc.

The OutputFile also contains the paths where those files should be “uplifted” to.