Struct cargo::core::compiler::custom_build::BuildScripts

pub struct BuildScripts {
    pub to_link: Vec<(PackageId, Metadata)>,
    seen_to_link: HashSet<(PackageId, Metadata)>,
    pub plugins: BTreeSet<(PackageId, Metadata)>,
}

Linking information for a Unit.

See build_map for more details.

Fields
to_link: Vec<(PackageId, Metadata)>

List of build script outputs this Unit needs to include for linking. Each element is an index into BuildScriptOutputs.

Cargo will use this to_link vector to add -L flags to compiles as we propagate them upwards towards the final build. Note, however, that we need to preserve the ordering of to_link to be topologically sorted. This will ensure that build scripts which print their paths properly will correctly pick up the files they generated (if there are duplicates elsewhere).

To preserve this ordering, the (id, metadata) is stored in two places, once in the Vec and once in seen_to_link for a fast lookup. We maintain this as we’re building interactively below to ensure that the memory usage here doesn’t blow up too much.

For more information, see #2354.

seen_to_link: HashSet<(PackageId, Metadata)>

This is only used while constructing to_link to avoid duplicates.

plugins: BTreeSet<(PackageId, Metadata)>

Host-only dependencies that have build scripts. Each element is an index into BuildScriptOutputs.

This is the set of transitive dependencies that are host-only (proc-macro, plugin, build-dependency) that contain a build script. Any BuildOutput::library_paths path relative to target will be added to LD_LIBRARY_PATH so that the compiler can find any dynamic libraries a build script may have generated.