cargo::core::compiler::fingerprint::dirty_reason
Enum DirtyReason

pub enum DirtyReason {

    RustcChanged,
    FeaturesChanged {
        old: String,
        new: String,
    },
    DeclaredFeaturesChanged {
        old: String,
        new: String,
    },
    TargetConfigurationChanged,
    PathToSourceChanged,
    ProfileConfigurationChanged,
    RustflagsChanged {
        old: Vec<String>,
        new: Vec<String>,
    },
    MetadataChanged,
    ConfigSettingsChanged,
    CompileKindChanged,
    LocalLengthsChanged,
    PrecalculatedComponentsChanged {
        old: String,
        new: String,
    },
    ChecksumUseChanged {
        old: bool,
    },
    DepInfoOutputChanged {
        old: PathBuf,
        new: PathBuf,
    },
    RerunIfChangedOutputFileChanged {
        old: PathBuf,
        new: PathBuf,
    },
    RerunIfChangedOutputPathsChanged {
        old: Vec<PathBuf>,
        new: Vec<PathBuf>,
    },
    EnvVarsChanged {
        old: String,
        new: String,
    },
    EnvVarChanged {
        name: String,
        old_value: Option<String>,
        new_value: Option<String>,
    },
    LocalFingerprintTypeChanged {
        old: &'static str,
        new: &'static str,
    },
    NumberOfDependenciesChanged {
        old: usize,
        new: usize,
    },
    UnitDependencyNameChanged {
        old: InternedString,
        new: InternedString,
    },
    UnitDependencyInfoChanged {
        old_name: InternedString,
        old_fingerprint: u64,
        new_name: InternedString,
        new_fingerprint: u64,
    },
    FsStatusOutdated(FsStatus),
    NothingObvious,
    Forced,
    FreshBuild,

}

Tells a better story of why a build is considered โdirtyโ that leads to a recompile. Usually constructed via Fingerprint::compare.

Variants

RustcChanged

FeaturesChanged

Fields

old: String

new: String

DeclaredFeaturesChanged

Fields

old: String

new: String

TargetConfigurationChanged

PathToSourceChanged

ProfileConfigurationChanged

RustflagsChanged

Fields

old: Vec<String>

new: Vec<String>

MetadataChanged

ConfigSettingsChanged

CompileKindChanged

LocalLengthsChanged

PrecalculatedComponentsChanged

Fields

old: String

new: String

ChecksumUseChanged

Fields

old: bool

DepInfoOutputChanged

Fields

old: PathBuf

new: PathBuf

RerunIfChangedOutputFileChanged

Fields

old: PathBuf

new: PathBuf

RerunIfChangedOutputPathsChanged

Fields

old: Vec<PathBuf>

new: Vec<PathBuf>

EnvVarsChanged

Fields

old: String

new: String

EnvVarChanged

Fields

name: String

old_value: Option<String>

new_value: Option<String>

LocalFingerprintTypeChanged

Fields

old: &'static str

new: &'static str

NumberOfDependenciesChanged

Fields

old: usize

new: usize

UnitDependencyNameChanged

Fields

old: InternedString

new: InternedString

UnitDependencyInfoChanged

Fields

old_name: InternedString

old_fingerprint: u64

new_name: InternedString

new_fingerprint: u64

FsStatusOutdated(FsStatus)

NothingObvious

Forced

FreshBuild

First time to build something.

Implementations

impl DirtyReason

pub fn is_fresh_build(&self) -> bool

Whether a build is dirty because it is a fresh build being kicked off.

fn after(old_time: FileTime, new_time: FileTime, what: &'static str) -> After

pub fn present_to( &self, s: &mut Shell, unit: &Unit, root: &Path, ) -> CargoResult<()>