Struct cargo::core::compiler::job_queue::DrainState

struct DrainState<'cfg> {

    total_units: usize,
    queue: DependencyQueue<Unit, Artifact, Job>,
    messages: Arc<Queue<Message>>,
    diag_dedupe: DiagDedupe<'cfg>,
    warning_count: HashMap<JobId, WarningCount>,
    active: HashMap<JobId, Unit>,
    compiled: HashSet<PackageId>,
    documented: HashSet<PackageId>,
    scraped: HashSet<PackageId>,
    counts: HashMap<PackageId, usize>,
    progress: Progress<'cfg>,
    next_id: u32,
    timings: Timings<'cfg>,
    tokens: Vec<Acquired>,
    pending_queue: Vec<(Unit, Job, usize)>,
    print: DiagnosticPrinter<'cfg>,
    finished: usize,
    per_package_future_incompat_reports: Vec<FutureIncompatReportPackage>,

}

This structure is backed by the DependencyQueue type and manages the actual compilation step of each package. Packages enqueue units of work and then later on the entire graph is processed and compiled.

It is created from JobQueue when we have fully assembled the crate graph (i.e., all package dependencies are known).

Fields
total_units: usizequeue: DependencyQueue<Unit, Artifact, Job>messages: Arc<Queue<Message>>diag_dedupe: DiagDedupe<'cfg>

Diagnostic deduplication support.

warning_count: HashMap<JobId, WarningCount>

Count of warnings, used to print a summary after the job succeeds

active: HashMap<JobId, Unit>compiled: HashSet<PackageId>documented: HashSet<PackageId>scraped: HashSet<PackageId>counts: HashMap<PackageId, usize>progress: Progress<'cfg>next_id: u32timings: Timings<'cfg>tokens: Vec<Acquired>

Tokens that are currently owned by this Cargo, and may be “associated” with a rustc process. They may also be unused, though if so will be dropped on the next loop iteration.

Note that the length of this may be zero, but we will still spawn work, as we share the implicit token given to this Cargo process with a single rustc process.

pending_queue: Vec<(Unit, Job, usize)>

The list of jobs that we have not yet started executing, but have retrieved from the queue. We eagerly pull jobs off the main queue to allow us to request jobserver tokens pretty early.

print: DiagnosticPrinter<'cfg>finished: usize

How many jobs we’ve finished

per_package_future_incompat_reports: Vec<FutureIncompatReportPackage>
Implementations

impl<'cfg> DrainState<'cfg>

fn spawn_work_if_possible<'s>( &mut self, cx: &mut Context<'_, '_>, jobserver_helper: &HelperThread, scope: &'s Scope<'s, '_> ) -> CargoResult<()>

fn has_extra_tokens(&self) -> bool

fn handle_event( &mut self, cx: &mut Context<'_, '_>, plan: &mut BuildPlan, event: Message ) -> Result<(), ErrorToHandle>

fn wait_for_events(&mut self) -> Vec<Message>

fn drain_the_queue<'s>( self, cx: &mut Context<'_, '_>, plan: &mut BuildPlan, scope: &'s Scope<'s, '_>, jobserver_helper: &HelperThread ) -> Option<Error>

This is the “main” loop, where Cargo does all work to run the compiler.

This returns an Option to prevent the use of ? on Result types because it is important for the loop to carefully handle errors.

fn handle_error( &self, shell: &mut Shell, err_state: &mut ErrorsDuringDrain, new_err: impl Into<ErrorToHandle> )

fn tick_progress(&mut self)

fn name_for_progress(&self, unit: &Unit) -> String

fn run<'s>( &mut self, unit: &Unit, job: Job, cx: &Context<'_, '_>, scope: &'s Scope<'s, '_> )

Executes a job.

Fresh jobs block until finished (which should be very fast!), Dirty jobs will spawn a thread in the background and return immediately.

fn emit_warnings( &mut self, msg: Option<&str>, unit: &Unit, cx: &mut Context<'_, '_> ) -> CargoResult<()>

fn bump_warning_count(&mut self, id: JobId, emitted: bool, fixable: bool)

fn report_warning_count( &mut self, config: &Config, id: JobId, rustc_workspace_wrapper: &Option<PathBuf> )

Displays a final report of the warnings emitted by a particular job.

fn finish( &mut self, id: JobId, unit: &Unit, artifact: Artifact, cx: &mut Context<'_, '_> ) -> CargoResult<()>

fn note_working_on( &mut self, config: &Config, ws_root: &Path, unit: &Unit, fresh: &Freshness ) -> CargoResult<()>

fn back_compat_notice( &self, cx: &Context<'_, '_>, unit: &Unit ) -> CargoResult<()>