Struct cargo::core::compiler::job_queue::JobQueue

pub struct JobQueue<'cfg> {
    queue: DependencyQueue<Unit, Artifact, Job>,
    counts: HashMap<PackageId, usize>,
    timings: Timings<'cfg>,
}

This structure is backed by the DependencyQueue type and manages the queueing of compilation steps for each package. Packages enqueue units of work and then later on the entire graph is converted to DrainState and executed.

Fields
queue: DependencyQueue<Unit, Artifact, Job>counts: HashMap<PackageId, usize>timings: Timings<'cfg>
Implementations

impl<'cfg> JobQueue<'cfg>

pub fn new(bcx: &BuildContext<'_, 'cfg>) -> JobQueue<'cfg>

pub fn enqueue( &mut self, cx: &Context<'_, 'cfg>, unit: &Unit, job: Job ) -> CargoResult<()>

pub fn execute( self, cx: &mut Context<'_, '_>, plan: &mut BuildPlan ) -> CargoResult<()>

Executes all jobs necessary to build the dependency graph.

This function will spawn off config.jobs() workers to build all of the necessary dependencies, in order. Freshness is propagated as far as possible along each dependency chain.