cargo::core::compiler

Module output_sbom

cargo-sbom precursor files for external tools to create SBOM files from. See build_sbom_graph for more.

Structs

Sbom

SbomCrate 🔒

SbomDependency 🔒

SbomFormatVersion

Typed version of a SBOM format version number.

SbomIndex 🔒

SbomRustc 🔒

Enums

SbomDependencyType 🔒

Functions

build_sbom

Build an Sbom for the given Unit.

build_sbom_graph 🔒

List all dependencies, including transitive ones. A dependency can also appear multiple times if it’s using different settings, e.g. profile, features or crate versions.