Module cargo::core::compiler::rustdoc

Utilities for building with rustdoc.

Structs

RustdocExternMap

A map of registry names to URLs where documentations are hosted. This is for unstable feature -Zrustdoc-map.

Enums

RustdocExternMode

Mode used for std. This is for unstable feature -Zrustdoc-map.

RustdocScrapeExamples

Indicates whether a target should have examples scraped from it by rustdoc. Configured within Cargo.toml and only for unstable feature -Zrustdoc-scrape-examples.

Constants

DOCS_RS_URL 🔒

Functions

add_output_format

Adds unstable flag --output-format to the given rustdoc invocation. This is for unstable feature [-Zunstable-features].

add_root_urls

Adds unstable flag --extern-html-root-url to the given rustdoc invocation. This is for unstable feature -Zrustdoc-map.

build_all_urls 🔒

Recursively generate html root url for all units and their children.

default_crates_io_to_docs_rs 🔒