Function cargo::core::compiler::standard_lib::generate_std_roots

pub fn generate_std_roots<'a>(
    bcx: &BuildContext<'a, '_>, 
    crates: &[String], 
    std_resolve: &'a Resolve, 
    kind: CompileKind
) -> CargoResult<Vec<Unit<'a>>>

Generate a list of root Units for the standard library.

The given slice of crate names is the root set.