Struct cargo::core::compiler::FileType

pub struct FileType {
    pub flavor: FileFlavor,
    pub crate_type: Option<CrateType>,
    suffix: String,
    prefix: String,
    should_replace_hyphens: bool,
}

Type of each file generated by a Unit.

Fields
flavor: FileFlavor

The kind of file.

crate_type: Option<CrateType>

The crate-type that generates this file.

None for things that aren’t associated with a specific crate type, for example rmeta files.

suffix: String

The suffix for the file (for example, .rlib). This is an empty string for executables on Unix-like platforms.

prefix: String

The prefix for the file (for example, lib). This is an empty string for things like executables.

should_replace_hyphens: bool

Flag to convert hyphen to underscore when uplifting.

Implementations

impl FileType

pub fn output_filename(&self, target: &Target, metadata: Option<&str>) -> String

The filename for this FileType crated by rustc.

pub fn uplift_filename(&self, target: &Target) -> String

The filename for this FileType that Cargo should use when “uplifting” it to the destination directory.

pub fn new_rmeta() -> FileType

Creates a new instance representing a .rmeta file.