cargo::core::compiler

Module timings

Timing tracking.

This module implements some simple tracking information for timing of how long it takes for different units to compile.

Structs

CompilationSection

Section of compilation (e.g. frontend, backend, linking).

Concurrency 🔒

Periodic concurrency tracking information.

SectionData 🔒

Postprocessed section data that has both start and an end.

SectionTiming

Represents a certain section (phase) of rustc compilation. It is emitted by rustc when the --json=timings flag is used.

Timings

Tracking information for the entire build.

UnitTime 🔒

Tracking information for an individual unit.

Enums

AggregatedSections 🔒

Contains post-processed data of individual compilation sections.

SectionTimingEvent

Start or end of a section timing.

Constants

CODEGEN_SECTION_NAME 🔒

FRONTEND_SECTION_NAME 🔒

Statics

HTML_CANVAS 🔒

HTML_TMPL 🔒

Functions

render_rustc_info 🔒