Struct cargo::core::compiler::timings::UnitTime

struct UnitTime {
    unit: Unit,
    target: String,
    start: f64,
    duration: f64,
    rmeta_time: Option<f64>,
    unlocked_units: Vec<Unit>,
    unlocked_rmeta_units: Vec<Unit>,
}

Tracking information for an individual unit.

Fields
unit: Unittarget: String

A string describing the cargo target.

start: f64

The time when this unit started as an offset in seconds from Timings::start.

duration: f64

Total time to build this unit in seconds.

rmeta_time: Option<f64>

The time when the .rmeta file was generated, an offset in seconds from start.

unlocked_units: Vec<Unit>

Reverse deps that are freed to run after this unit finished.

unlocked_rmeta_units: Vec<Unit>

Same as unlocked_units, but unlocked by rmeta.

Implementations

impl UnitTime

fn codegen_time(&self) -> Option<(f64, f64, f64)>

Returns the codegen time as (rmeta_time, codegen_time, percent of total)

fn name_ver(&self) -> String