Function cargo::core::compiler::unit_dependencies::build_unit_dependencies

pub fn build_unit_dependencies<'a, 'gctx>(
    ws: &'a Workspace<'gctx>,
    package_set: &'a PackageSet<'gctx>,
    resolve: &'a Resolve,
    features: &'a ResolvedFeatures,
    std_resolve: Option<&'a (Resolve, ResolvedFeatures)>,
    roots: &[Unit],
    scrape_units: &[Unit],
    std_roots: &HashMap<CompileKind, Vec<Unit>>,
    global_mode: CompileMode,
    target_data: &'a RustcTargetData<'gctx>,
    profiles: &'a Profiles,
    interner: &'a UnitInterner,
) -> CargoResult<UnitGraph>

Then entry point for building a dependency graph of compilation units.

You can find some information for arguments from doc of State.