Function cargo::core::compiler::unit_dependencies::dep_build_script

fn dep_build_script(
    unit: &Unit,
    unit_for: UnitFor,
    state: &State<'_, '_>
) -> CargoResult<Option<UnitDep>>

If a build script is scheduled to be run for the package specified by unit, this function will return the unit to run that build script.

Overriding a build script simply means that the running of the build script itself doesn’t have any dependencies, so even in that case a unit of work is still returned. None is only returned if the package has no build script.