Module cargo::core::dependency

Structs

Artifact

The presence of an artifact turns an ordinary dependency into an Artifact dependency. As such, it will build one or more different artifacts of possibly various kinds for making them available at build time for rustc invocations or runtime for build scripts.

Dependency

Information about a dependency requested by a Cargo manifest. Cheap to copy.

Inner 🔒

The data underlying a Dependency.

SerializedArtifact

SerializedDependency 🔒

Enums

ArtifactKind

ArtifactTarget

DepKind