Enum cargo::core::GitReference

pub enum GitReference {
    Tag(String),
    Branch(String),
    Rev(String),
}

Information to find a specific commit in a git repository

Variants
Tag(String)

from a tag

Branch(String)

from the HEAD of a branch

Rev(String)

from a specific revision

Methods

impl GitReference

pub fn pretty_ref(&self) -> Option<PrettyRef>

Returns a Displayable view of this git reference, or None if using the head of the "master" branch