Enum cargo::core::WorkspaceConfig

pub enum WorkspaceConfig {
    Root {
        members: Option<Vec<String>>,
    },
    Member {
        root: Option<String>,
    },
}

Configuration of a workspace in a manifest.

Variants
Root

Indicates that [workspace] was present and the members were optionally specified as well.

Fields

members: Option<Vec<String>>

Member

Indicates that [workspace] was present and the root field is the optional value of package.workspace, if present.

Fields

root: Option<String>