cargo::core::features
Struct Feature

pub struct Feature {
    name: &'static str,
    stability: Status,
    version: &'static str,
    docs: &'static str,
    get: fn(_: &Features) -> bool,
}

Status and metadata for a single unstable feature.

Fields
name: &'static str

Feature name. This is valid Rust identifier so no dash only underscore.

stability: Statusversion: &'static str

Version that this feature was stabilized or removed.

docs: &'static str

Link to the unstable documentation.

get: fn(_: &Features) -> bool
Implementations

impl Feature

pub const fn test_dummy_stable() -> &'static Feature

A dummy feature that doesn’t actually gate anything, but it’s used in testing to ensure that we can enable stable features.

See https://doc.rust-lang.org/nightly/cargo/.

pub const fn test_dummy_unstable() -> &'static Feature

A dummy feature that gates the usage of the im-a-teapot manifest entry. This is basically just intended for tests.

See https://doc.rust-lang.org/nightly/cargo/reference/unstable.html.

pub const fn alternative_registries() -> &'static Feature

Downloading packages from alternative registry indexes.

See https://doc.rust-lang.org/nightly/cargo/reference/registries.html.

pub const fn edition() -> &'static Feature

Using editions

See https://doc.rust-lang.org/nightly/cargo/reference/manifest.html#the-edition-field.

pub const fn rename_dependency() -> &'static Feature

Renaming a package in the manifest via the package key.

See https://doc.rust-lang.org/nightly/cargo/reference/specifying-dependencies.html#renaming-dependencies-in-cargotoml.

pub const fn publish_lockfile() -> &'static Feature

Whether a lock file is published with this crate.

See https://doc.rust-lang.org/nightly/cargo/reference/unstable.html#publish-lockfile.

pub const fn profile_overrides() -> &'static Feature

Overriding profiles for dependencies.

See https://doc.rust-lang.org/nightly/cargo/reference/profiles.html#overrides.

pub const fn default_run() -> &'static Feature

“default-run” manifest option.

See https://doc.rust-lang.org/nightly/cargo/reference/manifest.html#the-default-run-field.

pub const fn metabuild() -> &'static Feature

Declarative build scripts.

See https://doc.rust-lang.org/nightly/cargo/reference/unstable.html#metabuild.

pub const fn public_dependency() -> &'static Feature

Specifying the ‘public’ attribute on dependencies.

See https://doc.rust-lang.org/nightly/cargo/reference/unstable.html#public-dependency.

pub const fn named_profiles() -> &'static Feature

Allow to specify profiles other than ‘dev’, ‘release’, ‘test’, etc.

See https://doc.rust-lang.org/nightly/cargo/reference/profiles.html#custom-profiles.

pub const fn resolver() -> &'static Feature

Opt-in new-resolver behavior.

See https://doc.rust-lang.org/nightly/cargo/reference/resolver.html#resolver-versions.

pub const fn strip() -> &'static Feature

Allow to specify whether binaries should be stripped.

See https://doc.rust-lang.org/nightly/cargo/reference/profiles.html#strip-option.

pub const fn rust_version() -> &'static Feature

Specifying a minimal ‘rust-version’ attribute for crates.

See https://doc.rust-lang.org/nightly/cargo/reference/manifest.html#the-rust-version-field.

pub const fn edition2021() -> &'static Feature

Support for 2021 edition.

See https://doc.rust-lang.org/nightly/cargo/reference/manifest.html#the-edition-field.

pub const fn per_package_target() -> &'static Feature

Allow to specify per-package targets (compile kinds).

See https://doc.rust-lang.org/nightly/cargo/reference/unstable.html#per-package-target.

pub const fn codegen_backend() -> &'static Feature

Allow to specify which codegen backend should be used.

See https://doc.rust-lang.org/nightly/cargo/reference/unstable.html#codegen-backend.

pub const fn different_binary_name() -> &'static Feature

Allow specifying different binary name apart from the crate name.

See https://doc.rust-lang.org/nightly/cargo/reference/unstable.html#different-binary-name.

pub const fn profile_rustflags() -> &'static Feature

Allow specifying rustflags directly in a profile.

See https://doc.rust-lang.org/nightly/cargo/reference/unstable.html#profile-rustflags-option.

pub const fn workspace_inheritance() -> &'static Feature

Allow workspace members to inherit fields and dependencies from a workspace.

See https://doc.rust-lang.org/nightly/cargo/reference/unstable.html#workspace-inheritance.

pub const fn edition2024() -> &'static Feature

Support for 2024 edition.

See https://doc.rust-lang.org/nightly/cargo/reference/manifest.html#the-edition-field.

pub const fn trim_paths() -> &'static Feature

Allow setting trim-paths in a profile to control the sanitisation of file paths in build outputs.

See https://doc.rust-lang.org/nightly/cargo/reference/unstable.html#profile-trim-paths-option.

pub const fn open_namespaces() -> &'static Feature

Allow multiple packages to participate in the same API namespace

See https://doc.rust-lang.org/nightly/cargo/reference/unstable.html#open-namespaces.

pub const fn path_bases() -> &'static Feature

Allow paths that resolve relatively to a base specified in the config.

See https://doc.rust-lang.org/nightly/cargo/reference/unstable.html#path-bases.

fn is_enabled(&self, features: &Features) -> bool

Whether this feature is allowed to use in the given Features context.

pub(crate) fn name(&self) -> &str