cargo::core::features
Struct Features

pub struct Features {

    test_dummy_stable: bool,
    test_dummy_unstable: bool,
    alternative_registries: bool,
    edition: bool,
    rename_dependency: bool,
    publish_lockfile: bool,
    profile_overrides: bool,
    default_run: bool,
    metabuild: bool,
    public_dependency: bool,
    named_profiles: bool,
    resolver: bool,
    strip: bool,
    rust_version: bool,
    edition2021: bool,
    per_package_target: bool,
    codegen_backend: bool,
    different_binary_name: bool,
    profile_rustflags: bool,
    workspace_inheritance: bool,
    edition2024: bool,
    trim_paths: bool,
    open_namespaces: bool,
    path_bases: bool,
    activated: Vec<String>,
    nightly_features_allowed: bool,
    is_local: bool,

}

Unstable feature context for querying if a new Cargo.toml syntax is allowed to use.

See the module-level documentation for the usage.

Fields
test_dummy_stable: booltest_dummy_unstable: boolalternative_registries: booledition: boolrename_dependency: boolpublish_lockfile: boolprofile_overrides: booldefault_run: boolmetabuild: boolpublic_dependency: boolnamed_profiles: boolresolver: boolstrip: boolrust_version: booledition2021: boolper_package_target: boolcodegen_backend: booldifferent_binary_name: boolprofile_rustflags: boolworkspace_inheritance: booledition2024: booltrim_paths: boolopen_namespaces: boolpath_bases: boolactivated: Vec<String>

The current activated features.

nightly_features_allowed: bool

Whether is allowed to use any unstable features.

is_local: bool

Whether the source manifest is from a local package.

Implementations

impl Features

fn status(&mut self, feature: &str) -> Option<(&mut bool, &'static Feature)>

impl Features

pub fn new( features: &[String], gctx: &GlobalContext, warnings: &mut Vec<String>, is_local: bool, ) -> CargoResult<Features>

Creates a new unstable features context.

fn add( &mut self, feature_name: &str, gctx: &GlobalContext, warnings: &mut Vec<String>, ) -> CargoResult<()>

pub fn activated(&self) -> &[String]

Gets the current activated features.

pub fn require(&self, feature: &Feature) -> CargoResult<()>

Checks if the given feature is enabled.

pub fn is_enabled(&self, feature: &Feature) -> bool

Whether the given feature is allowed to use in this context.