Module cargo::core::manifest

Structs

DelayedWarning
When parsing Cargo.toml, some warnings should silenced if the manifest comes from a dependency. ManifestWarning allows this delayed emission of warnings.

Manifest
Contains all the information about a package, as loaded from a Cargo.toml.

ManifestMetadata
General metadata about a package which is just blindly uploaded to the registry.

Target
Information about a binary, a library, an example, etc. that is part of the package.

VirtualManifest

Warnings

Enums

EitherManifest

TargetKind

TargetSourcePath