Struct cargo::core::manifest::Manifest

pub struct Manifest {

    summary: Summary,
    targets: Vec<Target>,
    default_kind: Option<CompileKind>,
    forced_kind: Option<CompileKind>,
    links: Option<String>,
    warnings: Warnings,
    exclude: Vec<String>,
    include: Vec<String>,
    metadata: ManifestMetadata,
    custom_metadata: Option<Value>,
    profiles: Option<TomlProfiles>,
    publish: Option<Vec<String>>,
    replace: Vec<(PackageIdSpec, Dependency)>,
    patch: HashMap<Url, Vec<Dependency>>,
    workspace: WorkspaceConfig,
    original: Rc<TomlManifest>,
    unstable_features: Features,
    edition: Edition,
    rust_version: Option<String>,
    im_a_teapot: Option<bool>,
    default_run: Option<String>,
    metabuild: Option<Vec<String>>,
    resolve_behavior: Option<ResolveBehavior>,
    lint_rustflags: Vec<String>,
    embedded: bool,

}

Contains all the information about a package, as loaded from a Cargo.toml.

This is deserialized using the TomlManifest type.

Fields
summary: Summarytargets: Vec<Target>default_kind: Option<CompileKind>forced_kind: Option<CompileKind>links: Option<String>warnings: Warningsexclude: Vec<String>include: Vec<String>metadata: ManifestMetadatacustom_metadata: Option<Value>profiles: Option<TomlProfiles>publish: Option<Vec<String>>replace: Vec<(PackageIdSpec, Dependency)>patch: HashMap<Url, Vec<Dependency>>workspace: WorkspaceConfigoriginal: Rc<TomlManifest>unstable_features: Featuresedition: Editionrust_version: Option<String>im_a_teapot: Option<bool>default_run: Option<String>metabuild: Option<Vec<String>>resolve_behavior: Option<ResolveBehavior>lint_rustflags: Vec<String>embedded: bool
Implementations

impl Manifest

pub fn new( summary: Summary, default_kind: Option<CompileKind>, forced_kind: Option<CompileKind>, targets: Vec<Target>, exclude: Vec<String>, include: Vec<String>, links: Option<String>, metadata: ManifestMetadata, custom_metadata: Option<Value>, profiles: Option<TomlProfiles>, publish: Option<Vec<String>>, replace: Vec<(PackageIdSpec, Dependency)>, patch: HashMap<Url, Vec<Dependency>>, workspace: WorkspaceConfig, unstable_features: Features, edition: Edition, rust_version: Option<String>, im_a_teapot: Option<bool>, default_run: Option<String>, original: Rc<TomlManifest>, metabuild: Option<Vec<String>>, resolve_behavior: Option<ResolveBehavior>, lint_rustflags: Vec<String>, embedded: bool ) -> Manifest

pub fn dependencies(&self) -> &[Dependency]

pub fn default_kind(&self) -> Option<CompileKind>

pub fn forced_kind(&self) -> Option<CompileKind>

pub fn exclude(&self) -> &[String]

pub fn include(&self) -> &[String]

pub fn metadata(&self) -> &ManifestMetadata

pub fn name(&self) -> InternedString

pub fn package_id(&self) -> PackageId

pub fn summary(&self) -> &Summary

pub fn summary_mut(&mut self) -> &mut Summary

pub fn targets(&self) -> &[Target]

pub fn targets_mut(&mut self) -> &mut [Target]

pub fn version(&self) -> &Version

pub fn warnings_mut(&mut self) -> &mut Warnings

pub fn warnings(&self) -> &Warnings

pub fn profiles(&self) -> Option<&TomlProfiles>

pub fn publish(&self) -> &Option<Vec<String>>

pub fn replace(&self) -> &[(PackageIdSpec, Dependency)]

pub fn original(&self) -> &TomlManifest

pub fn patch(&self) -> &HashMap<Url, Vec<Dependency>>

pub fn links(&self) -> Option<&str>

pub fn is_embedded(&self) -> bool

pub fn workspace_config(&self) -> &WorkspaceConfig

pub fn unstable_features(&self) -> &Features

Unstable, nightly features that are enabled in this manifest.

pub fn resolve_behavior(&self) -> Option<ResolveBehavior>

The style of resolver behavior to use, declared with the resolver field.

Returns None if it is not specified.

pub fn lint_rustflags(&self) -> &[String]

RUSTFLAGS from the [lints] table

pub fn map_source( self, to_replace: SourceId, replace_with: SourceId ) -> Manifest

pub fn feature_gate(&self) -> CargoResult<()>

pub fn print_teapot(&self, config: &Config)

pub fn edition(&self) -> Edition

pub fn rust_version(&self) -> Option<&str>

pub fn custom_metadata(&self) -> Option<&Value>

pub fn default_run(&self) -> Option<&str>

pub fn metabuild(&self) -> Option<&Vec<String>>

pub fn metabuild_path(&self, target_dir: Filesystem) -> PathBuf