Struct cargo::core::manifest::Target

pub struct Target { /* fields omitted */ }

Information about a binary, a library, an example, etc. that is part of the package.

Implementations

impl Target

pub fn lib_target(
name: &str,
crate_targets: Vec<CrateType>,
src_path: PathBuf,
edition: Edition
) -> Target

pub fn bin_target(
name: &str,
src_path: PathBuf,
required_features: Option<Vec<String>>,
edition: Edition
) -> Target

pub fn custom_build_target(
name: &str,
src_path: PathBuf,
edition: Edition
) -> Target

Builds a Target corresponding to the build = "build.rs" entry.

pub fn metabuild_target(name: &str) -> Target

pub fn example_target(
name: &str,
crate_targets: Vec<CrateType>,
src_path: PathBuf,
required_features: Option<Vec<String>>,
edition: Edition
) -> Target

pub fn test_target(
name: &str,
src_path: PathBuf,
required_features: Option<Vec<String>>,
edition: Edition
) -> Target

pub fn bench_target(
name: &str,
src_path: PathBuf,
required_features: Option<Vec<String>>,
edition: Edition
) -> Target

pub fn name(&self) -> &str

pub fn crate_name(&self) -> String

pub fn src_path(&self) -> &TargetSourcePath

pub fn set_src_path(&mut self, src_path: TargetSourcePath)

pub fn required_features(&self) -> Option<&Vec<String>>

pub fn kind(&self) -> &TargetKind

pub fn tested(&self) -> bool

pub fn harness(&self) -> bool

pub fn documented(&self) -> bool

pub fn for_host(&self) -> bool

pub fn proc_macro(&self) -> bool

pub fn edition(&self) -> Edition

pub fn benched(&self) -> bool

pub fn doctested(&self) -> bool

pub fn doctestable(&self) -> bool

pub fn is_lib(&self) -> bool

pub fn is_dylib(&self) -> bool

pub fn is_cdylib(&self) -> bool

pub fn is_linkable(&self) -> bool

Returns whether this target produces an artifact which can be linked into a Rust crate.

This only returns true for certain kinds of libraries.

pub fn is_bin(&self) -> bool

pub fn is_example(&self) -> bool

pub fn is_executable(&self) -> bool

Returns true if it is a binary or executable example. NOTE: Tests are false!

pub fn is_exe_example(&self) -> bool

Returns true if it is an executable example.

pub fn is_test(&self) -> bool

pub fn is_bench(&self) -> bool

pub fn is_custom_build(&self) -> bool

pub fn rustc_crate_types(&self) -> Vec<CrateType>

Returns the arguments suitable for --crate-type to pass to rustc.

pub fn set_tested(&mut self, tested: bool) -> &mut Target

pub fn set_benched(&mut self, benched: bool) -> &mut Target

pub fn set_doctest(&mut self, doctest: bool) -> &mut Target

pub fn set_for_host(&mut self, for_host: bool) -> &mut Target

pub fn set_proc_macro(&mut self, proc_macro: bool) -> &mut Target

pub fn set_edition(&mut self, edition: Edition) -> &mut Target

pub fn set_harness(&mut self, harness: bool) -> &mut Target

pub fn set_doc(&mut self, doc: bool) -> &mut Target

pub fn set_kind(&mut self, kind: TargetKind) -> &mut Target

pub fn set_name(&mut self, name: &str) -> &mut Target

pub fn set_required_features(
&mut self,
required_features: Option<Vec<String>>
) -> &mut Target

pub fn description_named(&self) -> String