Module cargo::core::profiles

Structs

Profile

Profile settings used to determine which compiler flags to use for a target.

ProfileMaker 🔒

An object used for handling the profile hierarchy.

Profiles

Collection of all profiles.

UnitFor

Flags used in creating Units to indicate the purpose for the target, and to ensure the target’s dependencies have the correct settings.

Enums

Lto

The link-time-optimization setting.

PanicSetting 🔒

PanicStrategy

The panic setting.

ProfileRoot

The root profile (dev/release).

Strip

The setting for choosing which symbols to strip

Functions

get_config_profile 🔒

Helper for fetching a profile from config.

is_off 🔒

Returns true if a string is a toggle that turns an option off.

merge_config_profiles 🔒

Takes the manifest profiles, and overlays the config profiles on-top.

merge_profile 🔒

Merge the given TOML profile into the given Profile.

merge_toml_overrides 🔒

Merge package and build overrides from the given TOML profile into the given Profile.

validate_packages_unique 🔒

Validate that a package does not match multiple package override specs.

validate_packages_unmatched 🔒

Check for any profile override specs that do not match any known packages.