Trait cargo::core::registry::Registry

pub trait Registry {
    fn query(&mut self, name: &Dependency) -> CargoResult<Vec<Summary>>;

    fn supports_checksums(&self) -> bool { ... }
}

Source of information about a group of packages.

See also core::Source.

Required Methods

fn query(&mut self, name: &Dependency) -> CargoResult<Vec<Summary>>

Attempt to find the packages that match a dependency request.

Provided Methods

fn supports_checksums(&self) -> bool

Returns whether or not this registry will return summaries with checksums listed.

By default, registries do not support checksums.