Type Definition cargo::core::registry::LockedMap

type LockedMap = HashMap<(SourceId, InternedString), Vec<(PackageId, Vec<PackageId>)>>;

A map of all “locked packages” which is filled in when parsing a lock file and is used to guide dependency resolution by altering summaries as they’re queried from this source.

This map can be thought of as a glorified Vec<MySummary> where MySummary has a PackageId for which package it represents as well as a list of PackageId for the resolved dependencies. The hash map is otherwise structured though for easy access throughout this registry.