Module cargo::core::resolver

Resolution of the entire dependency graph for a crate

This module implements the core logic in taking the world of crates and constraints and creating a resolved graph with locked versions for all crates and their dependencies. This is separate from the registry module which is more worried about discovering crates from various sources, this module just uses the Registry trait as a source to learn about crates from.

Actually solving a constraint graph is an NP-hard (or NP-complete, I forget which) problem, this the algorithm is basically a nice heuristic to make sure we get roughly the best answer most of the time. The constraints that we're working with are:

Each crate can have any number of dependencies. Each dependency can declare a version range that it is compatible with.

Crates can be activated with multiple version (e.g. show up in the dependency graph twice) so long as each pairwise instance have semver-incompatible versions.

The algorithm employed here is fairly simple, we simply do a DFS, activating the "newest crate" (highest version) first and then going to the next option. The heuristics we employ are:

Never try to activate a crate version which is incompatible. This means we only try crates which will actually satisfy a dependency and we won't ever try to activate a crate that's semver compatible with something else activatd (as we're only allowed to have one).

Always try to activate the highest version crate first. The default dependency in Cargo (e.g. when you write foo = "0.1.2") is semver-compatible, so selecting the highest version possible will allow us to hopefully satisfy as many dependencies at once.

Beyond that, what's implemented below is just a naive backtracking version which should in theory try all possible combinations of dependencies and versions to see if one works. The first resolution that works causes everything to bail out immediately and return success, and only if nothing works do we actually return an error up the stack.

Performance

Note that this is a relatively performance-critical portion of Cargo. The data that we're processing is proportional to the size of the dependency graph, which can often be quite large (e.g. take a look at Servo). To make matters worse the DFS algorithm we're implemented is inherently quite inefficient and recursive. When we add the requirement of backtracking on top it means that we're implementing something that's very recursive and probably shouldn't be allocating all over the place.

Once we've avoided too many unnecessary allocations, however (e.g. using references, using reference counting, etc), it turns out that the performance in this module largely comes down to stack sizes due to the recursive nature of the implementation.

Small Stack Sizes (e.g. y u inline(never))

One of the most important optimizations in this module right now is the attempt to minimize the stack space taken up by the activate and activate_deps functions. These two functions are mutually recursive in a CPS fashion.

The recursion depth, if I'm getting this right, is something along the order of O(E) where E is the number of edges in the dependency graph, and that's on the order of O(N2) where N is the number of crates in the graph. As a result we need to watch our stack size!

Currently rustc is not great at producing small stacks because of landing pads and filling drop, so the first attempt at making small stacks is having literally small functions with very few owned values on the stack. This is also why there are many #[inline(never)] annotations in this module. By preventing these functions from being inlined we can make sure that these stack sizes stay small as the number of locals are under control.

Another hazard when watching out for small stacks is passing around large structures by value. For example the Context below is a relatively large struct, so we always place it behind a Box to ensure the size at runtime is just a word (e.g. very easy to pass around).

Combined together these tricks (plus a very recent version of LLVM) allow us to have a relatively small stack footprint for this implementation. Possible future optimizations include:

Turn off landing pads for all of Cargo

Wait for dynamic drop

Use a manual stack instead of the OS stack (I suspect this will be super painful to implement)

Spawn a new thread with a very large stack (this is what the compiler does)

Implement a form of segmented stacks where we manually check the stack limit every so often.

For now the current implementation of this module gets us past Servo's dependency graph (one of the largest known ones), so hopefully it'll work for a bit longer as well!

Structs

EncodableDependency

EncodablePackageId

EncodableResolve

Resolve
Represents a fully resolved package dependency graph. Each node in the graph is a package and edges represent dependencies between packages.

Enums

Method

Functions

resolve
Builds the list of all packages required to build the first argument.

Type Definitions

Metadata