Struct cargo::core::resolver::conflict_cache::ConflictCache

pub(super) struct ConflictCache {
    con_from_dep: HashMap<Dependency, ConflictStoreTrie>,
    dep_from_pid: HashMap<PackageId, HashSet<Dependency>>,
}

Fields
con_from_dep: HashMap<Dependency, ConflictStoreTrie>dep_from_pid: HashMap<PackageId, HashSet<Dependency>>
Implementations

impl ConflictCache

pub fn new() -> ConflictCache

pub fn find(
&self,
dep: &Dependency,
is_active: &impl Fn(PackageId) -> Option<usize>,
must_contain: Option<PackageId>,
max_age: usize
) -> Option<&BTreeMap<PackageId, ConflictReason>>

pub fn find_conflicting(
&self,
cx: &Context,
dep: &Dependency,
must_contain: Option<PackageId>
) -> Option<&BTreeMap<PackageId, ConflictReason>>

Finds any known set of conflicts, if any, which are activated in cx and contain PackageId specified. If more than one are activated, then it will return one that will allow for the most jump-back.

pub fn conflicting(
&self,
cx: &Context,
dep: &Dependency
) -> Option<&BTreeMap<PackageId, ConflictReason>>

pub fn insert(
&mut self,
dep: &Dependency,
con: &BTreeMap<PackageId, ConflictReason>
)

Adds to the cache a conflict of the form: dep is known to be unresolvable if all the PackageId entries are activated.

pub fn dependencies_conflicting_with(
&self,
pid: PackageId
) -> Option<&HashSet<Dependency>>