Struct cargo::core::resolver::context::Context

pub struct Context {
    pub age: usize,
    pub activations: HashMap<(InternedString, SourceId, SemverCompatibility), (Summary, usize)>,
    pub resolve_features: HashMap<PackageId, Rc<BTreeSet<InternedString>>>,
    pub links: HashMap<InternedString, PackageId>,
    pub public_dependency: Option<PublicDependency>,
    pub parents: Graph<PackageId, HashSet<Dependency>>,
}

Fields
age: usizeactivations: HashMap<(InternedString, SourceId, SemverCompatibility), (Summary, usize)>resolve_features: HashMap<PackageId, Rc<BTreeSet<InternedString>>>

list the features that are activated for each package

links: HashMap<InternedString, PackageId>

get the package that will be linking to a native library by its links attribute

public_dependency: Option<PublicDependency>

for each package the list of names it can see, then for each name the exact version that name represents and whether the name is public.

parents: Graph<PackageId, HashSet<Dependency>>

a way to look up for a package in activations what packages required it and all of the exact deps that it fulfilled.

Implementations

impl Context

pub fn new(check_public_visible_dependencies: bool) -> Context

pub fn flag_activated( &mut self, summary: &Summary, opts: &ResolveOpts, parent: Option<(&Summary, &Dependency)> ) -> ActivateResult<bool>

Activate this summary by inserting it into our list of known activations.

The parent passed in here is the parent summary/dependency edge which cased summary to get activated. This may not be present for the root crate, for example.

Returns true if this summary with the given features is already activated.

pub fn is_active(&self, id: PackageId) -> Option<usize>

If the package is active returns the ContextAge when it was added

pub fn still_applies( &self, id: PackageId, reason: &ConflictReason ) -> Option<usize>

If the conflict reason on the package still applies returns the ContextAge when it was added

pub fn is_conflicting( &self, parent: Option<PackageId>, conflicting_activations: &BTreeMap<PackageId, ConflictReason> ) -> Option<usize>

Checks whether all of parent and the keys of conflicting activations are still active. If so returns the ContextAge when the newest one was added.

pub fn resolve_replacements( &self, registry: &RegistryQueryer<'_> ) -> HashMap<PackageId, PackageId>

pub fn graph(&self) -> Graph<PackageId, HashSet<Dependency>>