cargo::core::resolver::context
Struct ResolverContext

pub struct ResolverContext {
    pub age: usize,
    pub activations: HashMap<ActivationsKey, (Summary, usize), FxBuildHasher>,
    pub resolve_features: HashMap<PackageId, Rc<BTreeSet<InternedString>>, FxBuildHasher>,
    pub links: HashMap<InternedString, PackageId, FxBuildHasher>,
    pub parents: Graph<PackageId, HashSet<Dependency, FxBuildHasher>>,
}

Fields
age: usizeactivations: HashMap<ActivationsKey, (Summary, usize), FxBuildHasher>resolve_features: HashMap<PackageId, Rc<BTreeSet<InternedString>>, FxBuildHasher>

list the features that are activated for each package

links: HashMap<InternedString, PackageId, FxBuildHasher>

get the package that will be linking to a native library by its links attribute

parents: Graph<PackageId, HashSet<Dependency, FxBuildHasher>>

a way to look up for a package in activations what packages required it and all of the exact deps that it fulfilled.

Implementations

impl ResolverContext

pub fn new() -> ResolverContext

pub fn flag_activated( &mut self, summary: &Summary, opts: &ResolveOpts, parent: Option<(&Summary, &Dependency)>, ) -> ActivateResult<bool>

Activate this summary by inserting it into our list of known activations.

The parent passed in here is the parent summary/dependency edge which cased summary to get activated. This may not be present for the root crate, for example.

Returns true if this summary with the given features is already activated.

pub fn is_active(&self, id: PackageId) -> Option<usize>

If the package is active returns the ContextAge when it was added

pub fn is_conflicting( &self, parent: Option<PackageId>, conflicting_activations: &BTreeMap<PackageId, ConflictReason>, ) -> Option<usize>

Checks whether all of parent and the keys of conflicting activations are still active. If so returns the ContextAge when the newest one was added.

pub fn resolve_replacements( &self, registry: &RegistryQueryer<'_>, ) -> HashMap<PackageId, PackageId>

pub fn graph(&self) -> Graph<PackageId, HashSet<Dependency>>