cargo::core::resolver::dep_cache

Function build_requirements

fn build_requirements<'a, 'b: 'a>(
    parent: Option<PackageId>,
    s: &'a Summary,
    opts: &'b ResolveOpts,
) -> ActivateResult<Requirements<'a>>

Takes requested features for a single package from the input ResolveOpts and recurses to find all requested features, dependencies and requested dependency features in a Requirements object, returning it to the resolver.