Function cargo::core::resolver::dep_cache::resolve_features

pub fn resolve_features<'b>(
    parent: Option<PackageId>,
    s: &'b Summary,
    opts: &'b ResolveOpts
) -> ActivateResult<(HashSet<InternedString>, Vec<(Dependency, Rc<BTreeSet<InternedString>>)>)>

Returns the features we ended up using and all dependencies and the features we want from each of them.