Module cargo::core::resolver::errors

Structs

ResolveError

Error during resolution providing a path of PackageIds.

Enums

ActivateError

Functions

activation_error 🔒

describe_path 🔒

Returns String representation of dependency chain for a particular pkgid.

describe_path_in_context 🔒

Returns String representation of dependency chain for a particular pkgid within given context.

Type Definitions

ActivateResult