cargo::core::resolver

Function activate

fn activate(
    cx: &mut ResolverContext,
    registry: &mut RegistryQueryer<'_>,
    parent: Option<(&Summary, &Dependency)>,
    candidate: Summary,
    first_version: Option<VersionOrdering>,
    opts: &ResolveOpts,
) -> ActivateResult<Option<(DepsFrame, Duration)>>

Attempts to activate the summary candidate in the context cx.

This function will pull dependency summaries from the registry provided, and the dependencies of the package will be determined by the opts provided. If candidate was activated, this function returns the dependency frame to iterate through next.