cargo::core::resolver

Function activate_deps_loop

fn activate_deps_loop(
    registry: &mut RegistryQueryer<'_>,
    summaries: &[(Summary, ResolveOpts)],
    first_version: Option<VersionOrdering>,
    gctx: Option<&GlobalContext>,
    past_conflicting_activations: &mut ConflictCache,
) -> CargoResult<ResolverContext>

Recursively activates the dependencies for summaries, in depth-first order, backtracking across possible candidates for each dependency as necessary.

If all dependencies can be activated and resolved to a version in the dependency graph, cx is returned.