cargo::core::resolver
Struct RemainingCandidates

struct RemainingCandidates {
    remaining: RcVecIter<Summary>,
    has_another: Option<Summary>,
}

A helper โiteratorโ used to extract candidates within a current Context of a dependency graph.

This struct doesnโt literally implement the Iterator trait (requires a few more inputs) but in general acts like one. Each RemainingCandidates is created with a list of candidates to choose from. When attempting to iterate over the list of candidates only valid candidates are returned. Validity is defined within a Context.

Candidates passed to new may not be returned from next as they could be filtered out, and as they are filtered the causes will be added to conflicting_prev_active.

Fields
remaining: RcVecIter<Summary>has_another: Option<Summary>
Implementations

impl RemainingCandidates

fn new(candidates: &Rc<Vec<Summary>>) -> RemainingCandidates

fn next( &mut self, conflicting_prev_active: &mut BTreeMap<PackageId, ConflictReason>, cx: &ResolverContext, ) -> Option<(Summary, bool)>

Attempts to find another candidate to check from this list.

This method will attempt to move this iterator forward, returning a candidate thatโs possible to activate. The cx argument is the current context which determines validity for candidates returned, and the dep is the dependency listing that weโre activating for.

If successful a (Candidate, bool) pair will be returned. The Candidate is the candidate to attempt to activate, and the bool is an indicator of whether there are remaining candidates to try of if weโve reached the end of iteration.

If weโve reached the end of the iterator here then Err will be returned. The error will contain a map of package ID to conflict reason, where each package ID caused a candidate to be filtered out from the original list for the reason listed.