Struct cargo::core::resolver::Resolve

pub struct Resolve {
    // some fields omitted
}

Represents a fully resolved package dependency graph. Each node in the graph is a package and edges represent dependencies between packages.

Each instance of Resolve also understands the full set of features used for each package.

Methods

impl Resolve

fn merge_from(&mut self, previous: &Resolve) -> CargoResult<()>

fn iter(&self) -> Nodes<PackageId>

fn deps(&self, pkg: &PackageId) -> Deps

fn deps_not_replaced(&self, pkg: &PackageId) -> DepsNotReplaced

fn replacement(&self, pkg: &PackageId) -> Option<&PackageId>

fn replacements(&self) -> &HashMap<PackageId, PackageId>

fn features(&self, pkg: &PackageId) -> Option<&HashSet<String>>

fn query(&self, spec: &str) -> CargoResult<&PackageId>