cargo::core::resolver
Struct VersionPreferences

pub struct VersionPreferences {
    try_to_use: HashSet<PackageId>,
    prefer_patch_deps: HashMap<InternedString, HashSet<Dependency>>,
    version_ordering: VersionOrdering,
    rust_versions: Vec<PartialVersion>,
}

A collection of preferences for particular package versions.

This is built up with Self::prefer_package_id and Self::prefer_dependency, then used to sort the set of summaries for a package during resolution via Self::sort_summaries.

As written, a version is either “preferred” or “not preferred”. Later extensions may introduce more granular preferences.

Fields
try_to_use: HashSet<PackageId>prefer_patch_deps: HashMap<InternedString, HashSet<Dependency>>version_ordering: VersionOrderingrust_versions: Vec<PartialVersion>
Implementations

impl VersionPreferences

pub fn prefer_package_id(&mut self, pkg_id: PackageId)

Indicate that the given package (specified as a PackageId) should be preferred.

pub fn prefer_dependency(&mut self, dep: Dependency)

Indicate that the given package (specified as a Dependency) should be preferred.

pub fn version_ordering(&mut self, ordering: VersionOrdering)

pub fn rust_versions(&mut self, vers: Vec<PartialVersion>)

pub fn sort_summaries( &self, summaries: &mut Vec<Summary>, first_version: Option<VersionOrdering>, )

Sort (and filter) the given vector of summaries in-place

Note: all summaries presumed to be for the same package.

Sort order:

Preferred packages

Most compatible VersionPreferences::rust_versions

first_version, falling back to VersionPreferences::version_ordering when None

Filtering:

first_version

fn msrv_compat_count(&self, summary: &Summary) -> usize