cargo::core
Module shell

Modules

imp 🔒

Structs

Hyperlink

Shell

An abstraction around console output that remembers preferences for output verbosity and color.

Enums

ColorChoice

Whether messages should use color output

ShellOut 🔒

A Writeable object, either with or without color support

TtyWidth

Verbosity

The requested verbosity of output.

Functions

supports_color 🔒

supports_hyperlinks 🔒

supports_unicode 🔒