Enum cargo::core::shell::ShellOut

enum ShellOut {
    Write(Box<dyn Write>),
    Stream {
        stdout: StandardStream,
        stderr: StandardStream,
        stderr_tty: bool,
        color_choice: ColorChoice,
    },
}

A Writeable object, either with or without color support

Variants

Write(Box<dyn Write>)

A plain write object without color support

Stream

Fields

stdout: StandardStream

stderr: StandardStream

stderr_tty: bool

color_choice: ColorChoice

Color-enabled stdio, with information on whether color should be used

Implementations

impl ShellOut

fn message_stderr(
&mut self,
status: &dyn Display,
message: Option<&dyn Display>,
color: Color,
justified: bool
) -> CargoResult<()>

Prints out a message with a status. The status comes first, and is bold plus the given color. The status can be justified, in which case the max width that will right align is 12 chars.

fn write_stdout(
&mut self,
fragment: impl Display,
color: &ColorSpec
) -> CargoResult<()>

Write a styled fragment

fn write_stderr(
&mut self,
fragment: impl Display,
color: &ColorSpec
) -> CargoResult<()>

Write a styled fragment

fn stdout(&mut self) -> &mut dyn Write

Gets stdout as a io::Write.

fn stderr(&mut self) -> &mut dyn Write

Gets stderr as a io::Write.