Struct cargo::core::shell::Shell

pub struct Shell {
    output: ShellOut,
    verbosity: Verbosity,
    needs_clear: bool,
}

An abstraction around console output that remembers preferences for output verbosity and color.

Fields
output: ShellOut

Wrapper around stdout/stderr. This helps with supporting sending output to a memory buffer which is useful for tests.

verbosity: Verbosity

How verbose messages should be.

needs_clear: bool

Flag that indicates the current line needs to be cleared before printing. Used when a progress bar is currently displayed.

Implementations

impl Shell

pub fn new() -> Shell

Creates a new shell (color choice and verbosity), defaulting to ‘auto’ color and verbose output.

pub fn from_write(out: Box<dyn Write>) -> Shell

Creates a shell from a plain writable object, with no color, and max verbosity.

fn print(
&mut self,
status: &dyn Display,
message: Option<&dyn Display>,
color: Color,
justified: bool
) -> CargoResult<()>

Prints a message, where the status will have color color, and can be justified. The messages follows without color.

pub fn set_needs_clear(&mut self, needs_clear: bool)

Sets whether the next print should clear the current line.

pub fn is_cleared(&self) -> bool

Returns true if the needs_clear flag is unset.

pub fn err_width(&self) -> TtyWidth

Returns the width of the terminal in spaces, if any.

pub fn is_err_tty(&self) -> bool

Returns true if stderr is a tty.

pub fn out(&mut self) -> &mut dyn Write

Gets a reference to the underlying stdout writer.

pub fn err(&mut self) -> &mut dyn Write

Gets a reference to the underlying stderr writer.

pub fn err_erase_line(&mut self)

Erase from cursor to end of line.

pub fn status<T, U>(&mut self, status: T, message: U) -> CargoResult<()>where
T: Display,
U: Display,

Shortcut to right-align and color green a status message.

pub fn status_header<T>(&mut self, status: T) -> CargoResult<()>where
T: Display,

pub fn status_with_color<T, U>(
&mut self,
status: T,
message: U,
color: Color
) -> CargoResult<()>where
T: Display,
U: Display,

Shortcut to right-align a status message.

pub fn verbose<F>(&mut self, callback: F) -> CargoResult<()>where
F: FnMut(&mut Shell) -> CargoResult<()>,

Runs the callback only if we are in verbose mode.

pub fn concise<F>(&mut self, callback: F) -> CargoResult<()>where
F: FnMut(&mut Shell) -> CargoResult<()>,

Runs the callback if we are not in verbose mode.

pub fn error<T: Display>(&mut self, message: T) -> CargoResult<()>

Prints a red ‘error’ message.

pub fn warn<T: Display>(&mut self, message: T) -> CargoResult<()>

Prints an amber ‘warning’ message.

pub fn note<T: Display>(&mut self, message: T) -> CargoResult<()>

Prints a cyan ‘note’ message.

pub fn set_verbosity(&mut self, verbosity: Verbosity)

Updates the verbosity of the shell.

pub fn verbosity(&self) -> Verbosity

Gets the verbosity of the shell.

pub fn set_color_choice(&mut self, color: Option<&str>) -> CargoResult<()>

Updates the color choice (always, never, or auto) from a string..

pub fn color_choice(&self) -> ColorChoice

Gets the current color choice.

If we are not using a color stream, this will always return Never, even if the color choice has been set to something else.

pub fn err_supports_color(&self) -> bool

Whether the shell supports color.

pub fn out_supports_color(&self) -> bool

pub fn write_stdout(
&mut self,
fragment: impl Display,
color: &ColorSpec
) -> CargoResult<()>

Write a styled fragment

Caller is responsible for deciding whether Shell::verbosity is affects output.

pub fn write_stderr(
&mut self,
fragment: impl Display,
color: &ColorSpec
) -> CargoResult<()>

Write a styled fragment

Caller is responsible for deciding whether Shell::verbosity is affects output.

pub fn print_ansi_stderr(&mut self, message: &[u8]) -> CargoResult<()>

Prints a message to stderr and translates ANSI escape code into console colors.

pub fn print_ansi_stdout(&mut self, message: &[u8]) -> CargoResult<()>

Prints a message to stdout and translates ANSI escape code into console colors.

pub fn print_json<T: Serialize>(&mut self, obj: &T) -> CargoResult<()>