Struct cargo::core::SourceId

pub struct SourceId { /* fields omitted */ }

Unique identifier for a source of packages.

Methods

impl SourceId

pub fn from_url(string: &str) -> CargoResult<SourceId>

Parses a source URL and returns the corresponding ID.

Example

use cargo::core::SourceId;
SourceId::from_url("git+https://github.com/alexcrichton/\
                    libssh2-static-sys#80e71a3021618eb05\
                    656c58fb7c5ef5f12bc747f");

pub fn to_url(&self) -> SourceIdToUrl

A view of the SourceId that can be Displayed as a URL

pub fn for_path(path: &Path) -> CargoResult<SourceId>

Create a SourceId from a filesystem path.

Pass absolute path

pub fn for_git(url: &Url, reference: GitReference) -> CargoResult<SourceId>

Crate a SourceId from a git reference

pub fn for_registry(url: &Url) -> CargoResult<SourceId>

Create a SourceId from a registry url

pub fn for_local_registry(path: &Path) -> CargoResult<SourceId>

Create a SourceId from a local registry path

pub fn for_directory(path: &Path) -> CargoResult<SourceId>

Create a SourceId from a directory path

pub fn crates_io(config: &Config) -> CargoResult<SourceId>

Returns the SourceId corresponding to the main repository.

This is the main cargo registry by default, but it can be overridden in a .cargo/config.

pub fn alt_registry(config: &Config, key: &str) -> CargoResult<SourceId>

pub fn url(&self) -> &Url

Get this source URL

pub fn display_registry(&self) -> String

pub fn is_path(&self) -> bool

Is this source from a filesystem path

pub fn is_registry(&self) -> bool

Is this source from a registry (either local or not)

pub fn is_alt_registry(&self) -> bool

Is this source from an alternative registry

pub fn is_git(&self) -> bool

Is this source from a git repository

pub fn load<'a>(&self, config: &'a Config) -> CargoResult<Box<Source + 'a>>

Creates an implementation of Source corresponding to this ID.

pub fn precise(&self) -> Option<&str>

Get the value of the precise field

pub fn git_reference(&self) -> Option<&GitReference>

Get the git reference if this is a git source, otherwise None.

pub fn with_precise(&self, v: Option<String>) -> SourceId

Create a new SourceId from this source with the given precise

pub fn is_default_registry(&self) -> bool

Whether the remote registry is the standard https://crates.io

pub fn stable_hash<S: Hasher>(&self, workspace: &Path, into: &mut S)

Hash self

For paths, remove the workspace prefix so the same source will give the same hash in different locations.