Module cargo::core::summary

Structs

Inner 🔒

Summary

Subset of a Manifest. Contains only the most important information about a package.

Enums

FeatureValue

FeatureValue represents the types of dependencies a feature can have.

Functions

build_feature_map 🔒

Checks features for errors, bailing out a CargoResult:Err if invalid, and creates FeatureValues for each feature.

validate_feature_name 🔒

Type Definitions

FeatureMap