Struct cargo::core::summary::Summary

pub struct Summary {
    inner: Rc<Inner>,
}

Subset of a Manifest. Contains only the most important information about a package.

Summaries are cloned, and should not be mutated after creation

Fields
inner: Rc<Inner>
Implementations

impl Summary

pub fn new(
config: &Config,
pkg_id: PackageId,
dependencies: Vec<Dependency>,
features: &BTreeMap<InternedString, Vec<InternedString>>,
links: Option<impl Into<InternedString>>
) -> CargoResult<Summary>

pub fn package_id(&self) -> PackageId

pub fn name(&self) -> InternedString

pub fn version(&self) -> &Version

pub fn source_id(&self) -> SourceId

pub fn dependencies(&self) -> &[Dependency]

pub fn features(&self) -> &FeatureMap

pub fn checksum(&self) -> Option<&str>

pub fn links(&self) -> Option<InternedString>

pub fn override_id(self, id: PackageId) -> Summary

pub fn set_checksum(&mut self, cksum: String)

pub fn map_dependencies<F>(self, f: F) -> Summarywhere
F: FnMut(Dependency) -> Dependency,

pub fn map_source(self, to_replace: SourceId, replace_with: SourceId) -> Summary