Module cargo::ops::cargo_compile::compile_filter

Filters and their rules to select which Cargo targets will be built.

Enums

CompileFilter

Filter to apply to the root package to select which Cargo targets will be built. (examples, bins, benches, tests, …)

FilterRule

Indicates which Cargo targets will be selected to be built.

LibRule

Indicates whether or not the library target gets included.