Enum cargo::ops::cargo_compile::packages::Packages

pub enum Packages {
    Default,
    All,
    OptOut(Vec<String>),
    Packages(Vec<String>),
}

Represents the selected pacakges that will be built.

Generally, it represents the combination of all -p flag. When working within a workspace, --exclude and --workspace flags also contribute to it.

Variants

Default

Pacakges selected by default. Ususally means no flag provided.

All

Opt in all packages.

As of the time of this writing, it only works on opting in all workspace members.

OptOut(Vec<String>)

Opt out of packages passed in.

As of the time of this writing, it only works on opting out workspace members.

Packages(Vec<String>)

A sequence of hand-picked packages that will be built. Normally done by -p flag.

Implementations

impl Packages

pub fn from_flags(
all: bool,
exclude: Vec<String>,
package: Vec<String>
) -> CargoResult<Self>

Creates a Packages from flags which are generally equivalent to command line flags.

pub fn to_package_id_specs(
&self,
ws: &Workspace<'_>
) -> CargoResult<Vec<PackageIdSpec>>

Converts selected packages to PackageIdSpecs.

pub fn get_packages<'ws>(
&self,
ws: &'ws Workspace<'_>
) -> CargoResult<Vec<&'ws Package>>

Gets a list of selected Packages.

pub fn needs_spec_flag(&self, ws: &Workspace<'_>) -> bool

Returns whether or not the user needs to pass a -p flag to target a specific package in the workspace.