Module cargo::ops::cargo_new

Structs

CargoNewConfig 🔒

IgnoreList 🔒

IgnoreList

MkOptions 🔒

NewOptions

SourceFileInformation 🔒

Enums

NewProjectKind

VersionControl

Functions

calculate_new_project_kind 🔒

check_name 🔒

check_path 🔒

Checks if the path contains any invalid PATH env characters.

detect_source_paths_and_types 🔒

get_name 🔒

init

init_vcs 🔒

Initializes the correct VCS system based on the provided config.

mk 🔒

new

plan_new_source_file 🔒

write_ignore_file 🔒

Writes the ignore file to the given directory. If the ignore file for the given vcs system already exists, its content is read and duplicate ignore file entries are filtered out.