Module cargo::ops::cargo_package

Structs

ArchiveFile ğŸ”’

GitVcsInfo ğŸ”’

PackageOpts

VcsInfo ğŸ”’

Enums

FileContents ğŸ”’

GeneratedFile ğŸ”’

Constants

ORIGINAL_MANIFEST_FILE ğŸ”’

VCS_INFO_FILE ğŸ”’

Functions

build_ar_list ğŸ”’

Builds list of files to archive.

build_lock ğŸ”’

Construct Cargo.lock for the package to be published.

check_filename ğŸ”’

check_for_file_and_add ğŸ”’

check_metadata ğŸ”’

check_repo_state ğŸ”’

Checks if the package source is in a git DVCS repository. If git, and the source is dirty (e.g., has uncommitted changes) then bail! with an informative message. Otherwise return the sha1 hash of the current HEAD commit, or None if no repo is found.

check_yanked

compare_resolve ğŸ”’

Generate warnings when packaging Cargo.lock, and the resolve have changed.

hash_all ğŸ”’

package

package_one

report_hash_difference ğŸ”’

run_verify ğŸ”’

tar ğŸ”’

Compresses and packages a list of ArchiveFiles and writes into the given file.