Module cargo::ops::cargo_test

Structs

TestOptions

UnitTestError ğŸ”’

A unit that failed to run.

Enums

TestKind ğŸ”’

The kind of test.

Functions

cmd_builds ğŸ”’

Creates a ProcessBuilder for executing a single test.

compile_tests ğŸ”’

display_no_run_information ğŸ”’

Displays human-readable descriptions of the test executables.

fail_fast_code ğŸ”’

Returns the error code to use when not using --no-fail-fast.

no_fail_fast_err ğŸ”’

Returns the CliError when using --no-fail-fast and there is at least one error.

report_test_error ğŸ”’

Displays an error on the console about a test failure.

run_benches

Compiles and runs benchmarks.

run_doc_tests ğŸ”’

Runs doc tests.

run_tests

Compiles and runs tests.

run_unit_tests ğŸ”’

Runs the unit and integration tests of a package.