Function cargo::ops::cargo_test::fail_fast_code

fn fail_fast_code(error: &Error) -> i32

Returns the error code to use when not using --no-fail-fast.

Cargo will return the error code from the test process itself. If some other error happened (like a failure to launch the process), then it will return a standard 101 error code.

When using --no-fail-fast, Cargo always uses the 101 exit code (since there may not be just one process to report).