Function cargo::ops::cargo_test::run_doc_tests

fn run_doc_tests(
    ws: &Workspace<'_>,
    options: &TestOptions,
    test_args: &[&str],
    compilation: &Compilation<'_>
) -> Result<Vec<UnitTestError>, CliError>

Runs doc tests.

Returns a Vec of tests that failed when --no-fail-fast is used. If --no-fail-fast is not used, then this returns an Err.