cargo::ops

Module cargo_update

Structs

PackageChange ğŸ”’

PackageDiff

All resolved versions of a package name within a SourceId

UpdateOptions

Enums

PackageChangeKind ğŸ”’

Functions

annotate_required_rust_version ğŸ”’

fill_with_deps ğŸ”’

generate_lockfile

is_latest ğŸ”’

print_lockfile_changes

Prints lockfile change statuses.

print_lockfile_generation ğŸ”’

print_lockfile_sync ğŸ”’

print_lockfile_updates ğŸ”’

report_latest ğŸ”’

report_required_rust_version ğŸ”’

required_rust_version ğŸ”’

status_locking ğŸ”’

update_lockfile

upgrade_dependency ğŸ”’

upgrade_manifests

write_manifest_upgrades

Update manifests with upgraded versions, and write to disk. Based on cargo-edit. Returns true if any file has changed.

Type Aliases

UpgradeMap