cargo::ops::common_for_install_and_uninstall
Struct CrateListingV2

struct CrateListingV2 {
    installs: BTreeMap<PackageId, InstallInfo>,
    other: BTreeMap<String, Value>,
}

Tracking information for the set of installed packages.

Fields
installs: BTreeMap<PackageId, InstallInfo>

Map of every installed package.

other: BTreeMap<String, Value>

Forwards compatibility. Unknown keys from future versions of Cargo will be stored here and retained when the file is saved.

Implementations

impl CrateListingV2

fn sync_v1(&mut self, v1: &CrateListingV1)

Incorporate any changes from v1 into self. This handles the initial upgrade to v2, and handles the case where v2 is in use, and a v1 update is made, then v2 is used again. i.e., cargo +new install foo ; cargo +old install bar ; cargo +new install bar For now, v1 is the source of truth, so its values are trusted over v2.

fn package_for_bin(&self, bin_name: &str) -> Option<PackageId>

fn mark_installed( &mut self, pkg: &Package, bins: &BTreeSet<String>, version_req: Option<String>, opts: &CompileOptions, target: &str, rustc: &str, )

fn remove(&mut self, pkg_id: PackageId, bins: &BTreeSet<String>)

fn remove_bin(&mut self, pkg_id: PackageId, bin: &str)

fn save(&self, lock: &FileLock) -> CargoResult<()>