Enum cargo::ops::CompileFilter

pub enum CompileFilter {
    Default {
        required_features_filterable: bool,
    },
    Only {
        all_targets: bool,
        lib: LibRule,
        bins: FilterRule,
        examples: FilterRule,
        tests: FilterRule,
        benches: FilterRule,
    },
}

Filter to apply to the root package to select which Cargo targets will be built. (examples, bins, benches, tests, …)

The actual filter process happens inside generate_root_units.

Not to be confused with Packages, which opts in packages to be built.

Variants

Default

Fields

required_features_filterable: bool

Flag whether targets can be safely skipped when required-features are not satisfied.

The default set of Cargo targets.

Only

Fields

all_targets: bool

Include all Cargo targets.

lib: LibRule

bins: FilterRule

examples: FilterRule

tests: FilterRule

benches: FilterRule

Only includes a subset of all Cargo targets.

Implementations

impl CompileFilter

pub fn from_raw_arguments( lib_only: bool, bins: Vec<String>, all_bins: bool, tsts: Vec<String>, all_tsts: bool, exms: Vec<String>, all_exms: bool, bens: Vec<String>, all_bens: bool, all_targets: bool ) -> CompileFilter

Constructs a filter from raw command line arguments.

pub fn new( rule_lib: LibRule, rule_bins: FilterRule, rule_tsts: FilterRule, rule_exms: FilterRule, rule_bens: FilterRule ) -> CompileFilter

Constructs a filter from underlying primitives.

pub fn new_all_targets() -> CompileFilter

Constructs a filter that includes all targets.

pub fn all_test_targets() -> Self

Constructs a filter that includes all test targets.

Being different from the behavior of CompileFilter::Default, this function only recognizes test targets, which means cargo might compile all targets with tested flag on, whereas CompileFilter::Default may include additional example targets to ensure they can be compiled.

Note that the actual behavior is subject to filter_default_targets and generate_root_units though.

pub fn lib_only() -> Self

Constructs a filter that includes lib target only.

pub fn single_bin(bin: String) -> Self

Constructs a filter that includes the given binary. No more. No less.

pub fn need_dev_deps(&self, mode: CompileMode) -> bool

Indicates if Cargo needs to build any dev dependency.

pub fn target_run(&self, target: &Target) -> bool

Selects targets for “cargo run”. for logic to select targets for other subcommands, see generate_root_units and filter_default_targets.

pub fn is_specific(&self) -> bool

pub fn is_all_targets(&self) -> bool

pub(crate) fn contains_glob_patterns(&self) -> bool

Checks if any specified target name contains glob patterns.