Function cargo::ops::resolve_with_previous

pub fn resolve_with_previous<'a>(
    registry: &mut PackageRegistry, 
    ws: &Workspace, 
    method: Method, 
    previous: Option<&'a Resolve>, 
    to_avoid: Option<&HashSet<&'a PackageId>>, 
    specs: &[PackageIdSpec], 
    warn: bool
) -> CargoResult<Resolve>

Resolve all dependencies for a package using an optional previous instance of resolve to guide the resolution process.

This also takes an optional hash set, to_avoid, which is a list of package ids that should be avoided when consulting the previous instance of resolve (often used in pairings with updates).

The previous resolve normally comes from a lockfile. This function does not read or write lockfiles from the filesystem.