Function cargo::ops::resolve_ws

pub fn resolve_ws<'a>(
    ws: &Workspace<'a>
) -> CargoResult<(PackageSet<'a>, Resolve)>

Resolves all dependencies for the workspace using the previous lock file as a guide if present.

This function will also write the result of resolution as a new lock file (unless it is an ephemeral workspace such as cargo install or cargo package).

This is a simple interface used by commands like clean, fetch, and package, which don’t specify any options or features.