Function cargo::ops::resolve::resolve_ws_with_opts

pub fn resolve_ws_with_opts<'cfg>(
    ws: &Workspace<'cfg>,
    target_data: &RustcTargetData<'cfg>,
    requested_targets: &[CompileKind],
    cli_features: &CliFeatures,
    specs: &[PackageIdSpec],
    has_dev_units: HasDevUnits,
    force_all_targets: ForceAllTargets
) -> CargoResult<WorkspaceResolve<'cfg>>

Resolves dependencies for some packages of the workspace, taking into account paths overrides and activated features.

This function will also write the result of resolution as a new lock file (unless Workspace::require_optional_deps is false, such as cargo install or -Z avoid-dev-deps), or it is an ephemeral workspace (cargo install or cargo package).

specs may be empty, which indicates it should resolve all workspace members. In this case, opts.all_features must be true.