Struct cargo::ops::BuildOutput

pub struct BuildOutput {
    pub library_paths: Vec<PathBuf>,
    pub library_links: Vec<String>,
    pub cfgs: Vec<String>,
    pub metadata: Vec<(String, String)>,
    pub rerun_if_changed: Vec<String>,
    pub warnings: Vec<String>,
}

Contains the parsed output of a custom build script.

Fields
library_paths: Vec<PathBuf>

Paths to pass to rustc with the -L flag

library_links: Vec<String>

Names and link kinds of libraries, suitable for the -l flag

cfgs: Vec<String>

Various --cfg flags to pass to the compiler

metadata: Vec<(String, String)>

Metadata to pass to the immediate dependencies

rerun_if_changed: Vec<String>

Glob paths to trigger a rerun of this build script.

warnings: Vec<String>

Warnings generated by this build,

Methods

impl BuildOutput

fn parse_file(path: &Path, pkg_name: &str) -> CargoResult<BuildOutput>

fn parse(input: &[u8], pkg_name: &str) -> CargoResult<BuildOutput>

fn parse_rustc_flags(value: &str, whence: &str) -> CargoResult<(Vec<PathBuf>, Vec<String>)>