Trait cargo::ops::Executor

pub trait Executor: Send + Sync + 'static {
    fn init(&self, _cx: &Context, _unit: &Unit) { ... }

    fn exec(
        &self, 
        cmd: ProcessBuilder, 
        _id: &PackageId, 
        _target: &Target
    ) -> CargoResult<()> { ... }

    fn exec_json(
        &self, 
        cmd: ProcessBuilder, 
        _id: &PackageId, 
        _target: &Target, 
        handle_stdout: &mut FnMut(&str) -> CargoResult<()>, 
        handle_stderr: &mut FnMut(&str) -> CargoResult<()>
    ) -> CargoResult<()> { ... }

    fn force_rebuild(&self, _unit: &Unit) -> bool { ... }
}

A glorified callback for executing calls to rustc. Rather than calling rustc directly, we'll use an Executor, giving clients an opportunity to intercept the build calls.

Provided Methods

fn init(&self, _cx: &Context, _unit: &Unit)

Called after a rustc process invocation is prepared up-front for a given unit of work (may still be modified for runtime-known dependencies, when the work is actually executed).

fn exec(
&self,
cmd: ProcessBuilder,
_id: &PackageId,
_target: &Target
) -> CargoResult<()>

In case of an Err, Cargo will not continue with the build process for this package.

fn exec_json(
&self,
cmd: ProcessBuilder,
_id: &PackageId,
_target: &Target,
handle_stdout: &mut FnMut(&str) -> CargoResult<()>,
handle_stderr: &mut FnMut(&str) -> CargoResult<()>
) -> CargoResult<()>

fn force_rebuild(&self, _unit: &Unit) -> bool

Queried when queuing each unit of work. If it returns true, then the unit will always be rebuilt, independent of whether it needs to be.