Module cargo::ops::tree

Implementation of cargo tree.

Modules

format 🔒

graph 🔒

Code for building the graph used by cargo tree.

Structs

Symbols 🔒

TreeOptions

Enums

Charset

EdgeKind

The kind of edge, for separating dependencies into different sections.

Node

Prefix

Target

Statics

ASCII_SYMBOLS 🔒

UTF8_SYMBOLS 🔒

Functions

build_and_print

Entry point for the cargo tree command.

print 🔒

Prints a tree for each given root.

print_dependencies 🔒

Prints all the dependencies of a package for the given dependency kind.

print_node 🔒

Prints a package and all of its dependencies.