Module cargo::ops::tree::graph

Code for building the graph used by cargo tree.

Structs

Edges ğŸ”’

Set of outgoing edges for a single node.

Graph

A graph of dependencies.

Enums

EdgeKind

The kind of edge, for separating dependencies into different sections.

Node

Functions

add_cli_features ğŸ”’

Adds nodes for features requested on the command-line for the given member.

add_feature ğŸ”’

Adds a feature node between two nodes.

add_feature_rec ğŸ”’

Recursively add feature nodes for all features enabled by the given feature.

add_internal_features ğŸ”’

Recursively adds connections between features in the [features] table for every package.

add_pkg ğŸ”’

Adds a single package node (if it does not already exist).

build

Builds the graph.