cargo::sources

Module config

Implementation of configuration for various sources.

This module will parse the various source.* TOML configuration keys into a structure usable by Cargo itself. Currently this is primarily used to map sources to one another via the replace-with key in .cargo/config.

Structs

SourceConfig 🔒

Configuration for a particular source, found in TOML looking like:

SourceConfigDef 🔒

Definition of a source in a config file.

SourceConfigMap

Represents the entire [source] replacement table in Cargo configuration.