Module cargo::sources::git::oxide

This module contains all code sporting gitoxide for operations on git repositories and it mirrors utils closely for now. One day it can be renamed into utils once git2 isn’t required anymore.

Enums

OpenMode

The reason we are opening a git repository.

Functions

amend_authentication_hints 🔒

cargo_config_to_gitoxide_overrides

Convert git related cargo configuration into the respective git configuration which can be used when opening new repositories.

open_repo

Produce a repository with everything pre-configured according to config. Most notably this includes transport configuration. Knowing its purpose helps to optimize the way we open the repository. Use config_overrides to configure the new repository.

reinitialize

Reinitializes a given Git repository. This is useful when a Git repoistory seems corrupted and we want to start over.

translate_progress_to_bar 🔒

with_retry_and_progress

For the time being, repo_path makes it easy to instantiate a gitoxide repo just for fetching. In future this may change to be the gitoxide repository itself.