Module cargo::sources::git::source

See GitSource.

Structs

GitSource

GitSource contains one or more packages gathering from a Git repository. Under the hood it uses RecursivePathSource to discover packages inside the repository.

Enums

Revision 🔒

Indicates a Git revision that might be locked or deferred to be resolved.

Functions

ident 🔒

Create an identifier from a URL, essentially turning proto://host/path/repo into repo-<hash-of-url>.

ident_shallow 🔒

Like ident(), but appends -shallow to it, turning proto://host/path/repo into repo-<hash-of-url>-shallow.