Struct cargo::sources::git::source::GitSource

pub struct GitSource<'cfg> {
    remote: GitRemote,
    manifest_reference: GitReference,
    locked_rev: Option<Oid>,
    source_id: SourceId,
    path_source: Option<PathSource<'cfg>>,
    ident: String,
    config: &'cfg Config,
    quiet: bool,
}

GitSource contains one or more packages gathering from a Git repository. Under the hood it uses PathSource to discover packages inside the repository.

Filesystem layout

During a successful GitSource download, at least two Git repositories are created: one is the shared Git database of this remote, and the other is the Git checkout to a specific revision, which contains the actual files to be compiled. Multiple checkouts can be cloned from a single Git database.

Those repositories are located at Cargo’s Git cache directory $CARGO_HOME/git. The file tree of the cache directory roughly looks like:

$CARGO_HOME/git/
├── checkouts/
│  ├── gimli-a0d193bd15a5ed96/
│  │  ├── 8e73ef0/     # Git short ID for a certain revision
│  │  ├── a2a4b78/
│  │  └── e33d1ac/
│  ├── log-c58e1db3de7c154d-shallow/
│  │  └── 11eda98/
└── db/
   ├── gimli-a0d193bd15a5ed96/
   └── log-c58e1db3de7c154d-shallow/

For more on Git cache directory, see “Cargo Home” in The Cargo Book.

For more on the directory format <pkg>-<hash>[-shallow], see ident and ident_shallow.

Locked to a revision

Once a GitSource is fetched, it will resolve to a specific commit revision. This is often mentioned as “locked revision” (locked_rev) throughout the codebase. The revision is written into Cargo.lock. This is essential since we want to ensure a package can compiles with the same set of files when a Cargo.lock is present. With the locked_rev provided, GitSource can precisely fetch the same revision from the Git repository.

Fields
remote: GitRemote

The git remote which we’re going to fetch from.

manifest_reference: GitReference

The Git reference from the manifest file.

locked_rev: Option<Oid>

The revision which a git source is locked to. This is expected to be set after the Git repository is fetched.

source_id: SourceId

The unique identifier of this source.

path_source: Option<PathSource<'cfg>>

The underlying path source to discover packages inside the Git repository.

ident: String

The identifer of this source for Cargo’s Git cache directory. See ident for more.

config: &'cfg Configquiet: bool

Disables status messages.

Implementations

impl<'cfg> GitSource<'cfg>

pub fn new( source_id: SourceId, config: &'cfg Config ) -> CargoResult<GitSource<'cfg>>

Creates a git source for the given SourceId.

pub fn url(&self) -> &Url

Gets the remote repository URL.

pub fn read_packages(&mut self) -> CargoResult<Vec<Package>>

Returns the packages discovered by this source. It may fetch the Git repository as well as walk the filesystem if package informations haven’t yet updated.