cargo::sources::path
Function list_files

pub fn list_files(
    pkg: &Package,
    gctx: &GlobalContext,
) -> CargoResult<Vec<PathBuf>>

List all files relevant to building this package inside this source.

This function will use the appropriate methods to determine the set of files underneath this source’s directory which are relevant for building pkg.

The basic assumption of this method is that all files in the directory are relevant for building this package, but it also contains logic to use other methods like .gitignore, package.include, or package.exclude to filter the list of files.