cargo::sources::registry
Enum IndexSummary

pub enum IndexSummary {
    Candidate(Summary),
    Yanked(Summary),
    Offline(Summary),
    Unsupported(Summary, u32),
    Invalid(Summary),
}

A parsed representation of a summary from the index. This is usually parsed from a line from a raw index file, or a JSON blob from on-disk index cache.

In addition to a full Summary, we have information on whether it is yanked.

Variants

Candidate(Summary)

Available for consideration

Yanked(Summary)

Yanked within its registry

Offline(Summary)

Not available as we are offline and create is not downloaded yet

Unsupported(Summary, u32)

From a newer schema version and is likely incomplete or inaccurate

Invalid(Summary)

An error was encountered despite being a supported schema version

Implementations

impl IndexSummary

pub fn as_summary(&self) -> &Summary

Extract the summary from any variant

pub fn into_summary(self) -> Summary

Extract the summary from any variant

pub fn map_summary(self, f: impl Fn(Summary) -> Summary) -> Self

pub fn package_id(&self) -> PackageId

Extract the package id from any variant

pub fn is_yanked(&self) -> bool

Returns true if the index summary is Yanked.

pub fn is_offline(&self) -> bool

Returns true if the index summary is Offline.

impl IndexSummary

fn parse( line: &[u8], source_id: SourceId, bindeps: bool, ) -> CargoResult<IndexSummary>

Parses a line from the registryโs index file into an IndexSummary for a package.

The line provided is expected to be valid JSON. It is supposed to be a IndexPackage.