cargo::sources::registry::index
Struct IndexPackage

pub struct IndexPackage<'a> {
    pub name: InternedString,
    pub vers: Version,
    pub deps: Vec<RegistryDependency<'a>>,
    pub features: BTreeMap<InternedString, Vec<InternedString>>,
    pub features2: Option<BTreeMap<InternedString, Vec<InternedString>>>,
    pub cksum: String,
    pub yanked: Option<bool>,
    pub links: Option<InternedString>,
    pub rust_version: Option<RustVersion>,
    pub v: Option<u32>,
}

A single line in the index representing a single version of a package.

Fields
name: InternedString

Name of the package.

vers: Version

The version of this dependency.

deps: Vec<RegistryDependency<'a>>

All kinds of direct dependencies of the package, including dev and build dependencies.

features: BTreeMap<InternedString, Vec<InternedString>>

Set of features defined for the package, i.e., [features] table.

features2: Option<BTreeMap<InternedString, Vec<InternedString>>>

This field contains features with new, extended syntax. Specifically, namespaced features (dep:) and weak dependencies (pkg?/feat).

This is separated from features because versions older than 1.19 will fail to load due to not being able to parse the new syntax, even with a Cargo.lock file.

cksum: String

Checksum for verifying the integrity of the corresponding downloaded package.

yanked: Option<bool>

If true, Cargo will skip this version when resolving.

This was added in 2014. Everything in the crates.io index has this set now, so this probably doesn’t need to be an option anymore.

links: Option<InternedString>

Native library name this package links to.

Added early 2018 (see https://github.com/rust-lang/cargo/pull/4978), can be None if published before then.

rust_version: Option<RustVersion>

Required version of rust

Corresponds to package.rust-version.

Added in 2023 (see https://github.com/rust-lang/crates.io/pull/6267), can be None if published before then or if not set in the manifest.

v: Option<u32>

The schema version for this entry.

If this is None, it defaults to version 1. Entries with unknown versions are ignored.

Version 2 schema adds the features2 field.

Version 3 schema adds artifact, bindep_targes, and lib for artifact dependencies support.

This provides a method to safely introduce changes to index entries and allow older versions of cargo to ignore newer entries it doesn’t understand. This is honored as of 1.51, so unfortunately older versions will ignore it, and potentially misinterpret version 2 and newer entries.

The intent is that versions older than 1.51 will work with a pre-existing Cargo.lock, but they may not correctly process cargo update or build a lock from scratch. In that case, cargo may incorrectly select a new package that uses a new index schema. A workaround is to downgrade any packages that are incompatible with the --precise flag of cargo update.

Implementations

impl IndexPackage<'_>

fn to_summary(&self, source_id: SourceId) -> CargoResult<Summary>