cargo::sources::registry::local
Struct LocalRegistry

pub struct LocalRegistry<'gctx> {
    index_path: Filesystem,
    root: Filesystem,
    src_path: Filesystem,
    gctx: &'gctx GlobalContext,
    updated: bool,
    quiet: bool,
}

A local registry is a registry that lives on the filesystem as a set of .crate files with an index directory in the same format as a remote registry.

This type is primarily accessed through the RegistryData trait.

When a local registry is requested for a package, it simply looks into what its index has under the index directory. When LocalRegistry::download is called, a local registry verifies the checksum of the requested .crate tarball and then unpacks it to $CARGO_HOME/.registry/src.

Note that there is a third-party subcommand cargo-local-registry, which happened to be developed by a former Cargo team member when local registry was introduced. The tool is to ease the burden of maintaining local registries. However, in general the Cargo team avoids recommending any specific third-party crate. Just FYI.

Filesystem hierarchy

Here is an example layout of a local registry on a local filesystem:

[registry root]/
├── index/                      # registry index
│  ├── an/
│  │  └── yh/
│  │     └── anyhow
│  ├── ru/
│  │  └── st/
│  │     ├── rustls
│  │     └── rustls-ffi
│  └── se/
│     └── mv/
│        └── semver
├── anyhow-1.0.71.crate         # pre-downloaded crate tarballs
├── rustls-0.20.8.crate
├── rustls-ffi-0.8.2.crate
└── semver-1.0.17.crate

For general concepts of registries, see the module-level documentation.

Fields
index_path: Filesystem

Path to the registry index.

root: Filesystem

Root path of this local registry.

src_path: Filesystem

Path where this local registry extract .crate tarballs to.

gctx: &'gctx GlobalContextupdated: bool

Whether this source has updated all package information it may contain.

quiet: bool

Disables status messages.

Implementations

impl<'gctx> LocalRegistry<'gctx>

pub fn new( root: &Path, gctx: &'gctx GlobalContext, name: &str, ) -> LocalRegistry<'gctx>

Creates a local registry at root.

name — Name of a path segment where .crate tarballs are stored. Expect to be unique.