Struct cargo::sources::registry::remote::RemoteRegistry

pub struct RemoteRegistry<'cfg> {
    index_path: Filesystem,
    cache_path: Filesystem,
    source_id: SourceId,
    index_git_ref: GitReference,
    config: &'cfg Config,
    tree: RefCell<Option<Tree<'static>>>,
    repo: LazyCell<Repository>,
    head: Cell<Option<Oid>>,
    current_sha: Cell<Option<InternedString>>,
    needs_update: bool,
    quiet: bool,
}

A remote registry is a registry that lives at a remote URL (such as crates.io). The git index is cloned locally, and .crate files are downloaded as needed and cached locally.

This type is primarily accessed through the RegistryData trait.

See the module-level documentation for the index format and layout.

History of Git-based index registry

Using Git to host this index used to be quite efficient. The full index can be stored efficiently locally on disk, and once it is downloaded, all queries of a registry can happen locally and needn’t touch the network. Git-based index was a reasonable design choice at the time when HTTP/2 was just introduced.

However, the full index keeps growing as crates.io grows. It becomes relatively big and slows down the first use of Cargo. Git (specifically libgit2) is not efficient at handling huge amounts of small files either. On the other hand, newer protocols like HTTP/2 are prevalent and capable to serve a bunch of tiny files. Today, it is encouraged to use HttpRegistry, which is the default from 1.70.0. That being said, Cargo will continue supporting Git-based index for a pretty long while.

Fields
index_path: Filesystem

Path to the registry index ($CARGO_HOME/registry/index/$REG-HASH).

cache_path: Filesystem

Path to the cache of .crate files ($CARGO_HOME/registry/cache/$REG-HASH).

source_id: SourceId

The unique identifier of this registry source.

index_git_ref: GitReference

This reference is stored so that when a registry needs update, it knows where to fetch from.

config: &'cfg Configtree: RefCell<Option<Tree<'static>>>

A Git tree object to help this registry find crate metadata from the underlying Git repository.

This is stored here to prevent Git from repeatly creating a tree object during each call into load().

repo: LazyCell<Repository>

A Git repository that contains the actual index we want.

head: Cell<Option<Oid>>

The current HEAD commit of the underlying Git repository.

current_sha: Cell<Option<InternedString>>

This stores sha value of the current HEAD commit for convenience.

needs_update: bool

Whether this registry needs to update package informations.

See RemoteRegistry::mark_updated on how to make sure a registry index is updated only once per session.

quiet: bool

Disables status messages.

Implementations

impl<'cfg> RemoteRegistry<'cfg>

pub fn new( source_id: SourceId, config: &'cfg Config, name: &str ) -> RemoteRegistry<'cfg>

Creates a Git-rebased remote registry for source_id.

name — Name of a path segment where .crate tarballs and the registry index are stored. Expect to be unique.

fn repo(&self) -> CargoResult<&Repository>

Creates intermediate dirs and initialize the repository.

fn head(&self) -> CargoResult<Oid>

Get the object ID of the HEAD commit from the underlying Git repository.

fn tree(&self) -> CargoResult<Ref<'_, Tree<'_>>>

Returns a git2::Tree object of the current HEAD commit of the underlying Git repository.

fn current_version(&self) -> Option<InternedString>

Gets the current version of the registry index.

It is usually sha of the HEAD commit from the underlying Git repository.

fn is_updated(&self) -> bool

Whether the registry is up-to-date. See Self::mark_updated for more.

fn mark_updated(&self)

Marks this registry as up-to-date.

This makes sure the index is only updated once per session since it is an expensive operation. This generally only happens when the resolver is run multiple times, such as during cargo publish.