Struct cargo::sources::registry::RegistrySource

pub struct RegistrySource<'cfg> {
    source_id: SourceId,
    src_path: Filesystem,
    config: &'cfg Config,
    ops: Box<dyn RegistryData + 'cfg>,
    index: RegistryIndex<'cfg>,
    yanked_whitelist: HashSet<PackageId>,
}

A “source” for a local (see local::LocalRegistry) or remote (see remote::RemoteRegistry) registry.

This contains common functionality that is shared between the two registry kinds, with the registry-specific logic implemented as part of the RegistryData trait referenced via the ops field.

Fields
source_id: SourceIdsrc_path: Filesystem

The path where crate files are extracted ($CARGO_HOME/registry/src/$REG-HASH).

config: &'cfg Config

Local reference to Config for convenience.

ops: Box<dyn RegistryData + 'cfg>

Abstraction for interfacing to the different registry kinds.

index: RegistryIndex<'cfg>

Interface for managing the on-disk index.

yanked_whitelist: HashSet<PackageId>

A set of packages that should be allowed to be used, even if they are yanked.

This is populated from the entries in Cargo.lock to ensure that cargo update -p somepkg won’t unlock yanked entries in Cargo.lock. Otherwise, the resolver would think that those entries no longer exist, and it would trigger updates to unrelated packages.

Implementations

impl<'cfg> RegistrySource<'cfg>

pub fn remote(
source_id: SourceId,
yanked_whitelist: &HashSet<PackageId>,
config: &'cfg Config
) -> CargoResult<RegistrySource<'cfg>>

pub fn local(
source_id: SourceId,
path: &Path,
yanked_whitelist: &HashSet<PackageId>,
config: &'cfg Config
) -> RegistrySource<'cfg>

fn new(
source_id: SourceId,
config: &'cfg Config,
name: &str,
ops: Box<dyn RegistryData + 'cfg>,
yanked_whitelist: &HashSet<PackageId>
) -> RegistrySource<'cfg>

pub fn config(&mut self) -> Poll<CargoResult<Option<RegistryConfig>>>

Decode the configuration stored within the registry.

This requires that the index has been at least checked out.

fn unpack_package(&self, pkg: PackageId, tarball: &File) -> CargoResult<PathBuf>

Unpacks a downloaded package into a location where it’s ready to be compiled.

No action is taken if the source looks like it’s already unpacked.

fn get_pkg(&mut self, package: PackageId, path: &File) -> CargoResult<Package>