Struct cargo::sources::replaced::ReplacedSource

pub struct ReplacedSource<'cfg> {
    to_replace: SourceId,
    replace_with: SourceId,
    inner: Box<dyn Source + 'cfg>,
}

A source that replaces one source with the other. This manages the source replacement feature.

The implementation is merely redirecting from the original to the replacement.

Fields
to_replace: SourceId

The identifier of the original source.

replace_with: SourceId

The identifier of the new replacement source.

inner: Box<dyn Source + 'cfg>
Implementations

impl<'cfg> ReplacedSource<'cfg>

pub fn new( to_replace: SourceId, replace_with: SourceId, src: Box<dyn Source + 'cfg> ) -> ReplacedSource<'cfg>

Creates a replaced source.

The src argument is the new replacement source.