cargo::sources
Module source

Source trait for sources of Cargo packages.

Structs

SourceMap

A HashMap of SourceId to Box<Source>.

Enums

MaybePackage

A download status that represents if a Package has already been downloaded, or if not then a location to download.

QueryKind

Defines how a dependency query will be performed for a Source.

Traits

Source

An abstraction of different sources of Cargo packages.