Module cargo::util

Re-exports

pub use self::config::homedir;

pub use self::config::Config;

pub use self::config::ConfigValue;

pub use self::diagnostic_server::RustfixDiagnosticServer;

pub use self::errors::internal;

pub use self::errors::process_error;

pub use self::errors::CargoResult;

pub use self::errors::CargoResultExt;

pub use self::errors::CliResult;

pub use self::errors::Test;

pub use self::errors::CargoTestError;

pub use self::errors::CliError;

pub use self::errors::ProcessError;

pub use self::graph::Graph;

pub use self::hex::hash_u64;

pub use self::hex::short_hash;

pub use self::hex::to_hex;

pub use self::into_url::IntoUrl;

pub use self::lev_distance::closest;

pub use self::lev_distance::closest_msg;

pub use self::lev_distance::lev_distance;

pub use self::paths::bytes2path;

pub use self::paths::dylib_path;

pub use self::paths::join_paths;

pub use self::paths::path2bytes;

pub use self::paths::dylib_path_envvar;

pub use self::paths::normalize_path;

pub use self::process_builder::process;

pub use self::process_builder::ProcessBuilder;

pub use self::restricted_names::validate_package_name;

pub use self::rustc::Rustc;

pub use self::to_semver::ToSemver;

Modules

command_prelude

config
Cargo's config system.

cpu

diagnostic_server
A small TCP server to handle collection of diagnostics information in a cross-platform way for the cargo fix command.

errors

graph

hex

important_paths

into_url

job
Job management (mostly for windows)

lev_distance

machine_message

network

paths

process_builder

profile

restricted_names
Helpers for validating and checking names like package and crate names.

rustc

to_semver

toml

Structs

CanonicalUrl
A newtype wrapper around Url which represents a "canonical" version of an original URL.

DependencyQueue

FileLock

Filesystem
A "filesystem" is intended to be a globally shared, hence locked, resource in Cargo.

FossilRepo

GitRepo

HgRepo

LockServer

LockServerClient

LockServerStarted

PijulRepo

Progress

Queue
A simple, threadsafe, queue of items of type T

Sha256

StableHasher

Enums

ProgressStyle

Traits

IntoUrlWithBase
A type that can be interpreted as a relative Url and converted to a Url.

Functions

elapsed

existing_vcs_repo

is_ci
Whether or not this running in a Continuous Integration environment.

print_available_benches

print_available_binaries

print_available_examples

print_available_tests

read2