Module cargo::util::command_prelude

Re-exports

pub use crate::CliError;

pub use crate::CliResult;

pub use crate::Config;

Macros

value_parser

Select a [ValueParser] implementation from the intended type

Structs

Arg

The abstract representation of a command line argument. Used to set all the options and relationships that define a valid argument for the program.

ArgMatches

Container for parse results.

Enums

AppSettings

Application level settings, which affect how Command operates

ArgAction

Behavior of arguments when they are encountered while parsing

CommandInfo

CompileMode

The general “mode” for what to do. This is used for two purposes. The commands themselves pass this in to compile_ws to tell it the general execution strategy. This influences the default targets selected. The other use is in the Unit struct to indicate what is being done with a specific target.

ProfileChecking

Determines whether or not to gate --profile as unstable when resolving it.

Traits

ArgMatchesExt

CommandExt

Functions

flag

ignore_unknown 🔒

multi_opt

opt

optional_multi_opt

optional_opt

subcommand

values

values_os

Type Definitions

Command