cargo::util::context
Module de

Support for deserializing configuration via serde

Macros

deserialize_method 🔒

Structs

ConfigMapAccess 🔒

ConfigSeqAccess 🔒

Deserializer 🔒

Serde deserializer used to convert config values to a target type using GlobalContext::get.

Tuple2Deserializer 🔒

A deserializer which takes two values and deserializes into a tuple of those two values. This is similar to types like StrDeserializer in upstream serde itself.

ValueDeserializer 🔒

This is a deserializer that deserializes into a Value<T> for configuration.

Enums

KeyKind 🔒