cargo::util::context::de
Struct ValueDeserializer

struct ValueDeserializer<'gctx> {
    hits: u32,
    definition: Definition,
    de: Option<Deserializer<'gctx>>,
    str_value: Option<String>,
}

This is a deserializer that deserializes into a Value<T> for configuration.

This is a special deserializer because it deserializes one of its struct fields into the location that this configuration value was defined in.

See more comments in value.rs for the protocol used here.

Fields
hits: u32definition: Definitionde: Option<Deserializer<'gctx>>

The deserializer, used to actually deserialize a Value struct. This is None if deserializing a string.

str_value: Option<String>

A string value to deserialize.

This is used for situations where you can’t address a string via a TOML key, such as a string inside an array. The ConfigSeqAccess doesn’t know if the type it should deserialize to is a String or Value<String>, so ValueDeserializer needs to be able to handle both.

Implementations

impl<'gctx> ValueDeserializer<'gctx>

fn new(de: Deserializer<'gctx>) -> Result<ValueDeserializer<'gctx>, ConfigError>

fn new_with_string( s: String, definition: Definition, ) -> ValueDeserializer<'gctx>