cargo::util::context::path
Struct ConfigRelativePath

pub struct ConfigRelativePath(Value<String>);

Use with the get API to fetch a string that will be converted to a PathBuf. Relative paths are converted to absolute paths based on the location of the config file.

Tuple Fields
0: Value<String>
Implementations

impl ConfigRelativePath

pub fn new(path: Value<String>) -> ConfigRelativePath

pub fn value(&self) -> &Value<String>

Returns the underlying value.

pub fn raw_value(&self) -> &str

Returns the raw underlying configuration value for this key.

pub fn resolve_path(&self, gctx: &GlobalContext) -> PathBuf

Resolves this configuration-relative path to an absolute path.

This will always return an absolute path where itโs relative to the location for configuration for this value.

pub fn resolve_program(&self, gctx: &GlobalContext) -> PathBuf

Resolves this configuration-relative path to either an absolute path or something appropriate to execute from PATH.

Values which donโt look like a filesystem path (donโt contain / or \) will be returned as-is, and everything else will fall through to an absolute path.