Module cargo::util::context::value

Deserialization of a Value<T> type which tracks where it was deserialized from.

Often Cargo wants to report semantic error information or other sorts of error information about configuration keys but it also may wish to indicate as an error context where the key was defined as well (to help user debugging). The Value<T> type here can be used to deserialize a T value from configuration, but also record where it was deserialized from when it was read.

Structs

DefinitionKey 🔒

FieldVisitor 🔒

Value

A type which can be deserialized as a configuration value which records where it was deserialized from.

ValueKey 🔒

Enums

Definition

Location where a config value is defined.

Constants

DEFINITION_FIELD 🔒

NAME 🔒

VALUE_FIELD 🔒

Statics

FIELDS 🔒

Type Aliases

OptValue