cargo::util
Module diagnostic_server

A small TCP server to handle collection of diagnostics information in a cross-platform way for the cargo fix command.

Structs

DiagnosticPrinter

A printer that will print diagnostics messages to the shell.

RustfixDiagnosticServer

StartedServer

Enums

Message

Constants

DIAGNOSTICS_SERVER_VAR 🔒

Functions

gen_please_report_this_bug_text 🔒

get_bug_report_url 🔒