cargo::util::flock
Struct FileLock

pub struct FileLock {
    f: Option<File>,
    path: PathBuf,
}

A locked file.

This provides access to file while holding a lock on the file. This type implements the Read, Write, and Seek traits to provide access to the underlying file.

Locks are either shared (multiple processes can access the file) or exclusive (only one process can access the file).

This type is created via methods on the Filesystem type.

When this value is dropped, the lock will be released.

Fields
f: Option<File>path: PathBuf
Implementations

impl FileLock

pub fn file(&self) -> &File

Returns the underlying file handle of this lock.

pub fn path(&self) -> &Path

Returns the underlying path that this lock points to.

Note that special care must be taken to ensure that the path is not referenced outside the lifetime of this lock.

pub fn parent(&self) -> &Path

Returns the parent path containing this file

pub fn remove_siblings(&self) -> CargoResult<()>

Removes all sibling files to this locked file.

This can be useful if a directory is locked with a sentinel file but it needs to be cleared out as it may be corrupt.