Module cargo::util::lev_distance

Functions

closest
Find the closest element from iter matching choice. The key callback is used to select a &str from the iterator to compare against choice.

closest_msg
Version of closest that returns a common "suggestion" that can be tacked onto the end of an error message.

lev_distance