Module cargo::util::machine_message

Structs

Artifact

ArtifactProfile

This is different from the regular Profile to maintain backwards compatibility (in particular, test is no longer in Profile, but we still want it to be included here).

BuildFinished

BuildScript

FromCompiler

TimingInfo

Enums

ArtifactDebuginfo

Internally this is an enum with different variants, but keep using 0/1/2 as integers for compatibility.

Traits

Message