Module cargo::util::network::sleep

Utility for tracking network requests that will be retried in the future.

Structs

SleepTracker

A tracker for network requests that have failed, and are awaiting to be retried in the future.

Sleeper 🔒

An individual network request that is waiting to be retried in the future.