Module cargo::util::restricted_names

Helpers for validating and checking names like package and crate names.

Functions

is_conflicting_artifact_name
An artifact with this name will conflict with one of Cargo's build directories.

is_keyword
A Rust keyword.

is_non_ascii_name
Returns true if the name contains non-ASCII characters.

is_windows_reserved
These names cannot be used on Windows, even with an extension.

is_windows_reserved_path

validate_package_name
Check the base requirements for a package name.