Module cargo::util::rustc

Structs

Cache 🔒

It is a well known fact that rustc is not the fastest compiler in the world. What is less known is that even rustc --version --verbose takes about a hundred milliseconds! Because we need compiler version info even for no-op builds, we cache it here, based on compiler’s mtime and rustup’s current toolchain.

CacheData 🔒

Output 🔒

Rustc

Information on the rustc executable

Functions

process_fingerprint 🔒

rustc_fingerprint 🔒