Struct cargo::util::rustc::Rustc

pub struct Rustc {
    pub path: PathBuf,
    pub wrapper: Option<PathBuf>,
    pub workspace_wrapper: Option<PathBuf>,
    pub verbose_version: String,
    pub version: Version,
    pub host: InternedString,
    cache: Mutex<Cache>,
}

Information on the rustc executable

Fields
path: PathBuf

The location of the exe

wrapper: Option<PathBuf>

An optional program that will be passed the path of the rust exe as its first argument, and rustc args following this.

workspace_wrapper: Option<PathBuf>

An optional wrapper to be used in addition to rustc.wrapper for workspace crates

verbose_version: String

Verbose version information (the output of rustc -vV)

version: Version

The rustc version (1.23.4-beta.2), this comes from verbose_version.

host: InternedString

The host triple (arch-platform-OS), this comes from verbose_version.

cache: Mutex<Cache>
Implementations

impl Rustc

pub fn new(
path: PathBuf,
wrapper: Option<PathBuf>,
workspace_wrapper: Option<PathBuf>,
rustup_rustc: &Path,
cache_location: Option<PathBuf>
) -> CargoResult<Rustc>

Runs the compiler at path to learn various pieces of information about it, with an optional wrapper.

If successful this function returns a description of the compiler along with a list of its capabilities.

pub fn process(&self) -> ProcessBuilder

Gets a process builder set up to use the found rustc version, with a wrapper if Some.

pub fn workspace_process(&self) -> ProcessBuilder

Gets a process builder set up to use the found rustc version, with a wrapper if Some.

pub fn process_no_wrapper(&self) -> ProcessBuilder

pub fn cached_output(
&self,
cmd: &ProcessBuilder,
extra_fingerprint: u64
) -> CargoResult<(String, String)>

Gets the output for the given command.

This will return the cached value if available, otherwise it will run the command and cache the output.

extra_fingerprint is extra data to include in the cache fingerprint. Use this if there is other information about the environment that may affect the output that is not part of cmd.

Returns a tuple of strings (stdout, stderr).