Struct cargo::util::FileLock

pub struct FileLock {
    // some fields omitted
}

Methods

impl FileLock

fn file(&self) -> &File

Returns the underlying file handle of this lock.

fn path(&self) -> &Path

Returns the underlying path that this lock points to.

Note that special care must be taken to ensure that the path is not referenced outside the lifetime of this lock.

fn parent(&self) -> &Path

Returns the parent path containing this file

fn remove_siblings(&self) -> Result<()>

Removes all sibling files to this locked file.

This can be useful if a directory is locked with a sentinel file but it needs to be cleared out as it may be corrupt.