Struct cargo::util::Rustc

pub struct Rustc {
    pub path: PathBuf,
    pub verbose_version: String,
    pub host: String,
    pub cap_lints: bool,
}

Fields
path: PathBuf verbose_version: String host: String cap_lints: bool

Backwards compatibility: does this compiler support --cap-lints flag?

Methods

impl Rustc

fn new(path: PathBuf) -> CargoResult<Rustc>

Run the compiler at path to learn various pieces of information about it.

If successful this function returns a description of the compiler along with a list of its capabilities.

fn process(&self) -> ProcessBuilder