cargo::util::toml

Module embedded

Structs

ScriptSource

Functions

expand_manifest 🔒

is_whitespace 🔒

True if c is considered a whitespace according to Rust language definition. See Rust language reference for definitions of these classes.

sanitize_name

Ensure the package name matches the validation from ops::cargo_new::check_name

strip_shebang 🔒