Struct cargo::util::toml::DetailedTomlDependency

pub struct DetailedTomlDependency<P: Clone = String> {

    version: Option<String>,
    registry: Option<String>,
    registry_index: Option<String>,
    path: Option<P>,
    git: Option<String>,
    branch: Option<String>,
    tag: Option<String>,
    rev: Option<String>,
    features: Option<Vec<String>>,
    optional: Option<bool>,
    default_features: Option<bool>,
    default_features2: Option<bool>,
    package: Option<String>,
    public: Option<bool>,
    artifact: Option<StringOrVec>,
    lib: Option<bool>,
    target: Option<String>,
    other: BTreeMap<String, Value>,

}

Fields
version: Option<String>registry: Option<String>registry_index: Option<String>

The URL of the registry field. This is an internal implementation detail. When Cargo creates a package, it replaces registry with registry-index so that the manifest contains the correct URL. All users won’t have the same registry names configured, so Cargo can’t rely on just the name for crates published by other users.

path: Option<P>git: Option<String>branch: Option<String>tag: Option<String>rev: Option<String>features: Option<Vec<String>>optional: Option<bool>default_features: Option<bool>default_features2: Option<bool>package: Option<String>public: Option<bool>artifact: Option<StringOrVec>

One or more of bin, cdylib, staticlib, bin:<name>.

lib: Option<bool>

If set, the artifact should also be a dependency

target: Option<String>

A platform name, like x86_64-apple-darwin

other: BTreeMap<String, Value>

This is here to provide a way to see the “unused manifest keys” when deserializing

Implementations

impl<P: ResolveToPath + Clone> DetailedTomlDependency<P>

fn to_dependency( &self, name_in_toml: &str, cx: &mut Context<'_, '_>, kind: Option<DepKind> ) -> CargoResult<Dependency>

impl DetailedTomlDependency

fn add_features(&mut self, features: Option<Vec<String>>)

fn update_optional(&mut self, optional: Option<bool>)

fn resolve_path( &mut self, name: &str, root_path: &Path, package_root: &Path ) -> CargoResult<()>