Struct cargo::util::toml::TomlProfile

pub struct TomlProfile {

    pub opt_level: Option<TomlOptLevel>,
    pub lto: Option<StringOrBool>,
    pub codegen_backend: Option<InternedString>,
    pub codegen_units: Option<u32>,
    pub debug: Option<TomlDebugInfo>,
    pub split_debuginfo: Option<String>,
    pub debug_assertions: Option<bool>,
    pub rpath: Option<bool>,
    pub panic: Option<String>,
    pub overflow_checks: Option<bool>,
    pub incremental: Option<bool>,
    pub dir_name: Option<InternedString>,
    pub inherits: Option<InternedString>,
    pub strip: Option<StringOrBool>,
    pub rustflags: Option<Vec<InternedString>>,
    pub package: Option<BTreeMap<ProfilePackageSpec, TomlProfile>>,
    pub build_override: Option<Box<TomlProfile>>,

}

Fields
opt_level: Option<TomlOptLevel>lto: Option<StringOrBool>codegen_backend: Option<InternedString>codegen_units: Option<u32>debug: Option<TomlDebugInfo>split_debuginfo: Option<String>debug_assertions: Option<bool>rpath: Option<bool>panic: Option<String>overflow_checks: Option<bool>incremental: Option<bool>dir_name: Option<InternedString>inherits: Option<InternedString>strip: Option<StringOrBool>rustflags: Option<Vec<InternedString>>package: Option<BTreeMap<ProfilePackageSpec, TomlProfile>>build_override: Option<Box<TomlProfile>>
Implementations

impl TomlProfile

pub fn validate( &self, name: &str, cli_unstable: &CliUnstable, features: &Features, warnings: &mut Vec<String> ) -> CargoResult<()>

Checks stytax validity and unstable feature gate for a given profile.

pub fn validate_name(name: &str) -> CargoResult<()>

Validate dir-names and profile names according to RFC 2678.

fn validate_profile( &self, name: &str, cli_unstable: &CliUnstable, features: &Features ) -> CargoResult<()>

Validates a profile.

This is a shallow check, which is reused for the profile itself and any overrides.

fn validate_override(&self, which: &str) -> CargoResult<()>

Validation that is specific to an override.

pub fn merge(&mut self, profile: &TomlProfile)

Overwrite self’s values with the given profile.