Type Alias cargo::util::toml::MaybeWorkspaceVecStringOrBool

type MaybeWorkspaceVecStringOrBool = MaybeWorkspace<VecStringOrBool, TomlWorkspaceField>;

Aliased Type

enum MaybeWorkspaceVecStringOrBool {
    Defined(VecStringOrBool),
    Workspace(TomlWorkspaceField),
}

Variants

Defined(VecStringOrBool)

The “defined” type, or the type that that is used when not inheriting from a workspace.

Workspace(TomlWorkspaceField)

The type when inheriting from a workspace.

Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference's “Type Layout” chapter for details on type layout guarantees.

Size: 32 bytes

Size for each variant:

Defined: 31 bytes

Workspace: 1 byte