cargo::util

Module toml_mut

Utilities for in-place editing of Cargo.toml manifests.

These utilities operate only on the level of a TOML document, and generally do not perform any processing of information beyond what is required for editing. For more comprehensive usage of manifests, see Manifest.

In most cases, the entrypoint for editing is LocalManifest, which contains editing functionality for a given manifest’s dependencies.

Modules

dependency

Information about dependencies in a manifest.

manifest

Parsing and editing of manifest files.

upgrade