Module cargo::util::workspace

Functions

add_path_args

get_available_targets 🔒

path_args

The path that we pass to rustc is actually fairly important because it will show up in error messages (important for readability), debug information (important for caching), etc. As a result we need to be pretty careful how we actually invoke rustc.

print_available_benches

print_available_binaries

print_available_examples

print_available_packages

print_available_targets 🔒

print_available_tests