Module cargo::core

Re-exports

pub use self::dependency::Dependency;

pub use self::features::CliUnstable;

pub use self::features::Edition;

pub use self::features::Feature;

pub use self::features::Features;

pub use self::manifest::EitherManifest;

pub use self::manifest::VirtualManifest;

pub use self::manifest::Manifest;

pub use self::manifest::Target;

pub use self::manifest::TargetKind;

pub use self::package::Package;

pub use self::package::PackageSet;

pub use self::package_id::PackageId;

pub use self::registry::Registry;

pub use self::resolver::Resolve;

pub use self::resolver::ResolveVersion;

pub use self::shell::Shell;

pub use self::shell::Verbosity;

pub use self::source::GitReference;

pub use self::source::QueryKind;

pub use self::source::Source;

pub use self::source::SourceId;

pub use self::source::SourceMap;

pub use self::summary::FeatureMap;

pub use self::summary::FeatureValue;

pub use self::summary::Summary;

pub use crate::util::toml::InheritableFields;

Modules

compiler

Interact with the compiler

dependency

features

Support for nightly features in Cargo itself.

manifest

package

package_id

package_id_spec 🔒

profiles

Handles built-in and customizable compiler flag presets.

registry

resolver

Resolution of the entire dependency graph for a crate.

shell

source

Fundamental types and traits for sources of Cargo packages.

summary

workspace 🔒

Structs

PackageIdSpec

Some or all of the data required to identify a package:

Workspace

The core abstraction in Cargo for working with a workspace of crates.

WorkspaceRootConfig

Intermediate configuration of a workspace root in a manifest.

Enums

MaybePackage

WorkspaceConfig

Configuration of a workspace in a manifest.

Functions

find_workspace_root

Finds the path of the root of the workspace.

resolve_relative_path