Module cargo::core

Re-exports

pub use self::dependency::Dependency;

pub use self::features::enable_nightly_features;

pub use self::features::maybe_allow_nightly_features;

pub use self::features::nightly_features_allowed;

pub use self::features::CliUnstable;

pub use self::features::Edition;

pub use self::features::Feature;

pub use self::features::Features;

pub use self::manifest::EitherManifest;

pub use self::manifest::VirtualManifest;

pub use self::manifest::Manifest;

pub use self::manifest::Target;

pub use self::manifest::TargetKind;

pub use self::package::Package;

pub use self::package::PackageSet;

pub use self::package_id::PackageId;

pub use self::registry::Registry;

pub use self::shell::Shell;

pub use self::shell::Verbosity;

pub use self::source::Source;

pub use self::source::SourceMap;

pub use self::summary::FeatureMap;

pub use self::summary::FeatureValue;

pub use self::summary::Summary;

Modules

compiler

dependency

features
Support for nightly features in Cargo itself.

manifest

package

package_id

profiles

registry

resolver
Resolution of the entire dependency graph for a crate.

shell

source

summary

Structs

InternedString

Members
An iterator over the member packages of a workspace, returned by Workspace::members

PackageIdSpec
Some or all of the data required to identify a package:

Resolve
Represents a fully-resolved package dependency graph. Each node in the graph is a package and edges represent dependencies between packages.

SourceId
Unique identifier for a source of packages.

Workspace
The core abstraction in Cargo for working with a workspace of crates.

WorkspaceRootConfig
Intermediate configuration of a workspace root in a manifest.

Enums

GitReference
Information to find a specific commit in a Git repository.

ResolveVersion
A version to indicate how a Cargo.lock should be serialized. Currently V2 is the default when creating a new lockfile. If a V1 lockfile already exists, it will stay as V1.

WorkspaceConfig
Configuration of a workspace in a manifest.