cargo

Module core

Re-exports

pub use self::dependency::Dependency;

pub use self::dependency::SerializedDependency;

pub use self::features::CliUnstable;

pub use self::features::Edition;

pub use self::features::Feature;

pub use self::features::Features;

pub use self::manifest::EitherManifest;

pub use self::manifest::VirtualManifest;

pub use self::manifest::Manifest;

pub use self::manifest::Target;

pub use self::manifest::TargetKind;

pub use self::package::Package;

pub use self::package::PackageSet;

pub use self::package_id::PackageId;

pub use self::registry::Registry;

pub use self::resolver::Resolve;

pub use self::resolver::ResolveVersion;

pub use self::shell::Shell;

pub use self::shell::Verbosity;

pub use self::summary::FeatureMap;

pub use self::summary::FeatureValue;

pub use self::summary::Summary;

Modules

compiler

Interact with the compiler

dependency

features

Support for nightly features in Cargo itself.

gc

Support for garbage collecting unused files from downloaded files or artifacts from the target directory.

global_cache_tracker

Support for tracking the last time files were used to assist with cleaning up those files if they haven’t been used in a while.

manifest

package

package_id

package_id_spec 🔒

profiles

Handles built-in and customizable compiler flag presets.

registry

Types that hold source information for a group of packages.

resolver

Resolution of the entire dependency graph for a crate.

shell

source_id 🔒

summary

workspace 🔒

Structs

PackageIdSpec

Some or all of the data required to identify a package:

SourceId

Unique identifier for a source of packages.

Workspace

The core abstraction in Cargo for working with a workspace of crates.

WorkspaceRootConfig

Intermediate configuration of a workspace root in a manifest.

Enums

GitReference

Information to find a specific commit in a Git repository.

MaybePackage

SourceKind

The possible kinds of code source.

WorkspaceConfig

Configuration of a workspace in a manifest.

Traits

PackageIdSpecQuery

Functions

find_workspace_root

Finds the path of the root of the workspace.

resolve_relative_path