Module cargo::core

Reexports

pub use self::dependency::Dependency;

pub use self::manifest::EitherManifest;

pub use self::manifest::VirtualManifest;

pub use self::manifest::Manifest;

pub use self::manifest::Target;

pub use self::manifest::TargetKind;

pub use self::manifest::Profile;

pub use self::manifest::LibKind;

pub use self::manifest::Profiles;

pub use self::package::Package;

pub use self::package::PackageSet;

pub use self::package_id::PackageId;

pub use self::registry::Registry;

pub use self::resolver::Resolve;

pub use self::shell::Shell;

pub use self::shell::Verbosity;

pub use self::source::Source;

pub use self::source::SourceMap;

pub use self::summary::Summary;

Modules

dependency

manifest

package

package_id

registry

resolver
Resolution of the entire dependency graph for a crate

shell

source

summary

Structs

CliUnstable
A parsed representation of all unstable flags that Cargo accepts.

Feature

Features

Members
An iterator over the member packages of a workspace, returned by Workspace::members

PackageIdSpec

SourceId
Unique identifier for a source of packages.

Workspace
The core abstraction in Cargo for working with a workspace of crates.

WorkspaceRootConfig
Intermediate configuration of a workspace root in a manifest.

Enums

GitReference
Information to find a specific commit in a git repository

WorkspaceConfig
Configuration of a workspace in a manifest.