Module cargo::core

Reexports

pub use self::dependency::{Dependency, DependencyInner};

pub use self::manifest::{Manifest, Target, TargetKind, Profile, LibKind, Profiles};

pub use self::manifest::{EitherManifest, VirtualManifest};

pub use self::package::{Package, PackageSet};

pub use self::package_id::{PackageId, Metadata};

pub use self::registry::Registry;

pub use self::resolver::Resolve;

pub use self::shell::{Shell, MultiShell, ShellConfig, Verbosity, ColorConfig};

pub use self::source::{Source, SourceId, SourceMap, GitReference};

pub use self::summary::Summary;

Modules

dependency

manifest

package

package_id

registry

resolver
Resolution of the entire dependency graph for a crate

shell

source

summary

Structs

PackageIdSpec

Workspace
The core abstraction in Cargo for working with a workspace of crates.

Enums

WorkspaceConfig
Configuration of a workspace in a manifest.