cargo::core::compiler::build_runner
Module compilation_files

See CompilationFiles.

Structs

CompilationFiles

Collection of information about the files emitted by the compiler, and the output directory structure.

Metadata

Metadata tracks several UnitHashs, including Metadata::unit_id, Metadata::c_metadata, and Metadata::c_extra_filename.

OutputFile

Info about a single file emitted by the compiler.

UnitHash

Uniquely identify a Unit under specific circumstances, see Metadata for more.

Constants

METADATA_VERSION 🔒

This is a generic version number that can be changed to make backwards-incompatible changes to any file structures in the output directory. For example, the fingerprint files or the build-script output files.

Functions

compute_metadata 🔒

Computes the metadata hash for the given Unit.

has_remap_path_prefix 🔒

HACK: Detect the potential presence of --remap-path-prefix

hash_rustc_version 🔒

Hash the version of rustc being used during the build process.

metadata_of 🔒

Gets the metadata hash for the given Unit.

use_extra_filename 🔒

Returns whether or not this unit should use a hash in the filename to make it unique.