Module cargo::core::compiler::build_runner::compilation_files

See CompilationFiles.

Structs

CompilationFiles

Collection of information about the files emitted by the compiler, and the output directory structure.

MetaInfo 🔒

Information about the metadata hashes used for a Unit.

Metadata

The Metadata is a hash used to make unique file names for each unit in a build. It is also used for symbol mangling.

OutputFile

Info about a single file emitted by the compiler.

Constants

METADATA_VERSION 🔒

This is a generic version number that can be changed to make backwards-incompatible changes to any file structures in the output directory. For example, the fingerprint files or the build-script output files.

Functions

compute_metadata 🔒

Computes the metadata hash for the given Unit.

hash_rustc_version 🔒

Hash the version of rustc being used during the build process.

metadata_of 🔒

Gets the metadata hash for the given Unit.

should_use_metadata 🔒

Returns whether or not this unit should use a metadata hash.