cargo::core::compiler
Module compilation

Type definitions for the result of a compilation.

Structs

Compilation

A structure returning the result of a compilation.

Doctest

Structure with enough information to run rustdoc --test.

UnitOutput

Information about the output of a unit.

Enums

ToolKind 🔒

Represents the kind of process we are creating.

Functions

fill_rustc_tool_env 🔒

Prepares a rustc_tool process with additional environment variables that are only relevant in a context that has a unit

get_sysroot_target_libdir 🔒

target_linker 🔒

Gets the user-specified linker for a particular host or target from the configuration.

target_runner 🔒