Module cargo::core::compiler::custom_build

Structs

BuildDeps

Dependency information as declared by a build script.

BuildOutput

Contains the parsed output of a custom build script.

BuildScriptOutputs

Map of packages to build script output.

BuildScripts

Linking information for a Unit.

Constants

CARGO_WARNING 🔒

Functions

build_map

Computes several maps in Context:

build_work 🔒

emit_build_output 🔒

insert_warnings_in_build_outputs 🔒

prepare

Prepares a Work that executes the target as a custom build script.

prepare_metabuild 🔒

prev_build_output 🔒

Returns the previous parsed BuildOutput, if any, from a previous execution.