cargo::core::compiler::fingerprint
Module dep_info

Types and functions managing dep-info files. For more, see the documentation in the fingerprint module.

Structs

Checksum

EncodedDepInfo

Same as RustcDepInfo except avoids absolute paths as much as possible to allow moving around the target directory.

RustcDepInfo

The representation of the .d dep-info file generated by rustc

Enums

ChecksumAlgo

Some algorithms are here to ensure compatibility with possible rustc outputs. The presence of an algorithm here is not a suggestion that it’s fit for use.

DepInfoPathType

Tells the associated path in EncodedDepInfo::files is relative to package root, target root, or absolute.

InvalidChecksum

Constants

CURRENT_ENCODED_DEP_INFO_VERSION 🔒

The current format version of EncodedDepInfo.

Functions

make_absolute_path 🔒

parse_dep_info

Parses Cargo’s internal EncodedDepInfo structure that was previously serialized to disk.

parse_rustc_dep_info

Parse the .d dep-info file generated by rustc.

translate_dep_info

Parses the dep-info file coming out of rustc into a Cargo-specific format.