Function cargo::core::compiler::standard_lib::generate_std_roots

pub fn generate_std_roots(
    crates: &[String], 
    std_resolve: &Resolve, 
    std_features: &ResolvedFeatures, 
    kinds: &[CompileKind], 
    package_set: &PackageSet<'_>, 
    interner: &UnitInterner, 
    profiles: &Profiles
) -> CargoResult<HashMap<CompileKind, Vec<Unit>>>

Generate a list of root Units for the standard library.

The given slice of crate names is the root set.