Struct cargo::core::compiler::BuildContext

pub struct BuildContext<'a, 'cfg> {
    pub ws: &'a Workspace<'cfg>,
    pub config: &'cfg Config,
    pub profiles: Profiles,
    pub build_config: &'a BuildConfig,
    pub extra_compiler_args: HashMap<Unit, Vec<String>>,
    pub packages: PackageSet<'cfg>,
    pub target_data: RustcTargetData,
    pub roots: Vec<Unit>,
    pub unit_graph: UnitGraph,
}

The build context, containing all information about a build task.

It is intended that this is mostly static information. Stuff that mutates during the build can be found in the parent Context. (I say mostly, because this has internal caching, but nothing that should be observable or require &mut.)

Fields
ws: &'a Workspace<'cfg>

The workspace the build is for.

config: &'cfg Config

The cargo configuration.

profiles: Profilesbuild_config: &'a BuildConfigextra_compiler_args: HashMap<Unit, Vec<String>>

Extra compiler args for either rustc or rustdoc.

packages: PackageSet<'cfg>

Package downloader.

This holds ownership of the Package objects.

target_data: RustcTargetData

Information about rustc and the target platform.

roots: Vec<Unit>

The root units of unit_graph (units requested on the command-line).

unit_graph: UnitGraph

The dependency graph of units to compile.

Implementations

impl<'a, 'cfg> BuildContext<'a, 'cfg>

pub fn new(
ws: &'a Workspace<'cfg>,
packages: PackageSet<'cfg>,
build_config: &'a BuildConfig,
profiles: Profiles,
extra_compiler_args: HashMap<Unit, Vec<String>>,
target_data: RustcTargetData,
roots: Vec<Unit>,
unit_graph: UnitGraph
) -> CargoResult<BuildContext<'a, 'cfg>>

pub fn rustc(&self) -> &Rustc

pub fn linker(&self, kind: CompileKind) -> Option<PathBuf>

Gets the user-specified linker for a particular host or target.

pub fn host_triple(&self) -> InternedString

Gets the host architecture triple.

For example, x86_64-unknown-linux-gnu, would be

machine: x86_64,

hardware-platform: unknown,

operating system: linux-gnu.

pub fn jobs(&self) -> u32

Gets the number of jobs specified for this build.

pub fn rustflags_args(&self, unit: &Unit) -> &[String]

pub fn rustdocflags_args(&self, unit: &Unit) -> &[String]

pub fn extra_args_for(&self, unit: &Unit) -> Option<&Vec<String>>