Struct cargo::core::compiler::Compilation

pub struct Compilation<'cfg> {

    pub tests: Vec<UnitOutput>,
    pub binaries: Vec<UnitOutput>,
    pub cdylibs: Vec<UnitOutput>,
    pub root_crate_names: Vec<String>,
    pub native_dirs: BTreeSet<PathBuf>,
    pub root_output: HashMap<CompileKind, PathBuf>,
    pub deps_output: HashMap<CompileKind, PathBuf>,
    sysroot_host_libdir: PathBuf,
    sysroot_target_libdir: HashMap<CompileKind, PathBuf>,
    pub extra_env: HashMap<Metadata, Vec<(String, String)>>,
    pub to_doc_test: Vec<Doctest>,
    pub host: String,
    config: &'cfg Config,
    rustc_process: ProcessBuilder,
    rustc_workspace_wrapper_process: ProcessBuilder,
    primary_rustc_process: Option<ProcessBuilder>,
    target_runners: HashMap<CompileKind, Option<(PathBuf, Vec<String>)>>,

}

A structure returning the result of a compilation.

Fields
tests: Vec<UnitOutput>

An array of all tests created during this compilation.

binaries: Vec<UnitOutput>

An array of all binaries created.

cdylibs: Vec<UnitOutput>

An array of all cdylibs created.

root_crate_names: Vec<String>

The crate names of the root units specified on the command-line.

native_dirs: BTreeSet<PathBuf>

All directories for the output of native build commands.

This is currently used to drive some entries which are added to the LD_LIBRARY_PATH as appropriate.

The order should be deterministic.

root_output: HashMap<CompileKind, PathBuf>

Root output directory (for the local package’s artifacts)

deps_output: HashMap<CompileKind, PathBuf>

Output directory for rust dependencies. May be for the host or for a specific target.

sysroot_host_libdir: PathBuf

The path to the host libdir for the compiler used

sysroot_target_libdir: HashMap<CompileKind, PathBuf>

The path to libstd for each target

extra_env: HashMap<Metadata, Vec<(String, String)>>

Extra environment variables that were passed to compilations and should be passed to future invocations of programs.

The key is the build script metadata for uniquely identifying the RunCustomBuild unit that generated these env vars.

to_doc_test: Vec<Doctest>

Libraries to test with rustdoc.

host: String

The target host triple.

config: &'cfg Configrustc_process: ProcessBuilder

Rustc process to be used by default

rustc_workspace_wrapper_process: ProcessBuilder

Rustc process to be used for workspace crates instead of rustc_process

primary_rustc_process: Option<ProcessBuilder>

Optional rustc process to be used for primary crates instead of either rustc_process or rustc_workspace_wrapper_process

target_runners: HashMap<CompileKind, Option<(PathBuf, Vec<String>)>>
Implementations

impl<'cfg> Compilation<'cfg>

pub fn new<'a>(bcx: &BuildContext<'a, 'cfg>) -> CargoResult<Compilation<'cfg>>

pub fn rustc_process(
&self,
unit: &Unit,
is_primary: bool,
is_workspace: bool
) -> CargoResult<ProcessBuilder>

Returns a [ProcessBuilder] for running rustc.

is_primary is true if this is a “primary package”, which means it was selected by the user on the command-line (such as with a -p flag), see crate::core::compiler::Context::primary_packages.

is_workspace is true if this is a workspace member.

pub fn rustdoc_process(
&self,
unit: &Unit,
script_meta: Option<Metadata>
) -> CargoResult<ProcessBuilder>

Returns a [ProcessBuilder] for running rustdoc.

pub fn host_process<T: AsRef<OsStr>>(
&self,
cmd: T,
pkg: &Package
) -> CargoResult<ProcessBuilder>

Returns a [ProcessBuilder] appropriate for running a process for the host platform.

This is currently only used for running build scripts. If you use this for anything else, please be extra careful on how environment variables are set!

pub fn target_runner(&self, kind: CompileKind) -> Option<&(PathBuf, Vec<String>)>

pub fn target_process<T: AsRef<OsStr>>(
&self,
cmd: T,
kind: CompileKind,
pkg: &Package,
script_meta: Option<Metadata>
) -> CargoResult<ProcessBuilder>

Returns a [ProcessBuilder] appropriate for running a process for the target platform. This is typically used for cargo run and cargo test.

script_meta is the metadata for the RunCustomBuild unit that this unit used for its build script. Use None if the package did not have a build script.

fn fill_env(
&self,
cmd: ProcessBuilder,
pkg: &Package,
script_meta: Option<Metadata>,
kind: CompileKind,
is_rustc_tool: bool
) -> CargoResult<ProcessBuilder>

Prepares a new process with an appropriate environment to run against the artifacts produced by the build process.

The package argument is also used to configure environment variables as well as the working directory of the child process.