cargo::core::compiler
Struct RustcTargetData

pub struct RustcTargetData<'gctx> {
    pub rustc: Rustc,
    pub gctx: &'gctx GlobalContext,
    requested_kinds: Vec<CompileKind>,
    host_config: TargetConfig,
    host_info: TargetInfo,
    target_config: HashMap<CompileTarget, TargetConfig>,
    target_info: HashMap<CompileTarget, TargetInfo>,
}

Collection of information about rustc and the host and target.

Fields
rustc: Rustc

Information about rustc itself.

gctx: &'gctx GlobalContext

Config

requested_kinds: Vec<CompileKind>host_config: TargetConfig

Build information for the “host”, which is information about when rustc is invoked without a --target flag. This is used for selecting a linker, and applying link overrides.

The configuration read into this depends on whether or not target-applies-to-host=true.

host_info: TargetInfo

Information about the host platform.

target_config: HashMap<CompileTarget, TargetConfig>

Build information for targets that we’re building for.

target_info: HashMap<CompileTarget, TargetInfo>

Information about the target platform that we’re building for.

Implementations

impl<'gctx> RustcTargetData<'gctx>

pub fn new( ws: &Workspace<'gctx>, requested_kinds: &[CompileKind], ) -> CargoResult<RustcTargetData<'gctx>>

pub fn merge_compile_kind(&mut self, kind: CompileKind) -> CargoResult<()>

Insert kind into our target_info and target_config members if it isn’t present yet.

pub fn short_name<'a>(&'a self, kind: &'a CompileKind) -> &'a str

Returns a “short” name for the given kind, suitable for keying off configuration in Cargo or presenting to users.

pub fn dep_platform_activated( &self, dep: &Dependency, kind: CompileKind, ) -> bool

Whether a dependency should be compiled for the host or target platform, specified by CompileKind.

pub fn cfg(&self, kind: CompileKind) -> &[Cfg]

Gets the list of cfgs printed out from the compiler for the specified kind.

pub fn info(&self, kind: CompileKind) -> &TargetInfo

Information about the given target platform, learned by querying rustc.

Panics

Panics, if the target platform described by kind can’t be found. See get_info for a non-panicking alternative.

pub fn get_info(&self, kind: CompileKind) -> Option<&TargetInfo>

Information about the given target platform, learned by querying rustc.

Returns None if the target platform described by kind can’t be found.

pub fn target_config(&self, kind: CompileKind) -> &TargetConfig

Gets the target configuration for a particular host or target.

pub fn get_unsupported_std_targets(&self) -> Vec<&str>