Struct cargo::core::compiler::TargetInfo

pub struct TargetInfo {
    crate_type_process: ProcessBuilder,
    crate_types: RefCell<HashMap<CrateType, Option<(String, String)>>>,
    cfg: Vec<Cfg>,
    pub sysroot: PathBuf,
    pub sysroot_host_libdir: PathBuf,
    pub sysroot_target_libdir: PathBuf,
    pub rustflags: Vec<String>,
    pub rustdocflags: Vec<String>,
    pub supports_split_debuginfo: bool,
}

Information about the platform target gleaned from querying rustc.

RustcTargetData keeps several of these, one for the host and the others for other specified targets. If no target is specified, it uses a clone from the host.

Fields
crate_type_process: ProcessBuilder

A base process builder for discovering crate type information. In particular, this is used to determine the output filename prefix and suffix for a crate type.

crate_types: RefCell<HashMap<CrateType, Option<(String, String)>>>

Cache of output filename prefixes and suffixes.

The key is the crate type name (like cdylib) and the value is Some((prefix, suffix)), for example libcargo.so would be Some(("lib", ".so")). The value is None` if the crate type is not supported.

cfg: Vec<Cfg>

cfg information extracted from rustc --print=cfg.

sysroot: PathBuf

Path to the sysroot.

sysroot_host_libdir: PathBuf

Path to the “lib” or “bin” directory that rustc uses for its dynamic libraries.

sysroot_target_libdir: PathBuf

Path to the “lib” directory in the sysroot which rustc uses for linking target libraries.

rustflags: Vec<String>

Extra flags to pass to rustc, see extra_args.

rustdocflags: Vec<String>

Extra flags to pass to rustdoc, see extra_args.

supports_split_debuginfo: bool

Whether or not rustc supports the -Csplit-debuginfo flag.

Implementations

impl TargetInfo

pub fn new(
config: &Config,
requested_kinds: &[CompileKind],
rustc: &Rustc,
kind: CompileKind
) -> CargoResult<TargetInfo>

Learns the information of target platform from rustc invocation(s).

Generally, the first time calling this function is expensive, as it may query rustc several times. To reduce the cost, output of each rustc invocation is cached by Rustc::cached_output.

Search Tricky to learn why querying rustc several times is needed.

fn not_user_specific_cfg(cfg: &CargoResult<Cfg>) -> bool

pub fn cfg(&self) -> &[Cfg]

All the target [Cfg] settings.

fn file_types(
&self,
crate_type: &CrateType,
flavor: FileFlavor,
target_triple: &str
) -> CargoResult<Option<Vec<FileType>>>

Returns the list of file types generated by the given crate type.

Returns None if the target does not support the given crate type.

fn discover_crate_type(
&self,
crate_type: &CrateType
) -> CargoResult<Option<(String, String)>>

pub fn rustc_outputs(
&self,
mode: CompileMode,
target_kind: &TargetKind,
target_triple: &str
) -> CargoResult<(Vec<FileType>, Vec<CrateType>)>

Returns all the file types generated by rustc for the given mode/target_kind.

The first value is a Vec of file types generated, the second value is a list of CrateTypes that are not supported by the given target.

fn calc_rustc_outputs(
&self,
target_kind: &TargetKind,
target_triple: &str
) -> CargoResult<(Vec<FileType>, Vec<CrateType>)>