cargo::core::compiler::unit_dependencies
Function connect_run_custom_build_deps

fn connect_run_custom_build_deps(state: &mut State<'_, '_>)

Fill in missing dependencies for units of the RunCustomBuild

As mentioned above in compute_deps_custom_build each build script execution has two dependencies. The first is compiling the build script itself (already added) and the second is that all crates the package of the build script depends on with links keys, their build script execution. (a bit confusing eh?)

Here we take the entire deps map and add more dependencies from execution of one build script to execution of another build script.