Struct cargo::core::dependency::Dependency

pub struct Dependency { /* fields omitted */ }

Information about a dependency requested by a Cargo manifest. Cheap to copy.

Methods

impl Dependency

pub fn parse(
name: &str,
version: Option<&str>,
source_id: &SourceId,
inside: &PackageId,
config: &Config
) -> CargoResult<Dependency>

Attempt to create a Dependency from an entry in the manifest.

pub fn parse_no_deprecated(
name: &str,
version: Option<&str>,
source_id: &SourceId
) -> CargoResult<Dependency>

Attempt to create a Dependency from an entry in the manifest.

pub fn new_override(name: &str, source_id: &SourceId) -> Dependency

pub fn version_req(&self) -> &VersionReq

pub fn name(&self) -> &str

pub fn source_id(&self) -> &SourceId

pub fn kind(&self) -> Kind

pub fn specified_req(&self) -> bool

pub fn platform(&self) -> Option<&Platform>

If none, this dependencies must be built for all platforms. If some, it must only be built for the specified platform.

pub fn set_kind(&mut self, kind: Kind) -> &mut Dependency

pub fn set_features(&mut self, features: Vec<String>) -> &mut Dependency

Sets the list of features requested for the package.

pub fn set_default_features(
&mut self,
default_features: bool
) -> &mut Dependency

Sets whether the dependency requests default features of the package.

pub fn set_optional(&mut self, optional: bool) -> &mut Dependency

Sets whether the dependency is optional.

pub fn set_source_id(&mut self, id: SourceId) -> &mut Dependency

Set the source id for this dependency

pub fn set_version_req(&mut self, req: VersionReq) -> &mut Dependency

Set the version requirement for this dependency

pub fn set_platform(&mut self, platform: Option<Platform>) -> &mut Dependency

pub fn lock_to(&mut self, id: &PackageId) -> &mut Dependency

Lock this dependency to depending on the specified package id

pub fn is_locked(&self) -> bool

Returns whether this is a "locked" dependency, basically whether it has an exact version req.

pub fn is_transitive(&self) -> bool

Returns false if the dependency is only used to build the local package.

pub fn is_build(&self) -> bool

pub fn is_optional(&self) -> bool

pub fn uses_default_features(&self) -> bool

Returns true if the default features of the dependency are requested.

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn features(&self) -> &[String]

Returns the list of features that are requested by the dependency.

pub fn matches(&self, sum: &Summary) -> bool

Returns true if the package (sum) can fulfill this dependency request.

pub fn matches_ignoring_source(&self, sum: &Summary) -> bool

Returns true if the package (sum) can fulfill this dependency request.

pub fn matches_id(&self, id: &PackageId) -> bool

Returns true if the package (id) can fulfill this dependency request.

pub fn map_source(
self,
to_replace: &SourceId,
replace_with: &SourceId
) -> Dependency