Enum cargo::core::WorkspaceConfig

pub enum WorkspaceConfig {
    Root(WorkspaceRootConfig),
    Member {
        root: Option<String>,
    },
}

Configuration of a workspace in a manifest.

Variants
Root(WorkspaceRootConfig)

Indicates that [workspace] was present and the members were optionally specified as well.

Member

Indicates that [workspace] was present and the root field is the optional value of package.workspace, if present.

Fields of Member

root: Option<String>