Struct cargo::core::features::CliUnstable

pub struct CliUnstable {

    pub allow_features: Option<BTreeSet<String>>,
    pub print_im_a_teapot: bool,
    pub advanced_env: bool,
    pub avoid_dev_deps: bool,
    pub binary_dep_depinfo: bool,
    pub bindeps: bool,
    pub build_std: Option<Vec<String>>,
    pub build_std_features: Option<Vec<String>>,
    pub check_cfg: Option<(bool, bool, bool, bool)>,
    pub codegen_backend: bool,
    pub config_include: bool,
    pub credential_process: bool,
    pub direct_minimal_versions: bool,
    pub doctest_xcompile: bool,
    pub dual_proc_macros: bool,
    pub features: Option<Vec<String>>,
    pub gitoxide: Option<GitoxideFeatures>,
    pub host_config: bool,
    pub lints: bool,
    pub minimal_versions: bool,
    pub msrv_policy: bool,
    pub mtime_on_use: bool,
    pub next_lockfile_bump: bool,
    pub no_index_update: bool,
    pub panic_abort_tests: bool,
    pub profile_rustflags: bool,
    pub publish_timeout: bool,
    pub registry_auth: bool,
    pub rustdoc_map: bool,
    pub rustdoc_scrape_examples: bool,
    pub script: bool,
    pub separate_nightlies: bool,
    pub skip_rustdoc_fingerprint: bool,
    pub target_applies_to_host: bool,
    pub unstable_options: bool,

}

A parsed representation of all unstable flags that Cargo accepts.

Cargo, like rustc, accepts a suite of -Z flags which are intended for gating unstable functionality to Cargo. These flags are only available on the nightly channel of Cargo.

Fields
allow_features: Option<BTreeSet<String>>print_im_a_teapot: booladvanced_env: boolavoid_dev_deps: boolbinary_dep_depinfo: boolbindeps: boolbuild_std: Option<Vec<String>>build_std_features: Option<Vec<String>>check_cfg: Option<(bool, bool, bool, bool)>codegen_backend: boolconfig_include: boolcredential_process: booldirect_minimal_versions: booldoctest_xcompile: booldual_proc_macros: boolfeatures: Option<Vec<String>>gitoxide: Option<GitoxideFeatures>host_config: boollints: boolminimal_versions: boolmsrv_policy: boolmtime_on_use: boolnext_lockfile_bump: boolno_index_update: boolpanic_abort_tests: boolprofile_rustflags: boolpublish_timeout: boolregistry_auth: boolrustdoc_map: boolrustdoc_scrape_examples: boolscript: boolseparate_nightlies: boolskip_rustdoc_fingerprint: booltarget_applies_to_host: boolunstable_options: bool
Implementations

impl CliUnstable

pub fn help() -> Vec<(&'static str, &'static str)>

impl CliUnstable

pub fn parse( &mut self, flags: &[String], nightly_features_allowed: bool ) -> CargoResult<Vec<String>>

fn add(&mut self, flag: &str, warnings: &mut Vec<String>) -> CargoResult<()>

pub fn fail_if_stable_opt(&self, flag: &str, issue: u32) -> CargoResult<()>

Generates an error if -Z unstable-options was not used for a new, unstable command-line flag.

pub fn fail_if_stable_command( &self, config: &Config, command: &str, issue: u32 ) -> CargoResult<()>

Generates an error if -Z unstable-options was not used for a new, unstable subcommand.