Struct cargo::core::features::CliUnstable

pub struct CliUnstable {

    pub allow_features: Option<BTreeSet<String>>,
    pub print_im_a_teapot: bool,
    pub advanced_env: bool,
    pub avoid_dev_deps: bool,
    pub binary_dep_depinfo: bool,
    pub bindeps: bool,
    pub build_std: Option<Vec<String>>,
    pub build_std_features: Option<Vec<String>>,
    pub config_include: bool,
    pub credential_process: bool,
    pub check_cfg: Option<(bool, bool, bool, bool)>,
    pub doctest_in_workspace: bool,
    pub doctest_xcompile: bool,
    pub dual_proc_macros: bool,
    pub features: Option<Vec<String>>,
    pub jobserver_per_rustc: bool,
    pub minimal_versions: bool,
    pub mtime_on_use: bool,
    pub no_index_update: bool,
    pub panic_abort_tests: bool,
    pub host_config: bool,
    pub sparse_registry: bool,
    pub registry_auth: bool,
    pub target_applies_to_host: bool,
    pub rustdoc_map: bool,
    pub separate_nightlies: bool,
    pub publish_timeout: bool,
    pub unstable_options: bool,
    pub skip_rustdoc_fingerprint: bool,
    pub rustdoc_scrape_examples: bool,

}

A parsed representation of all unstable flags that Cargo accepts.

Cargo, like rustc, accepts a suite of -Z flags which are intended for gating unstable functionality to Cargo. These flags are only available on the nightly channel of Cargo.

Fields
allow_features: Option<BTreeSet<String>>print_im_a_teapot: booladvanced_env: boolavoid_dev_deps: boolbinary_dep_depinfo: boolbindeps: boolbuild_std: Option<Vec<String>>build_std_features: Option<Vec<String>>config_include: boolcredential_process: boolcheck_cfg: Option<(bool, bool, bool, bool)>doctest_in_workspace: booldoctest_xcompile: booldual_proc_macros: boolfeatures: Option<Vec<String>>jobserver_per_rustc: boolminimal_versions: boolmtime_on_use: boolno_index_update: boolpanic_abort_tests: boolhost_config: boolsparse_registry: boolregistry_auth: booltarget_applies_to_host: boolrustdoc_map: boolseparate_nightlies: boolpublish_timeout: boolunstable_options: boolskip_rustdoc_fingerprint: boolrustdoc_scrape_examples: bool
Implementations

impl CliUnstable

pub fn help() -> Vec<(&'static str, &'static str)>

impl CliUnstable

pub fn parse(
&mut self,
flags: &[String],
nightly_features_allowed: bool
) -> CargoResult<Vec<String>>

fn add(&mut self, flag: &str, warnings: &mut Vec<String>) -> CargoResult<()>

pub fn fail_if_stable_opt(&self, flag: &str, issue: u32) -> CargoResult<()>

Generates an error if -Z unstable-options was not used for a new, unstable command-line flag.

pub fn fail_if_stable_command(
&self,
config: &Config,
command: &str,
issue: u32
) -> CargoResult<()>

Generates an error if -Z unstable-options was not used for a new, unstable subcommand.