cargo::core

Module manifest

Macros

get_metadata_env 🔒

metadata_envs 🔒

Structs

DelayedWarning

When parsing Cargo.toml, some warnings should silenced if the manifest comes from a dependency. ManifestWarning allows this delayed emission of warnings.

Manifest

Contains all the information about a package, as loaded from a Cargo.toml.

ManifestMetadata

General metadata about a package which is just blindly uploaded to the registry.

MetadataEnvs 🔒

SerializedTarget 🔒

Target

Information about a binary, a library, an example, etc. that is part of the package.

TargetInner 🔒

VirtualManifest

Warnings

Enums

EitherManifest

TargetKind

TargetSourcePath

Constants

MANIFEST_PREAMBLE