Module cargo::core::manifest

Structs

Manifest
Contains all the information about a package, as loaded from a Cargo.toml.

ManifestMetadata
General metadata about a package which is just blindly uploaded to the registry.

Profile

Profiles

Target
Information about a binary, a library, an example, etc. that is part of the package.

VirtualManifest

Enums

EitherManifest

LibKind

TargetKind