cargo::core::manifest
Struct Manifest

pub struct Manifest {

    contents: Rc<String>,
    document: Rc<ImDocument<String>>,
    original_toml: Rc<TomlManifest>,
    normalized_toml: Rc<TomlManifest>,
    summary: Summary,
    targets: Vec<Target>,
    default_kind: Option<CompileKind>,
    forced_kind: Option<CompileKind>,
    links: Option<String>,
    warnings: Warnings,
    exclude: Vec<String>,
    include: Vec<String>,
    metadata: ManifestMetadata,
    custom_metadata: Option<Value>,
    publish: Option<Vec<String>>,
    replace: Vec<(PackageIdSpec, Dependency)>,
    patch: HashMap<Url, Vec<Dependency>>,
    workspace: WorkspaceConfig,
    unstable_features: Features,
    edition: Edition,
    rust_version: Option<RustVersion>,
    im_a_teapot: Option<bool>,
    default_run: Option<String>,
    metabuild: Option<Vec<String>>,
    resolve_behavior: Option<ResolveBehavior>,
    lint_rustflags: Vec<String>,
    embedded: bool,

}

Contains all the information about a package, as loaded from a Cargo.toml.

This is deserialized using the TomlManifest type.

Fields
contents: Rc<String>document: Rc<ImDocument<String>>original_toml: Rc<TomlManifest>normalized_toml: Rc<TomlManifest>summary: Summarytargets: Vec<Target>default_kind: Option<CompileKind>forced_kind: Option<CompileKind>links: Option<String>warnings: Warningsexclude: Vec<String>include: Vec<String>metadata: ManifestMetadatacustom_metadata: Option<Value>publish: Option<Vec<String>>replace: Vec<(PackageIdSpec, Dependency)>patch: HashMap<Url, Vec<Dependency>>workspace: WorkspaceConfigunstable_features: Featuresedition: Editionrust_version: Option<RustVersion>im_a_teapot: Option<bool>default_run: Option<String>metabuild: Option<Vec<String>>resolve_behavior: Option<ResolveBehavior>lint_rustflags: Vec<String>embedded: bool
Implementations

impl Manifest

pub fn new( contents: Rc<String>, document: Rc<ImDocument<String>>, original_toml: Rc<TomlManifest>, normalized_toml: Rc<TomlManifest>, summary: Summary, default_kind: Option<CompileKind>, forced_kind: Option<CompileKind>, targets: Vec<Target>, exclude: Vec<String>, include: Vec<String>, links: Option<String>, metadata: ManifestMetadata, custom_metadata: Option<Value>, publish: Option<Vec<String>>, replace: Vec<(PackageIdSpec, Dependency)>, patch: HashMap<Url, Vec<Dependency>>, workspace: WorkspaceConfig, unstable_features: Features, edition: Edition, rust_version: Option<RustVersion>, im_a_teapot: Option<bool>, default_run: Option<String>, metabuild: Option<Vec<String>>, resolve_behavior: Option<ResolveBehavior>, lint_rustflags: Vec<String>, embedded: bool, ) -> Manifest

pub fn contents(&self) -> &str

The raw contents of the original TOML

pub fn to_normalized_contents(&self) -> CargoResult<String>

See Manifest::normalized_toml for what “normalized” means

pub fn document(&self) -> &ImDocument<String>

Collection of spans for the original TOML

pub fn original_toml(&self) -> &TomlManifest

The TomlManifest as parsed from Manifest::document

pub fn normalized_toml(&self) -> &TomlManifest

The TomlManifest with all fields expanded

This is the intersection of what fields need resolving for cargo-publish that also are useful for the operation of cargo, including

workspace inheritance

target discovery

pub fn summary(&self) -> &Summary

pub fn summary_mut(&mut self) -> &mut Summary

pub fn dependencies(&self) -> &[Dependency]

pub fn default_kind(&self) -> Option<CompileKind>

pub fn forced_kind(&self) -> Option<CompileKind>

pub fn exclude(&self) -> &[String]

pub fn include(&self) -> &[String]

pub fn metadata(&self) -> &ManifestMetadata

pub fn name(&self) -> InternedString

pub fn package_id(&self) -> PackageId

pub fn targets(&self) -> &[Target]

pub fn targets_mut(&mut self) -> &mut [Target]

pub fn version(&self) -> &Version

pub fn warnings_mut(&mut self) -> &mut Warnings

pub fn warnings(&self) -> &Warnings

pub fn profiles(&self) -> Option<&TomlProfiles>

pub fn publish(&self) -> &Option<Vec<String>>

pub fn replace(&self) -> &[(PackageIdSpec, Dependency)]

pub fn patch(&self) -> &HashMap<Url, Vec<Dependency>>

pub fn links(&self) -> Option<&str>

pub fn is_embedded(&self) -> bool

pub fn workspace_config(&self) -> &WorkspaceConfig

pub fn unstable_features(&self) -> &Features

Unstable, nightly features that are enabled in this manifest.

pub fn resolve_behavior(&self) -> Option<ResolveBehavior>

The style of resolver behavior to use, declared with the resolver field.

Returns None if it is not specified.

pub fn lint_rustflags(&self) -> &[String]

RUSTFLAGS from the [lints] table

pub fn map_source( self, to_replace: SourceId, replace_with: SourceId, ) -> Manifest

pub fn feature_gate(&self) -> CargoResult<()>

pub fn print_teapot(&self, gctx: &GlobalContext)

pub fn edition(&self) -> Edition

pub fn rust_version(&self) -> Option<&RustVersion>

pub fn custom_metadata(&self) -> Option<&Value>

pub fn default_run(&self) -> Option<&str>

pub fn metabuild(&self) -> Option<&Vec<String>>

pub fn metabuild_path(&self, target_dir: Filesystem) -> PathBuf