Struct cargo::core::manifest::ManifestMetadata

pub struct ManifestMetadata {
    pub authors: Vec<String>,
    pub keywords: Vec<String>,
    pub license: Option<String>,
    pub license_file: Option<String>,
    pub description: Option<String>,
    pub readme: Option<String>,
    pub homepage: Option<String>,
    pub repository: Option<String>,
    pub documentation: Option<String>,
}

General metadata about a package which is just blindly uploaded to the registry.

Note that many of these fields can contain invalid values such as the homepage, repository, documentation, or license. These fields are not validated by cargo itself, but rather it is up to the registry when uploaded to validate these fields. Cargo will itself accept any valid TOML specification for these values.

Fields
authors: Vec<String> keywords: Vec<String> license: Option<String> license_file: Option<String> description: Option<String> readme: Option<String> homepage: Option<String> repository: Option<String> documentation: Option<String>