Struct cargo::core::package::Package

pub struct Package { /* fields omitted */ }

Information about a package that is available somewhere in the file system.

A package is a Cargo.toml file plus all the files that are part of it.

Methods

impl Package

pub fn new(manifest: Manifest, manifest_path: &Path) -> Package

Create a package from a manifest and its location

pub fn for_path(manifest_path: &Path, config: &Config) -> CargoResult<Package>

Calculate the Package from the manifest path (and cargo configuration).

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn dependencies(&self) -> &[Dependency]

Get the manifest dependencies

pub fn manifest(&self) -> &Manifest

Get the manifest

pub fn manifest_path(&self) -> &Path

Get the path to the manifest

pub fn name(&self) -> &str

Get the name of the package

pub fn package_id(&self) -> &PackageId

Get the PackageId object for the package (fully defines a packge)

pub fn root(&self) -> &Path

Get the root folder of the package

pub fn summary(&self) -> &Summary

Get the summary for the package

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn targets(&self) -> &[Target]

Get the targets specified in the manifest

pub fn version(&self) -> &Version

Get the current package version

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn authors(&self) -> &Vec<String>

Get the package authors

pub fn publish(&self) -> &Option<Vec<String>>

Whether the package is set to publish

pub fn has_custom_build(&self) -> bool

Whether the package uses a custom build script for any target

pub fn find_closest_target(
&self,
target: &str,
is_expected_kind: fn(_: &Target) -> bool
) -> Option<&Target>

pub fn map_source(
self,
to_replace: &SourceId,
replace_with: &SourceId
) -> Package

pub fn to_registry_toml(&self) -> CargoResult<String>