Struct cargo::core::package::Package

pub struct Package {
    inner: Rc<PackageInner>,
}

Information about a package that is available somewhere in the file system.

A package is a Cargo.toml file plus all the files that are part of it.

Fields
inner: Rc<PackageInner>
Implementations

impl Package

pub fn new(manifest: Manifest, manifest_path: &Path) -> Package

Creates a package from a manifest and its location.

pub fn dependencies(&self) -> &[Dependency]

Gets the manifest dependencies.

pub fn manifest(&self) -> &Manifest

Gets the manifest.

pub fn manifest_mut(&mut self) -> &mut Manifest

Gets the manifest.

pub fn manifest_path(&self) -> &Path

Gets the path to the manifest.

pub fn name(&self) -> InternedString

Gets the name of the package.

pub fn package_id(&self) -> PackageId

Gets the PackageId object for the package (fully defines a package).

pub fn root(&self) -> &Path

Gets the root folder of the package.

pub fn summary(&self) -> &Summary

Gets the summary for the package.

pub fn targets(&self) -> &[Target]

Gets the targets specified in the manifest.

pub fn library(&self) -> Option<&Target>

Gets the library crate for this package, if it exists.

pub fn version(&self) -> &Version

Gets the current package version.

pub fn authors(&self) -> &Vec<String>

Gets the package authors.

pub fn publish(&self) -> &Option<Vec<String>>

Returns None if the package is set to publish. Returns Some(allowed_registries) if publishing is limited to specified registries or if package is set to not publish.

pub fn proc_macro(&self) -> bool

Returns true if this package is a proc-macro.

pub fn rust_version(&self) -> Option<&str>

Gets the package’s minimum Rust version.

pub fn has_custom_build(&self) -> bool

Returns true if the package uses a custom build script for any target.

pub fn map_source(self, to_replace: SourceId, replace_with: SourceId) -> Package

pub fn to_registry_toml(&self, ws: &Workspace<'_>) -> CargoResult<String>

pub fn include_lockfile(&self) -> bool

Returns if package should include Cargo.lock.

pub fn serialized(&self) -> SerializedPackage