Struct cargo::core::package::Package

pub struct Package {
    // some fields omitted
}

Informations about a package that is available somewhere in the file system.

A package is a Cargo.toml file, plus all the files that are part of it.

Methods

impl Package

fn new(manifest: Manifest, manifest_path: &Path) -> Package

fn dependencies(&self) -> &[Dependency]

fn manifest(&self) -> &Manifest

fn manifest_path(&self) -> &Path

fn name(&self) -> &str

fn package_id(&self) -> &PackageId

fn root(&self) -> &Path

fn summary(&self) -> &Summary

fn targets(&self) -> &[Target]

fn version(&self) -> &Version

fn has_custom_build(&self) -> bool

fn generate_metadata(&self) -> Metadata