Struct cargo::core::package::PackageSet

pub struct PackageSet<'cfg> { /* private fields */ }

A set of packages, with the intent to download.

This is primarily used to convert a set of PackageIds to Packages. It will download as needed, or used the cached download if available.

Implementations

impl<'cfg> PackageSet<'cfg>

pub fn new(
package_ids: &[PackageId],
sources: SourceMap<'cfg>,
config: &'cfg Config
) -> CargoResult<PackageSet<'cfg>>

pub fn package_ids(&self) -> impl Iterator<Item = PackageId> + '_

pub fn packages(&self) -> impl Iterator<Item = &Package>

pub fn enable_download<'a>(&'a self) -> CargoResult<Downloads<'a, 'cfg>>

pub fn get_one(&self, id: PackageId) -> CargoResult<&Package>

pub fn get_many(
&self,
ids: impl IntoIterator<Item = PackageId>
) -> CargoResult<Vec<&Package>>

pub fn download_accessible(
&self,
resolve: &Resolve,
root_ids: &[PackageId],
has_dev_units: HasDevUnits,
requested_kinds: &[CompileKind],
target_data: &RustcTargetData<'cfg>,
force_all_targets: ForceAllTargets
) -> CargoResult<()>

Downloads any packages accessible from the give root ids.

pub fn sources(&self) -> Ref<'_, SourceMap<'cfg>>

pub fn sources_mut(&self) -> RefMut<'_, SourceMap<'cfg>>

pub fn add_set(&mut self, set: PackageSet<'cfg>)

Merge the given set into self.