cargo::core::registry
Struct PackageRegistry

pub struct PackageRegistry<'gctx> {
    gctx: &'gctx GlobalContext,
    sources: SourceMap<'gctx>,
    overrides: Vec<SourceId>,
    source_ids: HashMap<SourceId, (SourceId, Kind)>,
    locked: HashMap<(SourceId, InternedString), Vec<(PackageId, Vec<PackageId>)>>,
    yanked_whitelist: HashSet<PackageId>,
    source_config: SourceConfigMap<'gctx>,
    patches: HashMap<CanonicalUrl, Vec<Summary>>,
    patches_locked: bool,
    patches_available: HashMap<CanonicalUrl, Vec<PackageId>>,
}

This structure represents a registry of known packages. It internally contains a number of Source instances which are used to load a Package from.

The resolution phase of Cargo uses this to drive knowledge about new packages as well as querying for lists of new packages. It is here that sources are updated (e.g., network operations) and overrides/patches are handled.

The general idea behind this registry is that it is centered around the SourceMap structure, contained within which is a mapping of a SourceId to a Source. Each Source in the map has been updated (using network operations if necessary) and is ready to be queried for packages.

Fields
gctx: &'gctx GlobalContextsources: SourceMap<'gctx>overrides: Vec<SourceId>

A list of sources which are considered “path-overrides” which take precedent when querying for packages.

source_ids: HashMap<SourceId, (SourceId, Kind)>

Use for tracking sources that are already loaded into the registry.

locked: HashMap<(SourceId, InternedString), Vec<(PackageId, Vec<PackageId>)>>

This is constructed via PackageRegistry::register_lock. See also LockedMap.

yanked_whitelist: HashSet<PackageId>

A group of packages tha allows to use even when yanked.

source_config: SourceConfigMap<'gctx>patches: HashMap<CanonicalUrl, Vec<Summary>>

Patches registered during calls to PackageRegistry::patch.

These are available for query after calling PackageRegistry::lock_patches, which locks them all to specific versions.

patches_locked: bool

Whether patches are locked. That is, they are available to resolution.

See PackageRegistry::lock_patches and PackageRegistry::patch for more.

patches_available: HashMap<CanonicalUrl, Vec<PackageId>>

Patches available for each source.

This is for determining whether a dependency entry from a lockfile happened through [patch], during calls to lock to rewrite summaries to point directly at these patched entries.

This is constructed during calls to PackageRegistry::patch, along with the patches field, thoough these entries never get locked.

Implementations

impl<'gctx> PackageRegistry<'gctx>

pub fn new_with_source_config( gctx: &'gctx GlobalContext, source_config: SourceConfigMap<'gctx>, ) -> CargoResult<PackageRegistry<'gctx>>

pub fn get(self, package_ids: &[PackageId]) -> CargoResult<PackageSet<'gctx>>

fn ensure_loaded(&mut self, namespace: SourceId, kind: Kind) -> CargoResult<()>

Ensures the Source of the given SourceId is loaded. If not, this will block until the source is ready.

pub fn add_sources( &mut self, ids: impl IntoIterator<Item = SourceId>, ) -> CargoResult<()>

pub fn add_preloaded(&mut self, source: Box<dyn Source + 'gctx>)

Adds a source which will be locked. Useful for path sources such as the source of a workspace member.

fn add_source(&mut self, source: Box<dyn Source + 'gctx>, kind: Kind)

Adds a source to the registry.

pub fn add_override(&mut self, source: Box<dyn Source + 'gctx>)

Adds a source from a path override.

pub fn add_to_yanked_whitelist(&mut self, iter: impl Iterator<Item = PackageId>)

Allows a group of package to be available to query even if they are yanked.

pub fn clear_lock(&mut self)

remove all residual state from previous lock files.

pub fn register_lock(&mut self, id: PackageId, deps: Vec<PackageId>)

Registers one “locked package” to the registry, for guiding the dependency resolution. See LockedMap for more.

pub fn patch( &mut self, url: &Url, patch_deps: &[PatchDependency<'_>], ) -> CargoResult<Vec<(Dependency, PackageId)>>

Insert a [patch] section into this registry.

This method will insert a [patch] section for the url specified, with the given list of dependencies. The url specified is the URL of the source to patch (for example this is crates-io in the manifest). The deps is an array of all the entries in the [patch] section of the manifest.

Here the patch_deps will be resolved to a precise version and stored internally for future calls to query below.

Note that the patch list specified here will not be available to Registry::query until PackageRegistry::lock_patches is called below, which should be called once all patches have been added.

The return value is a Vec of patches that should not be locked. This happens when the patch is locked, but the patch has been updated so the locked value is no longer correct.

pub fn lock_patches(&mut self)

Lock all patch summaries added via patch, making them available to resolution via Registry::query.

pub fn patches(&self) -> &HashMap<CanonicalUrl, Vec<Summary>>

Gets all patches grouped by the source URLs they are going to patch.

These patches are mainly collected from patch. They might not be the same as patches actually used during dependency resolving.

fn load(&mut self, source_id: SourceId, kind: Kind) -> CargoResult<()>

Loads the Source for a given SourceId to this registry, making them available to resolution.

fn query_overrides( &mut self, dep: &Dependency, ) -> Poll<CargoResult<Option<IndexSummary>>>

Queries path overrides from this registry.

pub fn lock(&self, summary: Summary) -> Summary

This function is used to transform a summary to another locked summary if possible. This is where the concept of a lock file comes into play.

If a summary points at a package ID which was previously locked, then we override the summary’s ID itself, as well as all dependencies, to be rewritten to the locked versions. This will transform the summary’s source to a precise source (listed in the locked version) as well as transforming all of the dependencies from range requirements on imprecise sources to exact requirements on precise sources.

If a summary does not point at a package ID which was previously locked, or if any dependencies were added and don’t have a previously listed version, we still want to avoid updating as many dependencies as possible to keep the graph stable. In this case we map all of the summary’s dependencies to be rewritten to a locked version wherever possible. If we’re unable to map a dependency though, we just pass it on through.

fn warn_bad_override( &self, override_summary: &Summary, real_summary: &Summary, ) -> CargoResult<()>