Trait cargo::core::registry::Registry

pub trait Registry {
    fn query(
        &mut self, 
        dep: &Dependency, 
        f: &mut FnMut(Summary)
    ) -> CargoResult<()>;

    fn supports_checksums(&self) -> bool;

    fn requires_precise(&self) -> bool;

    fn query_vec(&mut self, dep: &Dependency) -> CargoResult<Vec<Summary>> { ... }
}

Source of information about a group of packages.

See also core::Source.

Required Methods

fn query(&mut self, dep: &Dependency, f: &mut FnMut(Summary)) -> CargoResult<()>

Attempt to find the packages that match a dependency request.

fn supports_checksums(&self) -> bool

Returns whether or not this registry will return summaries with checksums listed.

fn requires_precise(&self) -> bool

Returns whether or not this registry will return summaries with the precise field in the source id listed.

Provided Methods

fn query_vec(&mut self, dep: &Dependency) -> CargoResult<Vec<Summary>>