cargo::core::resolver::encode
Struct EncodableSourceId

pub struct EncodableSourceId {
    inner: SourceId,
    encoded: bool,
}

Pretty much equivalent to SourceId with a different serialization method.

The serialization for SourceId doesn’t do URL encode for parameters. In contrast, this type is aware of that whenever ResolveVersion allows us to do so (v4 or later).

Fields
inner: SourceIdencoded: bool

We don’t care about the deserialization of this, as the url crate will always decode as the URL was encoded. Only when a Resolve turns into a EncodableResolve will it set the value accordingly via encodable_source_id.

Implementations

impl EncodableSourceId

fn new(inner: SourceId) -> Self

Creates a EncodableSourceId that always encodes URL params.

fn without_url_encoded(inner: SourceId) -> Self

Creates a EncodableSourceId that doesn’t encode URL params. This is for backward compatibility for order lockfile version.

fn as_url(&self) -> impl Display + '_

Encodes the inner SourceId as a URL.