Struct cargo::core::resolver::features::ResolvedFeatures

pub struct ResolvedFeatures {
    activated_features: HashMap<(PackageId, FeaturesFor), BTreeSet<InternedString>>,
    activated_dependencies: HashMap<(PackageId, FeaturesFor), BTreeSet<InternedString>>,
    opts: FeatureOpts,
}

Set of all activated features for all packages in the resolve graph.

Fields
activated_features: HashMap<(PackageId, FeaturesFor), BTreeSet<InternedString>>activated_dependencies: HashMap<(PackageId, FeaturesFor), BTreeSet<InternedString>>

Optional dependencies that should be built.

The value is the name_in_toml of the dependencies.

opts: FeatureOpts
Implementations

impl ResolvedFeatures

pub fn activated_features(
&self,
pkg_id: PackageId,
features_for: FeaturesFor
) -> Vec<InternedString>

Returns the list of features that are enabled for the given package.

pub fn is_dep_activated(
&self,
pkg_id: PackageId,
features_for: FeaturesFor,
dep_name: InternedString
) -> bool

Returns if the given dependency should be included.

This handles dependencies disabled via cfg expressions and optional dependencies which are not enabled.

pub fn activated_features_unverified(
&self,
pkg_id: PackageId,
features_for: FeaturesFor
) -> Option<Vec<InternedString>>

Variant of activated_features that returns None if this is not a valid pkg_id/is_build combination. Used in places which do not know which packages are activated (like cargo clean).

fn activated_features_int(
&self,
pkg_id: PackageId,
features_for: FeaturesFor
) -> CargoResult<Vec<InternedString>>

pub fn compare_legacy(&self, legacy: &ResolvedFeatures) -> DiffMap

Compares the result against the original resolver behavior.

Used by cargo fix --edition to display any differences.