Struct cargo::core::resolver::Resolve

pub struct Resolve { /* fields omitted */ }

Represents a fully resolved package dependency graph. Each node in the graph is a package and edges represent dependencies between packages.

Each instance of Resolve also understands the full set of features used for each package.

Methods

impl Resolve

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn path_to_top<'a>(&'a self, pkg: &'a PackageId) -> Vec<&'a PackageId>

Resolves one of the paths from the given dependent package up to the root.

pub fn register_used_patches(&mut self, patches: &HashMap<Url, Vec<Summary>>)

pub fn merge_from(&mut self, previous: &Resolve) -> CargoResult<()>

pub fn iter(&self) -> Nodes<PackageId>

Important traits for Deps<'a>
impl<'a> Iterator for Deps<'a> type Item = &'a PackageId;

pub fn deps(&self, pkg: &PackageId) -> Deps

Important traits for DepsNotReplaced<'a>
impl<'a> Iterator for DepsNotReplaced<'a> type Item = &'a PackageId;

pub fn deps_not_replaced(&self, pkg: &PackageId) -> DepsNotReplaced

pub fn replacement(&self, pkg: &PackageId) -> Option<&PackageId>

pub fn replacements(&self) -> &HashMap<PackageId, PackageId>

pub fn features(&self, pkg: &PackageId) -> &HashSet<String>

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn features_sorted(&self, pkg: &PackageId) -> Vec<&str>

pub fn query(&self, spec: &str) -> CargoResult<&PackageId>

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn unused_patches(&self) -> &[PackageId]