Struct cargo::core::resolver::Resolve

pub struct Resolve {
    // some fields omitted
}

Represents a fully resolved package dependency graph. Each node in the graph is a package and edges represent dependencies between packages.

Each instance of Resolve also understands the full set of features used for each package as well as what the root package is.

Methods

impl Resolve

fn copy_metadata(&mut self, other: &Resolve)

fn iter(&self) -> Nodes<PackageId>

fn root(&self) -> &PackageId

fn deps(&self, pkg: &PackageId) -> Option<Edges<PackageId>>

fn query(&self, spec: &str) -> CargoResult<&PackageId>

fn features(&self, pkg: &PackageId) -> Option<&HashSet<String>>