Type Alias cargo::core::resolver::types::FeaturesSet

pub type FeaturesSet = Rc<BTreeSet<InternedString>>;

The preferred way to store the set of activated features for a package. This is sorted so that it impls Hash, and owns its contents, needed so it can be part of the key for caching in the DepsCache. It is also cloned often as part of Context, hence the RC. im-rs::OrdSet was slower of small sets like this, but this can change with improvements to std, im, or llvm. Using a consistent type for this allows us to use the highly optimized comparison operators like is_subset at the interfaces.

Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference’s “Type Layout” chapter for details on type layout guarantees.

Size:8 bytes