Struct cargo::core::shell::Shell

pub struct Shell { /* fields omitted */ }

An abstraction around a Writeable object that remembers preferences for output verbosity and color.

Methods

impl Shell

pub fn new() -> Shell

Create a new shell (color choice and verbosity), defaulting to 'auto' color and verbose output.

pub fn from_write(out: Box<Write>) -> Shell

Create a shell from a plain writable object, with no color, and max verbosity.

pub fn err_width(&self) -> Option<usize>

Returns the width of the terminal in spaces, if any

pub fn is_err_tty(&self) -> bool

Returns whether stderr is a tty

pub fn err(&mut self) -> &mut Write

Get a reference to the underlying writer

pub fn status<T, U>(&mut self, status: T, message: U) -> CargoResult<()> where
T: Display,
U: Display,

Shortcut to right-align and color green a status message.

pub fn status_header<T>(&mut self, status: T) -> CargoResult<()> where
T: Display,

pub fn status_with_color<T, U>(
&mut self,
status: T,
message: U,
color: Color
) -> CargoResult<()> where
T: Display,
U: Display,

Shortcut to right-align a status message.

pub fn verbose<F>(&mut self, callback: F) -> CargoResult<()> where
F: FnMut(&mut Shell) -> CargoResult<()>,

Run the callback only if we are in verbose mode

pub fn concise<F>(&mut self, callback: F) -> CargoResult<()> where
F: FnMut(&mut Shell) -> CargoResult<()>,

Run the callback if we are not in verbose mode.

pub fn error<T: Display>(&mut self, message: T) -> CargoResult<()>

Print a red 'error' message

pub fn warn<T: Display>(&mut self, message: T) -> CargoResult<()>

Print an amber 'warning' message

pub fn set_verbosity(&mut self, verbosity: Verbosity)

Update the verbosity of the shell

pub fn verbosity(&self) -> Verbosity

Get the verbosity of the shell

pub fn set_color_choice(&mut self, color: Option<&str>) -> CargoResult<()>

Update the color choice (always, never, or auto) from a string.

pub fn color_choice(&self) -> ColorChoice

Get the current color choice

If we are not using a color stream, this will always return Never, even if the color choice has been set to something else.