Module cargo::core::source

Structs

SourceId
Unique identifier for a source of packages.

SourceMap
A HashMap of SourceId -> Box<Source>.

Enums

GitReference
Information to find a specific commit in a Git repository.

MaybePackage

Traits

Source
Something that finds and downloads remote packages based on names and versions.