Module cargo::core::source

Structs

SourceId
Unique identifier for a source of packages.

SourceMap
A HashMap of SourceId -> Box<Source>

SourcesMut
A std::collection::hash_map::IterMut for SourceMap

Enums

GitReference
Information to find a specific commit in a git repository

Traits

Source
A Source finds and downloads remote packages based on names and versions.

Type Definitions

Sources
A std::collection::hash_map::Values for SourceMap