Struct cargo::core::source::SourceId

pub struct SourceId {
    // some fields omitted
}

Unique identifier for a source of packages.

Methods

impl SourceId

fn from_url(string: &str) -> CargoResult<SourceId>

Parses a source URL and returns the corresponding ID.

Example

use cargo::core::SourceId;
SourceId::from_url("git+https://github.com/alexcrichton/\
                    libssh2-static-sys#80e71a3021618eb05\
                    656c58fb7c5ef5f12bc747f");

fn to_url(&self) -> String

fn for_path(path: &Path) -> CargoResult<SourceId>

fn for_git(url: &Url, reference: GitReference) -> SourceId

fn for_registry(url: &Url) -> SourceId

fn for_local_registry(path: &Path) -> CargoResult<SourceId>

fn for_directory(path: &Path) -> CargoResult<SourceId>

fn crates_io(config: &Config) -> CargoResult<SourceId>

Returns the SourceId corresponding to the main repository.

This is the main cargo registry by default, but it can be overridden in a .cargo/config.

fn url(&self) -> &Url

fn is_path(&self) -> bool

fn is_registry(&self) -> bool

fn is_git(&self) -> bool

fn load<'a>(&self, config: &'a Config) -> Box<Source + 'a>

Creates an implementation of Source corresponding to this ID.

fn precise(&self) -> Option<&str>

fn git_reference(&self) -> Option<&GitReference>

fn with_precise(&self, v: Option<String>) -> SourceId

fn is_default_registry(&self) -> bool