Struct cargo::core::Workspace

pub struct Workspace<'cfg> { /* fields omitted */ }

The core abstraction in Cargo for working with a workspace of crates.

A workspace is often created very early on and then threaded through all other functions. It's typically through this object that the current package is loaded and/or learned about.

Methods

impl<'cfg> Workspace<'cfg>

pub fn new(
manifest_path: &Path,
config: &'cfg Config
) -> CargoResult<Workspace<'cfg>>

Creates a new workspace given the target manifest pointed to by manifest_path.

This function will construct the entire workspace by determining the root and all member packages. It will then validate the workspace before returning it, so Ok is only returned for valid workspaces.

pub fn current_manifest(&self) -> &Path

pub fn ephemeral(
package: Package,
config: &'cfg Config,
target_dir: Option<Filesystem>,
require_optional_deps: bool
) -> CargoResult<Workspace<'cfg>>

Creates a "temporary workspace" from one package which only contains that package.

This constructor will not touch the filesystem and only creates an in-memory workspace. That is, all configuration is ignored, it's just intended for that one package.

This is currently only used in niche situations like cargo install or cargo package.

pub fn current(&self) -> CargoResult<&Package>

Returns the current package of this workspace.

Note that this can return an error if it the current manifest is actually a "virtual Cargo.toml", in which case an error is returned indicating that something else should be passed.

pub fn current_opt(&self) -> Option<&Package>

pub fn is_virtual(&self) -> bool

pub fn config(&self) -> &'cfg Config

Returns the Config this workspace is associated with.

pub fn profiles(&self) -> &Profiles

pub fn root(&self) -> &Path

Returns the root path of this workspace.

That is, this returns the path of the directory containing the Cargo.toml which is the root of this workspace.

pub fn target_dir(&self) -> Filesystem

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn root_replace(&self) -> &[(PackageIdSpec, Dependency)]

Returns the root [replace] section of this workspace.

This may be from a virtual crate or an actual crate.

pub fn root_patch(&self) -> &HashMap<Url, Vec<Dependency>>

Returns the root [patch] section of this workspace.

This may be from a virtual crate or an actual crate.

Important traits for Members<'a, 'cfg>
impl<'a, 'cfg> Iterator for Members<'a, 'cfg> type Item = &'a Package;

pub fn members<'a>(&'a self) -> Members<'a, 'cfg>

Returns an iterator over all packages in this workspace

Important traits for Members<'a, 'cfg>
impl<'a, 'cfg> Iterator for Members<'a, 'cfg> type Item = &'a Package;

pub fn default_members<'a>(&'a self) -> Members<'a, 'cfg>

Returns an iterator over default packages in this workspace

pub fn is_ephemeral(&self) -> bool

pub fn require_optional_deps(&self) -> bool