Struct cargo::core::WorkspaceRootConfig

pub struct WorkspaceRootConfig { /* fields omitted */ }

Intermediate configuration of a workspace root in a manifest.

Knows the Workspace Root path, as well as members and exclude lists of path patterns, which together tell if some path is recognized as a member by this root or not.

Methods

impl WorkspaceRootConfig

pub fn new(
root_dir: &Path,
members: &Option<Vec<String>>,
default_members: &Option<Vec<String>>,
exclude: &Option<Vec<String>>
) -> WorkspaceRootConfig

Create a new Intermediate Workspace Root configuration.