Struct cargo::core::summary::Summary

pub struct Summary { /* fields omitted */ }

Subset of a Manifest. Contains only the most important informations about a package.

Summaries are cloned, and should not be mutated after creation

Methods

impl Summary

pub fn new(
pkg_id: PackageId,
dependencies: Vec<Dependency>,
features: BTreeMap<String, Vec<String>>
) -> CargoResult<Summary>

pub fn package_id(&self) -> &PackageId

pub fn name(&self) -> &str

pub fn version(&self) -> &Version

pub fn source_id(&self) -> &SourceId

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn dependencies(&self) -> &[Dependency]

pub fn features(&self) -> &BTreeMap<String, Vec<String>>

pub fn checksum(&self) -> Option<&str>

pub fn override_id(self, id: PackageId) -> Summary

pub fn set_checksum(self, cksum: String) -> Summary

pub fn map_dependencies<F>(self, f: F) -> Summary where
F: FnMut(Dependency) -> Dependency,

pub fn map_source(
self,
to_replace: &SourceId,
replace_with: &SourceId
) -> Summary