cargo::core
Module workspace

Structs

LookBehind 🔒

LookBehindWindow 🔒

Packages 🔒

Workspace

The core abstraction in Cargo for working with a workspace of crates.

WorkspaceRootConfig

Intermediate configuration of a workspace root in a manifest.

Enums

MaybePackage

WorkspaceConfig

Configuration of a workspace in a manifest.

Functions

find_root_iter 🔒

find_workspace_root

Finds the path of the root of the workspace.

find_workspace_root_with_loader 🔒

Finds the path of the root of the workspace.

read_root_pointer 🔒

resolve_relative_path