Module cargo::ops::cargo_add

Core of cargo-add command

Modules

crate_spec ğŸ”’

Crate name parsing.

Structs

AddOptions

Information on what dependencies should be added

DepOp

Dependency entry operation

DependencyUI

Track presentation-layer information with the editable representation of a [dependencies] entry (Dependency)

Functions

add

Add dependencies to a manifest

check_invalid_ws_keys ğŸ”’

When { workspace = true } you cannot define other keys that configure the source of the dependency such as version, registry, registry-index, path, git, branch, tag, rev, or package. You can also not define default-features.

find_workspace_dep ğŸ”’

get_existing_dependency ğŸ”’

Provide the existing dependency for the target table

get_latest_dependency ğŸ”’

infer_package_for_git_source ğŸ”’

is_sorted ğŸ”’

populate_available_features ğŸ”’

Lookup available features

populate_dependency ğŸ”’

precise_version ğŸ”’

Convert a semver::VersionReq into a rendered semver::Version if all fields are fully specified.

print_action_msg ğŸ”’

print_dep_table_msg ğŸ”’

resolve_dependency ğŸ”’

select_package ğŸ”’