cargo::ops::cargo_compile::unit_generator
Struct UnitGenerator

pub(super) struct UnitGenerator<'a, 'gctx> {

    pub ws: &'a Workspace<'gctx>,
    pub packages: &'a [&'a Package],
    pub target_data: &'a RustcTargetData<'gctx>,
    pub filter: &'a CompileFilter,
    pub requested_kinds: &'a [CompileKind],
    pub explicit_host_kind: CompileKind,
    pub mode: CompileMode,
    pub resolve: &'a Resolve,
    pub workspace_resolve: &'a Option<Resolve>,
    pub resolved_features: &'a ResolvedFeatures,
    pub package_set: &'a PackageSet<'gctx>,
    pub profiles: &'a Profiles,
    pub interner: &'a UnitInterner,
    pub has_dev_units: HasDevUnits,

}

The context needed for generating root units, which are packages the user has requested to compile.

To generate a full UnitGraph, generally you need to call generate_root_units first, and then provide the output to build_unit_dependencies.

Fields
ws: &'a Workspace<'gctx>packages: &'a [&'a Package]target_data: &'a RustcTargetData<'gctx>filter: &'a CompileFilterrequested_kinds: &'a [CompileKind]explicit_host_kind: CompileKindmode: CompileModeresolve: &'a Resolveworkspace_resolve: &'a Option<Resolve>resolved_features: &'a ResolvedFeaturespackage_set: &'a PackageSet<'gctx>profiles: &'a Profilesinterner: &'a UnitInternerhas_dev_units: HasDevUnits
Implementations

impl<'a> UnitGenerator<'a, '_>

fn new_units( &self, pkg: &Package, target: &Target, initial_target_mode: CompileMode, ) -> Vec<Unit>

Helper for creating a list of Unit structures

fn filter_default_targets<'b>(&self, targets: &'b [Target]) -> Vec<&'b Target>

Given a list of all targets for a package, filters out only the targets that are automatically included when the user doesn’t specify any targets.

fn filter_targets( &self, predicate: impl Fn(&Target) -> bool, requires_features: bool, mode: CompileMode, ) -> Vec<Proposal<'a>>

Filters the set of all possible targets based on the provided predicate.

fn find_named_targets( &self, target_name: &str, target_desc: &'static str, is_expected_kind: fn(_: &Target) -> bool, mode: CompileMode, ) -> CargoResult<Vec<Proposal<'a>>>

Finds the targets for a specifically named target.

fn list_rule_targets( &self, rule: &FilterRule, target_desc: &'static str, is_expected_kind: fn(_: &Target) -> bool, mode: CompileMode, ) -> CargoResult<Vec<Proposal<'a>>>

Returns a list of proposed targets based on command-line target selection flags.

fn create_proposals(&self) -> CargoResult<Vec<Proposal<'_>>>

Create a list of proposed targets given the context in UnitGenerator

fn create_docscrape_proposals( &self, doc_units: &[Unit], ) -> CargoResult<Vec<Proposal<'a>>>

Proposes targets from which to scrape examples for documentation

fn unmatched_target_filters(&self, units: &[Unit]) -> CargoResult<()>

Checks if the unit list is empty and the user has passed any combination of –tests, –examples, –benches or –bins, and we didn’t match on any targets. We want to emit a warning to make sure the user knows that this run is a no-op, and their code remains unchecked despite cargo not returning any errors

fn validate_required_features( &self, target_name: &str, required_features: &[String], summary: &Summary, ) -> CargoResult<()>

Warns if a target’s required-features references a feature that doesn’t exist.

This is a warning because historically this was not validated, and it would cause too much breakage to make it an error.

fn proposals_to_units( &self, proposals: Vec<Proposal<'_>>, ) -> CargoResult<Vec<Unit>>

Converts proposals to units based on each target’s required features.

pub fn generate_root_units(&self) -> CargoResult<Vec<Unit>>

Generates all the base units for the packages the user has requested to compile. Dependencies for these units are computed later in unit_dependencies.

pub fn generate_scrape_units( &self, doc_units: &[Unit], ) -> CargoResult<Vec<Unit>>

Generates units specifically for doc-scraping.

This requires a separate entrypoint from generate_root_units because it takes the documented units as input.