cargo::ops
Module cargo_install

Structs

InstallablePackage 🔒

Transaction 🔒

Functions

install

install_list

Display a list of installed binaries.

installed_exact_package 🔒

Checks if vers can only be satisfied by exactly one version of a package in a registry, and it’s already installed. If this is the case, we can skip interacting with a registry to check if newer versions may be installable, as no newer version can exist.

is_installed 🔒

make_warning_about_missing_features 🔒

make_ws_rustc_target 🔒

remove_orphaned_bins 🔒

Removes executables that are no longer part of a package that was previously installed.