cargo::ops
Module cargo_new

Structs

CargoNewConfig ğŸ”’

IgnoreList ğŸ”’

IgnoreList

MkOptions ğŸ”’

NewOptions

SourceFileInformation ğŸ”’

Enums

NewProjectKind

VersionControl

Functions

calculate_new_project_kind ğŸ”’

can_be_workspace_member ğŸ”’

check_name ğŸ”’

See also util::toml::embedded::sanitize_name

check_path ğŸ”’

Checks if the path contains any invalid PATH env characters.

detect_source_paths_and_types ğŸ”’

get_display_path ğŸ”’

get_name ğŸ”’

init

init_vcs ğŸ”’

Initializes the correct VCS system based on the provided config.

mk ğŸ”’

new

plan_new_source_file ğŸ”’

update_manifest_with_inherited_workspace_package_keys ğŸ”’

update_manifest_with_new_member ğŸ”’

Adds the new package member to the [workspace.members] array.

write_ignore_file ğŸ”’

Writes the ignore file to the given directory. If the ignore file for the given vcs system already exists, its content is read and duplicate ignore file entries are filtered out.