cargo::ops

Module cargo_new

Structs

CargoNewConfig 🔒

IgnoreList 🔒

IgnoreList

MkOptions 🔒

NewOptions

SourceFileInformation 🔒

Enums

NewProjectKind

VersionControl

Functions

calculate_new_project_kind 🔒

can_be_workspace_member 🔒

check_name 🔒

See also util::toml::embedded::sanitize_name

check_path 🔒

Checks if the path contains any invalid PATH env characters.

detect_source_paths_and_types 🔒

get_display_path 🔒

get_name 🔒

init

init_vcs 🔒

Initializes the correct VCS system based on the provided config.

mk 🔒

new

plan_new_source_file 🔒

update_manifest_with_inherited_workspace_package_keys 🔒

update_manifest_with_new_member 🔒

Adds the new package member to the [workspace.members] array.

write_ignore_file 🔒

Writes the ignore file to the given directory. If the ignore file for the given vcs system already exists, its content is read and duplicate ignore file entries are filtered out.