cargo::ops
Module common_for_install_and_uninstall

Structs

CrateListingV1

Tracking information for the set of installed packages.

CrateListingV2 🔒

Tracking information for the set of installed packages.

InstallInfo 🔒

Tracking information for the installation of a single package.

InstallTracker

On-disk tracking for which package installed which binary.

Functions

exe_names

Helper to get the executable names from a filter.

feature_set 🔒

Helper to convert features to a BTreeSet.

one 🔒

Get one element from the iterator. Returns None if none left. Returns error if there is more than one item in the iterator.

path_source

Determines the PathSource from a SourceId.

resolve_root

Determines the root directory where installation is done.

select_dep_pkg

Gets a Package based on command-line requirements.

select_pkg