cargo::ops::common_for_install_and_uninstall
Struct InstallInfo

struct InstallInfo {
    version_req: Option<String>,
    bins: BTreeSet<String>,
    features: BTreeSet<String>,
    all_features: bool,
    no_default_features: bool,
    profile: String,
    target: Option<String>,
    rustc: Option<String>,
    other: BTreeMap<String, Value>,
}

Tracking information for the installation of a single package.

This tracks the settings that were used when the package was installed. Future attempts to install the same package will check these settings to determine if it needs to be rebuilt/reinstalled. If nothing has changed, then Cargo will inform the user that it is “up to date”.

This is only used for the v2 format.

Fields
version_req: Option<String>

Version requested via --version. None if --version not specified. Currently not used, possibly may be used in the future.

bins: BTreeSet<String>

Set of binary names installed.

features: BTreeSet<String>

Set of features explicitly enabled.

all_features: boolno_default_features: boolprofile: String

Either “debug” or “release”.

target: Option<String>

The installation target. Either the host or the value specified in --target. None if unknown (when loading from v1).

rustc: Option<String>

Output of rustc -V. None if unknown (when loading from v1). Currently not used, possibly may be used in the future.

other: BTreeMap<String, Value>

Forwards compatibility.

Implementations

impl InstallInfo

fn from_v1(set: &BTreeSet<String>) -> InstallInfo

fn is_up_to_date( &self, opts: &CompileOptions, target: &str, exes: &BTreeSet<String>, ) -> bool

Determine if this installation is “up to date”, or if it needs to be reinstalled.

This does not do Package/Source/Version checking.