Function cargo::ops::fix::rustfix_and_fix

fn rustfix_and_fix(
    files: &mut HashMap<String, FixedFile>,
    rustc: &ProcessBuilder,
    filename: &Path,
    args: &FixArgs,
    gctx: &GlobalContext,
) -> CargoResult<(Output, bool)>

Executes rustc to apply one round of suggestions to the crate in question.

This will fill in the fixes map with original code, suggestions applied, and any errors encountered while fixing files.