cargo::ops::fix
Struct FixArgs

struct FixArgs {
    file: PathBuf,
    prepare_for_edition: Option<Edition>,
    idioms: bool,
    enabled_edition: Option<Edition>,
    other: Vec<OsString>,
    rustc: PathBuf,
    sysroot: Option<PathBuf>,
}

Various command-line options and settings used when cargo is running as a proxy for rustc during the fix operation.

Fields
file: PathBuf

This is the .rs file that is being fixed.

prepare_for_edition: Option<Edition>

If --edition is used to migrate to the next edition, this is the edition we are migrating towards.

idioms: bool

true if --edition-idioms is enabled.

enabled_edition: Option<Edition>

The current edition.

None if on 2015.

other: Vec<OsString>

Other command-line arguments not reflected by other fields in FixArgs.

rustc: PathBuf

Path to the rustc executable.

sysroot: Option<PathBuf>

Path to host sysroot.

Implementations

impl FixArgs

fn get() -> CargoResult<FixArgs>

fn from_args(argv: impl IntoIterator<Item = OsString>) -> CargoResult<Self>

fn apply(&self, cmd: &mut ProcessBuilder)

fn can_run_rustfix(&self, gctx: &GlobalContext) -> CargoResult<bool>

Validates the edition, and sends a message indicating what is being done. Returns a flag indicating whether this fix should be run.