cargo::ops

Function fix_exec_rustc

pub fn fix_exec_rustc(gctx: &GlobalContext, lock_addr: &str) -> CargoResult<()>

Entry point for cargo running as a proxy for rustc.

This is called every time cargo is run to check if it is in proxy mode.

If there are warnings or errors, this does not return, and the process exits with the corresponding rustc exit code.

See fix_get_proxy_lock_addr