Module cargo::ops::resolve

High-level APIs for executing the resolver.

This module provides functions for running the resolver given a workspace. There are roughly 3 main functions:

resolve_ws: A simple, high-level function with no options.

resolve_ws_with_opts: A medium-level function with options like user-provided features. This is the most appropriate function to use in most cases.

resolve_with_previous: A low-level function for running the resolver, providing the most power and flexibility.

Structs

WorkspaceResolve

Result for resolve_ws_with_opts.

Constants

UNUSED_PATCH_WARNING 🔒

Functions

add_overrides

Read the paths configuration variable to discover all path overrides that have been configured.

emit_warnings_of_unused_patches 🔒

Emits warnings of unused patches case by case.

get_resolved_packages

master_branch_git_source 🔒

register_previous_locks 🔒

In this function we’re responsible for informing the registry of all locked dependencies from the previous lock file we had, resolve.

resolve_with_previous

Resolves all dependencies for a package using an optional previous instance of resolve to guide the resolution process.

resolve_with_registry 🔒

resolve_ws

Resolves all dependencies for the workspace using the previous lock file as a guide if present.

resolve_ws_with_opts

Resolves dependencies for some packages of the workspace, taking into account paths overrides and activated features.