Struct cargo::ops::BuildOutput

pub struct BuildOutput {
    pub library_paths: Vec<PathBuf>,
    pub library_links: Vec<String>,
    pub cfgs: Vec<String>,
    pub env: Vec<(String, String)>,
    pub metadata: Vec<(String, String)>,
    pub rerun_if_changed: Vec<PathBuf>,
    pub rerun_if_env_changed: Vec<String>,
    pub warnings: Vec<String>,
}

Contains the parsed output of a custom build script.

Fields
library_paths: Vec<PathBuf>

Paths to pass to rustc with the -L flag

library_links: Vec<String>

Names and link kinds of libraries, suitable for the -l flag

cfgs: Vec<String>

Various --cfg flags to pass to the compiler

env: Vec<(String, String)>

Additional environment variables to run the compiler with.

metadata: Vec<(String, String)>

Metadata to pass to the immediate dependencies

rerun_if_changed: Vec<PathBuf>

Paths to trigger a rerun of this build script.

rerun_if_env_changed: Vec<String>

Environment variables which, when changed, will cause a rebuild.

warnings: Vec<String>

Warnings generated by this build,

Methods

impl BuildOutput

pub fn parse_file(
path: &Path,
pkg_name: &str,
root_output_when_generated: &Path,
root_output: &Path
) -> CargoResult<BuildOutput>

pub fn parse(
input: &[u8],
pkg_name: &str,
root_output_when_generated: &Path,
root_output: &Path
) -> CargoResult<BuildOutput>

pub fn parse_rustc_flags(
value: &str,
whence: &str
) -> CargoResult<(Vec<PathBuf>, Vec<String>)>

pub fn parse_rustc_env(
value: &str,
whence: &str
) -> CargoResult<(String, String)>