Struct cargo::ops::BuildOutput

pub struct BuildOutput {
    pub library_paths: Vec<PathBuf>,
    pub library_links: Vec<String>,
    pub cfgs: Vec<String>,
    pub metadata: Vec<(String, String)>,
}

Contains the parsed output of a custom build script.

Fields

library_paths

Paths to pass to rustc with the -L flag

library_links

Names and link kinds of libraries, suitable for the -l flag

cfgs

Various --cfg flags to pass to the compiler

metadata

Metadata to pass to the immediate dependencies

Methods

impl BuildOutput

fn parse(input: &str, pkg_name: &str) -> CargoResult<BuildOutput>

fn parse_rustc_flags(value: &str, whence: &str) -> CargoResult<(Vec<PathBuf>, Vec<String>)>