Struct cargo::ops::Compilation

pub struct Compilation<'cfg> {
    pub libraries: HashMap<PackageId, HashSet<(Target, PathBuf)>>,
    pub tests: Vec<(Package, TargetKind, String, PathBuf)>,
    pub binaries: Vec<PathBuf>,
    pub native_dirs: HashSet<PathBuf>,
    pub root_output: PathBuf,
    pub deps_output: PathBuf,
    pub host_deps_output: PathBuf,
    pub host_dylib_path: Option<PathBuf>,
    pub target_dylib_path: Option<PathBuf>,
    pub extra_env: HashMap<PackageId, Vec<(String, String)>>,
    pub to_doc_test: Vec<Package>,
    pub cfgs: HashMap<PackageId, HashSet<String>>,
    pub rustdocflags: HashMap<PackageId, Vec<String>>,
    pub target: String,
    // some fields omitted
}

A structure returning the result of a compilation.

Fields
libraries: HashMap<PackageId, HashSet<(Target, PathBuf)>>

A mapping from a package to the list of libraries that need to be linked when working with that package.

tests: Vec<(Package, TargetKind, String, PathBuf)>

An array of all tests created during this compilation.

binaries: Vec<PathBuf>

An array of all binaries created.

native_dirs: HashSet<PathBuf>

All directories for the output of native build commands.

This is currently used to drive some entries which are added to the LD_LIBRARY_PATH as appropriate.

root_output: PathBuf

Root output directory (for the local package's artifacts)

deps_output: PathBuf

Output directory for rust dependencies. May be for the host or for a specific target.

host_deps_output: PathBuf

Output directory for the rust host dependencies.

host_dylib_path: Option<PathBuf>

The path to rustc's own libstd

target_dylib_path: Option<PathBuf>

The path to libstd for the target

extra_env: HashMap<PackageId, Vec<(String, String)>>

Extra environment variables that were passed to compilations and should be passed to future invocations of programs.

to_doc_test: Vec<Package> cfgs: HashMap<PackageId, HashSet<String>>

Features per package enabled during this compilation.

rustdocflags: HashMap<PackageId, Vec<String>>

Flags to pass to rustdoc when invoked from cargo test, per package.

target: String
Methods

impl<'cfg> Compilation<'cfg>

pub fn new(config: &'cfg Config) -> Compilation<'cfg>

pub fn rustc_process(&self, pkg: &Package) -> CargoResult<ProcessBuilder>

See process.

pub fn rustdoc_process(&self, pkg: &Package) -> CargoResult<ProcessBuilder>

See process.

pub fn host_process<T: AsRef<OsStr>>(
&self,
cmd: T,
pkg: &Package
) -> CargoResult<ProcessBuilder>

See process.

pub fn target_process<T: AsRef<OsStr>>(
&self,
cmd: T,
pkg: &Package
) -> CargoResult<ProcessBuilder>

See process.