Struct cargo::ops::CompileOptions

pub struct CompileOptions<'a> {
    pub config: &'a Config,
    pub jobs: Option<u32>,
    pub target: Option<&'a str>,
    pub features: &'a [String],
    pub all_features: bool,
    pub no_default_features: bool,
    pub spec: Packages<'a>,
    pub filter: CompileFilter<'a>,
    pub release: bool,
    pub mode: CompileMode,
    pub message_format: MessageFormat,
    pub target_rustdoc_args: Option<&'a [String]>,
    pub target_rustc_args: Option<&'a [String]>,
}

Contains information about how a package should be compiled.

Fields
config: &'a Config jobs: Option<u32>

Number of concurrent jobs to use.

target: Option<&'a str>

The target platform to compile for (example: i686-unknown-linux-gnu).

features: &'a [String]

Extra features to build for the root package

all_features: bool

Flag whether all available features should be built for the root package

no_default_features: bool

Flag if the default feature should be built for the root package

spec: Packages<'a>

A set of packages to build.

filter: CompileFilter<'a>

Filter to apply to the root package to select which targets will be built.

release: bool

Whether this is a release build or not

mode: CompileMode

Mode for this compile.

message_format: MessageFormat

--error_format flag for the compiler.

target_rustdoc_args: Option<&'a [String]>

Extra arguments to be passed to rustdoc (for main crate and dependencies)

target_rustc_args: Option<&'a [String]>

The specified target will be compiled with all the available arguments, note that this only accounts for the final invocation of rustc

Methods

impl<'a> CompileOptions<'a>

pub fn default(config: &'a Config, mode: CompileMode) -> CompileOptions<'a>