Struct cargo::ops::CompileOptions

pub struct CompileOptions<'a> {
    pub config: &'a Config,
    pub jobs: Option<u32>,
    pub target: Option<&'a str>,
    pub features: &'a [String],
    pub no_default_features: bool,
    pub spec: &'a [String],
    pub filter: CompileFilter<'a>,
    pub exec_engine: Option<Arc<Box<ExecEngine>>>,
    pub release: bool,
    pub mode: CompileMode,
    pub target_rustc_args: Option<&'a [String]>,
}

Contains information about how a package should be compiled.

Fields

config

jobs

Number of concurrent jobs to use.

target

The target platform to compile for (example: i686-unknown-linux-gnu).

features

Extra features to build for the root package

no_default_features

Flag if the default feature should be built for the root package

spec

Root package to build (if None it's the current one)

filter

Filter to apply to the root package to select which targets will be built.

exec_engine

Engine which drives compilation

release

Whether this is a release build or not

mode

Mode for this compile.

target_rustc_args

The specified target will be compiled with all the available arguments, note that this only accounts for the final invocation of rustc