Struct cargo::ops::Context

pub struct Context<'a, 'cfg: 'a> {
    pub ws: &'a Workspace<'cfg>,
    pub config: &'cfg Config,
    pub resolve: &'a Resolve,
    pub compilation: Compilation<'cfg>,
    pub packages: &'a PackageSet<'cfg>,
    pub build_state: Arc<BuildState>,
    pub build_script_overridden: HashSet<(PackageId, Kind)>,
    pub build_explicit_deps: HashMap<Unit<'a>, BuildDeps>,
    pub fingerprints: HashMap<Unit<'a>, Arc<Fingerprint>>,
    pub compiled: HashSet<Unit<'a>>,
    pub build_config: BuildConfig,
    pub build_scripts: HashMap<Unit<'a>, Arc<BuildScripts>>,
    pub links: Links<'a>,
    pub used_in_plugin: HashSet<Unit<'a>>,
    pub jobserver: Client,
    // some fields omitted
}

The build context, containing all information about a build task

Fields
ws: &'a Workspace<'cfg>

The workspace the build is for

config: &'cfg Config

The cargo configuration

resolve: &'a Resolve

The dependency graph for our build

compilation: Compilation<'cfg>

Information on the compilation output

packages: &'a PackageSet<'cfg> build_state: Arc<BuildState> build_script_overridden: HashSet<(PackageId, Kind)> build_explicit_deps: HashMap<Unit<'a>, BuildDeps> fingerprints: HashMap<Unit<'a>, Arc<Fingerprint>> compiled: HashSet<Unit<'a>> build_config: BuildConfig build_scripts: HashMap<Unit<'a>, Arc<BuildScripts>> links: Links<'a> used_in_plugin: HashSet<Unit<'a>> jobserver: Client
Methods

impl<'a, 'cfg> Context<'a, 'cfg>

pub fn new(
ws: &'a Workspace<'cfg>,
resolve: &'a Resolve,
packages: &'a PackageSet<'cfg>,
config: &'cfg Config,
build_config: BuildConfig,
profiles: &'a Profiles
) -> CargoResult<Context<'a, 'cfg>>

pub fn prepare(&mut self) -> CargoResult<()>

Prepare this context, ensuring that all filesystem directories are in place.

pub fn probe_target_info(&mut self, units: &[Unit<'a>]) -> CargoResult<()>

Ensure that we've collected all target-specific information to compile all the units mentioned in units.

pub fn build_used_in_plugin_map(
&mut self,
units: &[Unit<'a>]
) -> CargoResult<()>

Builds up the used_in_plugin internal to this context from the list of top-level units.

This will recursively walk units and all of their dependencies to determine which crate are going to be used in plugins or not.

pub fn deps_dir(&self, unit: &Unit) -> &Path

Returns the directories where Rust crate dependencies are found for the specified unit.

pub fn fingerprint_dir(&mut self, unit: &Unit<'a>) -> PathBuf

Returns the directory for the specified unit where fingerprint information is stored.

pub fn build_script_dir(&mut self, unit: &Unit<'a>) -> PathBuf

Returns the appropriate directory layout for either a plugin or not.

pub fn build_script_out_dir(&mut self, unit: &Unit<'a>) -> PathBuf

Returns the appropriate directory layout for either a plugin or not.

pub fn host_deps(&self) -> &Path

pub fn target_root(&self) -> &Path

Return the root of the build output tree

pub fn out_dir(&mut self, unit: &Unit<'a>) -> PathBuf

Returns the appropriate output directory for the specified package and target.

pub fn host_triple(&self) -> &str

Return the host triple for this context

pub fn target_triple(&self) -> &str

Return the target triple which this context is targeting.

pub fn requested_target(&self) -> Option<&str>

Requested (not actual) target for the build

pub fn target_short_hash(&self, unit: &Unit) -> String

Get the short hash based only on the PackageId Used for the metadata when target_metadata returns None

pub fn target_metadata(&mut self, unit: &Unit<'a>) -> Option<Metadata>

Get the metadata for a target in a specific profile We build to the path: "{filename}-{target_metadata}" We use a linking step to link/copy to a predictable filename like target/debug/libfoo.{a,so,rlib} and such.

pub fn file_stem(&mut self, unit: &Unit<'a>) -> String

Returns the file stem for a given target/profile combo (with metadata)

pub fn link_stem(&mut self, unit: &Unit<'a>) -> Option<(PathBuf, String)>

Returns a tuple with the directory and name of the hard link we expect our target to be copied to. Eg, file_stem may be out_dir/deps/foo-abcdef and link_stem would be out_dir/foo This function returns it in two parts so the caller can add prefix/suffix to filename separately

Returns an Option because in some cases we don't want to link (eg a dependent lib)

pub fn target_filenames(
&mut self,
unit: &Unit<'a>
) -> CargoResult<Arc<Vec<(PathBuf, Option<PathBuf>, TargetFileType)>>>

Return the filenames that the given target for the given profile will generate as a list of 3-tuples (filename, link_dst, linkable)

filename: filename rustc compiles to. (Often has metadata suffix).

link_dst: Optional file to link/copy the result to (without metadata suffix)

linkable: Whether possible to link against file (eg it's a library)

pub fn dep_targets(&self, unit: &Unit<'a>) -> CargoResult<Vec<Unit<'a>>>

For a package, return all targets which are registered as dependencies for that package.

pub fn dep_run_custom_build(
&self,
unit: &Unit<'a>
) -> CargoResult<Vec<Unit<'a>>>

Returns the dependencies needed to run a build script.

The unit provided must represent an execution of a build script, and the returned set of units must all be run before unit is run.

pub fn get_package(&self, id: &PackageId) -> CargoResult<&'a Package>

Gets a package for the given package id.

pub fn linker(&self, kind: Kind) -> Option<&Path>

Get the user-specified linker for a particular host or target

pub fn ar(&self, kind: Kind) -> Option<&Path>

Get the user-specified ar program for a particular host or target

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn cfg(&self, kind: Kind) -> &[Cfg]

Get the list of cfg printed out from the compiler for the specified kind

pub fn jobs(&self) -> u32

Number of jobs specified for this build

pub fn lib_profile(&self) -> &'a Profile

pub fn lib_or_check_profile(&self, unit: &Unit, target: &Target) -> &'a Profile

pub fn build_script_profile(&self, _pkg: &PackageId) -> &'a Profile

pub fn incremental_args(&self, unit: &Unit) -> CargoResult<Vec<String>>

pub fn rustflags_args(&self, unit: &Unit) -> CargoResult<Vec<String>>

pub fn rustdocflags_args(&self, unit: &Unit) -> CargoResult<Vec<String>>

pub fn show_warnings(&self, pkg: &PackageId) -> bool