cargo::ops::tree

Module graph

Code for building the graph used by cargo tree.

Structs

Edge

Edges 🔒

Set of outgoing edges for a single node.

Graph

A graph of dependencies.

NodeId

Enums

EdgeKind

The kind of edge, for separating dependencies into different sections.

Node

Functions

add_cli_features 🔒

Adds nodes for features requested on the command-line for the given member.

add_feature 🔒

Adds a feature node between two nodes.

add_feature_rec 🔒

Recursively add feature nodes for all features enabled by the given feature.

add_internal_features 🔒

Recursively adds connections between features in the [features] table for every package.

add_pkg 🔒

Adds a single package node (if it does not already exist).

build

Builds the graph.