Module cargo::sources::git

Home of the GitSource.

Apparently, the most important type in this module is GitSource. utils provides libgit2 utilities like fetch and checkout, whereas oxide is the couterpart for gitoxide integration. known_hosts is the mitigation of CVE-2022-46176.

Modules

fetch

For -Zgitoxide integration.

known_hosts 🔒

SSH host key validation support.

oxide 🔒

This module contains all code sporting gitoxide for operations on git repositories and it mirrors utils closely for now. One day it can be renamed into utils once git2 isn’t required anymore.

source 🔒

See GitSource.

utils 🔒

Utilities for handling git repositories, mainly around authentication/cloning.

Structs

GitCheckout

A local checkout of a particular revision from a GitDatabase.

GitDatabase

A local clone of a remote repository’s database. Multiple GitCheckouts can be cloned from a single GitDatabase.

GitRemote

A remote repository. It gets cloned into a local GitDatabase.

GitSource

GitSource contains one or more packages gathering from a Git repository. Under the hood it uses PathSource to discover packages inside the repository.

Functions

fetch

Attempts to fetch the given git reference for a Git repository.