cargo::sources

Module path

Structs

PathEntry

PathBuf with extra metadata.

PathSource

A source that represents a package gathered at the root path on the filesystem.

RecursivePathSource

A source that represents one or multiple packages gathered from a given root path on the filesystem.

Enums

FileType 🔒

Type that abstracts over [gix::dir::entry::Kind] and fs::FileType.

Functions

_list_files 🔒

See PathSource::list_files.

discover_gix_repo 🔒

Returns Some(gix::Repository) if the discovered repository (searched upwards from root) contains a tracked <root>/Cargo.toml. Otherwise, the caller should fall back on full file list.

first_package 🔒

has_manifest 🔒

last_modified_file 🔒

Gets the last modified file in a package.

list_files

List all files relevant to building this package inside this source.

list_files_gix 🔒

Lists files relevant to building this package inside this source by traversing the git working tree, while avoiding ignored files.

list_files_walk 🔒

Lists files relevant to building this package inside this source by walking the filesystem from the package root path.

nested_paths 🔒

read_nested_packages 🔒

read_packages 🔒

walk 🔒