Struct cargo::sources::path::PathSource

pub struct PathSource<'cfg> { /* fields omitted */ }

Methods

impl<'cfg> PathSource<'cfg>

pub fn new(path: &Path, id: &SourceId, config: &'cfg Config) -> PathSource<'cfg>

Invoked with an absolute path to a directory that contains a Cargo.toml.

This source will only return the package at precisely the path specified, and it will be an error if there's not a package at path.

pub fn new_recursive(
root: &Path,
id: &SourceId,
config: &'cfg Config
) -> PathSource<'cfg>

Creates a new source which is walked recursively to discover packages.

This is similar to the new method except that instead of requiring a valid package to be present at root the folder is walked entirely to crawl for packages.

Note that this should be used with care and likely shouldn't be chosen by default!

pub fn root_package(&mut self) -> CargoResult<Package>

pub fn read_packages(&self) -> CargoResult<Vec<Package>>

pub fn list_files(&self, pkg: &Package) -> CargoResult<Vec<PathBuf>>

List all files relevant to building this package inside this source.

This function will use the appropriate methods to determine the set of files underneath this source's directory which are relevant for building pkg.

The basic assumption of this method is that all files in the directory are relevant for building this package, but it also contains logic to use other methods like .gitignore to filter the list of files.

Pattern matching strategy

Migrating from a glob-like pattern matching (using glob crate) to a gitignore-like pattern matching (using ignore crate). The migration stages are:

1) Only warn users about the future change iff their matching rules are affected. (CURRENT STAGE)

2) Switch to the new strategy and upate documents. Still keep warning affected users.

3) Drop the old strategy and no mor warnings.

See https://github.com/rust-lang/cargo/issues/4268 for more info.