Struct cargo::sources::path::PathSource

pub struct PathSource<'cfg> {
    // some fields omitted
}

Methods

impl<'cfg> PathSource<'cfg>

fn for_path(path: &Path, config: &'cfg Config) -> CargoResult<PathSource<'cfg>>

fn new(path: &Path, id: &SourceId, config: &'cfg Config) -> PathSource<'cfg>

Invoked with an absolute path to a directory that contains a Cargo.toml. The source will read the manifest and find any other packages contained in the directory structure reachable by the root manifest.

fn root_package(&mut self) -> CargoResult<Package>

fn list_files(&self, pkg: &Package) -> CargoResult<Vec<PathBuf>>

List all files relevant to building this package inside this source.

This function will use the appropriate methods to determine the set of files underneath this source's directory which are relevant for building pkg.

The basic assumption of this method is that all files in the directory are relevant for building this package, but it also contains logic to use other methods like .gitignore to filter the list of files.