Struct cargo::sources::registry::RegistryConfig

pub struct RegistryConfig {
    pub dl: String,
    pub api: String,
}

Fields
dl: String

Download endpoint for all crates. This will be appended with /<crate>/<version>/download and then will be hit with an HTTP GET request to download the tarball for a crate.

api: String

API endpoint for the registry. This is what's actually hit to perform operations like yanks, owner modifications, publish new crates, etc.