Struct cargo::sources::registry::RegistryConfig

pub struct RegistryConfig {
    pub dl: String,
    pub api: Option<String>,
}

Fields
dl: String

Download endpoint for all crates.

The string is a template which will generate the download URL for the tarball of a specific version of a crate. The substrings {crate} and {version} will be replaced with the crate's name and version respectively.

For backwards compatibility, if the string does not contain {crate} or {version}, it will be extended with /{crate}/{version}/download to support registries like crates.io which were crated before the templating setup was created.

api: Option<String>

API endpoint for the registry. This is what's actually hit to perform operations like yanks, owner modifications, publish new crates, etc.