Struct cargo::sources::registry::RegistrySource

pub struct RegistrySource<'cfg> { /* private fields */ }

A “source” for a local or remote registry.

This contains common functionality that is shared between the two registry kinds, with the registry-specific logic implemented as part of the RegistryData trait referenced via the ops field.

Implementations

impl<'cfg> RegistrySource<'cfg>

pub fn remote(
source_id: SourceId,
yanked_whitelist: &HashSet<PackageId>,
config: &'cfg Config
) -> RegistrySource<'cfg>

pub fn local(
source_id: SourceId,
path: &Path,
yanked_whitelist: &HashSet<PackageId>,
config: &'cfg Config
) -> RegistrySource<'cfg>

pub fn config(&mut self) -> CargoResult<Option<RegistryConfig>>

Decode the configuration stored within the registry.

This requires that the index has been at least checked out.