cargo::sources::replaced
Struct ReplacedSource

pub struct ReplacedSource<'gctx> {
    to_replace: SourceId,
    replace_with: SourceId,
    inner: Box<dyn Source + 'gctx>,
}

A source that replaces one source with the other. This manages the source replacement feature.

The implementation is merely redirecting from the original to the replacement.

Fields
to_replace: SourceId

The identifier of the original source.

replace_with: SourceId

The identifier of the new replacement source.

inner: Box<dyn Source + 'gctx>
Implementations

impl<'gctx> ReplacedSource<'gctx>

pub fn new( to_replace: SourceId, replace_with: SourceId, src: Box<dyn Source + 'gctx>, ) -> ReplacedSource<'gctx>

Creates a replaced source.

The src argument is the new replacement source.

fn is_builtin_replacement(&self) -> bool

Is this source a built-in replacement of crates.io?

Built-in source replacement of crates.io for sparse registry or tests should not show messages indicating source replacement.