cargo::util

Module auth

Registry authentication support.

Structs

AuthorizationError

An authorization error from accessing a registry.

RegistryConfig

[registries.NAME] tables.

RegistryConfigExtended

The [registry] table, which more keys than the [registries.NAME] tables.

Enums

AuthorizationErrorReason

Statics

BUILT_IN_PROVIDERS 🔒

List of credential providers built-in to Cargo. Keep in sync with the match in credential_action.

Functions

auth_token

Returns the token to use for the given registry. If a login_url is provided and a token is not available, the login_url will be included in the returned error.

auth_token_optional 🔒

Returns the token to use for the given registry.

cache_token_from_commandline

Store a token in the cache for future calls.

credential_action 🔒

credential_provider 🔒

Get the list of credential providers for a registry source.

get_credential_libsecret 🔒

Retrieves a cached instance of LibSecretCredential. Must be cached to avoid repeated load/unload cycles, which are not supported by glib.

login

Log in to the given registry.

logout

Log out from the given registry.

registry_credential_config_raw

Get the credential configuration for a SourceId.

registry_credential_config_raw_uncached 🔒

resolve_credential_alias 🔒

Use the [credential-alias] table to see if the provider name has been aliased.