Module cargo::util::auth

Registry authentication support.

Structs

AuthorizationError

An authorization error from accessing a registry.

Enums

Action 🔒

AuthorizationErrorReason

Functions

auth_token

Returns the token to use for the given registry. If a login_url is provided and a token is not available, the login_url will be included in the returned error.

auth_token_optional 🔒

Returns the token to use for the given registry.

cache_token

login

Saves the given token.

logout

Removes the token for the given registry.

registry_credential_config

Get the credential configuration for a SourceId.

run_command 🔒

sysroot_credential 🔒

Gets the path to the libexec processes in the sysroot.