Module cargo::util::command_prelude

Re-exports

pub use crate::core::compiler::CompileMode;

pub use crate::CliError;

pub use crate::CliResult;

pub use crate::Config;

Modules

heading

Macros

value_parser

Select a [ValueParser] implementation from the intended type

Structs

Arg

The abstract representation of a command line argument. Used to set all the options and relationships that define a valid argument for the program.

ArgMatches

Container for parse results.

Command

Build a command-line interface.

Enums

ArgAction

Behavior of arguments when they are encountered while parsing

CommandInfo

ProfileChecking

Determines whether or not to gate --profile as unstable when resolving it.

Traits

ArgMatchesExt

CommandExt

Functions

flag

ignore_unknown

multi_opt

opt

optional_multi_opt

optional_opt

root_manifest

subcommand

values

values_os