cargo::util

Module command_prelude

Re-exports

pub use crate::core::compiler::UserIntent;

pub use crate::CliError;

pub use crate::CliResult;

pub use crate::GlobalContext;

Modules

heading

Macros

value_parser

Select a [ValueParser] implementation from the intended type

Structs

Arg

The abstract representation of a command line argument. Used to set all the options and relationships that define a valid argument for the program.

ArgMatches

Container for parse results.

Command

Build a command-line interface.

Enums

ArgAction

Behavior of arguments when they are encountered while parsing

CommandInfo

ProfileChecking

Determines whether or not to gate --profile as unstable when resolving it.

Traits

ArgMatchesExt

CommandExt

Functions

default_profile_candidates 🔒

flag

get_bench_candidates 🔒

get_bin_candidates 🔒

get_dependencies_from_metadata 🔒

get_direct_dependencies_pkg_name_candidates

get_example_candidates 🔒

get_packages 🔒

get_pkg_id_spec_candidates

get_profile_candidates 🔒

get_registry_candidates

get_target_triples 🔒

get_target_triples_from_rustc 🔒

get_target_triples_from_rustup 🔒

get_targets_from_metadata 🔒

get_test_candidates 🔒

get_workspace_profile_candidates 🔒

ignore_unknown

lockfile_path

multi_opt

new_gctx_for_completions

opt

optional_multi_opt

optional_opt

root_manifest

subcommand

values

values_os