Trait cargo::util::command_prelude::CommandExt

pub trait CommandExt: Sized {

    // Required method
    fn _arg(self, arg: Arg) -> Self;

    // Provided methods
    fn arg_package_spec(
        self,
        package: &'static str,
        all: &'static str,
        exclude: &'static str
    ) -> Self { ... }
    fn arg_package_spec_no_all(
        self,
        package: &'static str,
        all: &'static str,
        exclude: &'static str
    ) -> Self { ... }
    fn arg_package_spec_simple(self, package: &'static str) -> Self { ... }
    fn arg_package(self, package: &'static str) -> Self { ... }
    fn arg_parallel(self) -> Self { ... }
    fn arg_jobs(self) -> Self { ... }
    fn arg_unsupported_keep_going(self) -> Self { ... }
    fn arg_targets_all(
        self,
        lib: &'static str,
        bin: &'static str,
        bins: &'static str,
        example: &'static str,
        examples: &'static str,
        test: &'static str,
        tests: &'static str,
        bench: &'static str,
        benches: &'static str,
        all: &'static str
    ) -> Self { ... }
    fn arg_targets_lib_bin_example(
        self,
        lib: &'static str,
        bin: &'static str,
        bins: &'static str,
        example: &'static str,
        examples: &'static str
    ) -> Self { ... }
    fn arg_targets_bins_examples(
        self,
        bin: &'static str,
        bins: &'static str,
        example: &'static str,
        examples: &'static str
    ) -> Self { ... }
    fn arg_targets_bin_example(
        self,
        bin: &'static str,
        example: &'static str
    ) -> Self { ... }
    fn arg_features(self) -> Self { ... }
    fn arg_release(self, release: &'static str) -> Self { ... }
    fn arg_profile(self, profile: &'static str) -> Self { ... }
    fn arg_doc(self, doc: &'static str) -> Self { ... }
    fn arg_target_triple(self, target: &'static str) -> Self { ... }
    fn arg_target_dir(self) -> Self { ... }
    fn arg_manifest_path(self) -> Self { ... }
    fn arg_message_format(self) -> Self { ... }
    fn arg_build_plan(self) -> Self { ... }
    fn arg_unit_graph(self) -> Self { ... }
    fn arg_new_opts(self) -> Self { ... }
    fn arg_index(self) -> Self { ... }
    fn arg_dry_run(self, dry_run: &'static str) -> Self { ... }
    fn arg_ignore_rust_version(self) -> Self { ... }
    fn arg_future_incompat_report(self) -> Self { ... }
    fn arg_quiet(self) -> Self { ... }
    fn arg_timings(self) -> Self { ... }

}

Required Methods

fn _arg(self, arg: Arg) -> Self

Provided Methods

fn arg_package_spec( self, package: &'static str, all: &'static str, exclude: &'static str ) -> Self

Do not use this method, it is only for backwards compatibility. Use arg_package_spec_no_all instead.

fn arg_package_spec_no_all( self, package: &'static str, all: &'static str, exclude: &'static str ) -> Self

Variant of arg_package_spec that does not include the --all flag (but does include --workspace). Used to avoid confusion with historical uses of --all.

fn arg_package_spec_simple(self, package: &'static str) -> Self

fn arg_package(self, package: &'static str) -> Self

fn arg_parallel(self) -> Self

fn arg_jobs(self) -> Self

fn arg_unsupported_keep_going(self) -> Self

fn arg_targets_all( self, lib: &'static str, bin: &'static str, bins: &'static str, example: &'static str, examples: &'static str, test: &'static str, tests: &'static str, bench: &'static str, benches: &'static str, all: &'static str ) -> Self

fn arg_targets_lib_bin_example( self, lib: &'static str, bin: &'static str, bins: &'static str, example: &'static str, examples: &'static str ) -> Self

fn arg_targets_bins_examples( self, bin: &'static str, bins: &'static str, example: &'static str, examples: &'static str ) -> Self

fn arg_targets_bin_example( self, bin: &'static str, example: &'static str ) -> Self

fn arg_features(self) -> Self

fn arg_release(self, release: &'static str) -> Self

fn arg_profile(self, profile: &'static str) -> Self

fn arg_doc(self, doc: &'static str) -> Self

fn arg_target_triple(self, target: &'static str) -> Self

fn arg_target_dir(self) -> Self

fn arg_manifest_path(self) -> Self

fn arg_message_format(self) -> Self

fn arg_build_plan(self) -> Self

fn arg_unit_graph(self) -> Self

fn arg_new_opts(self) -> Self

fn arg_index(self) -> Self

fn arg_dry_run(self, dry_run: &'static str) -> Self

fn arg_ignore_rust_version(self) -> Self

fn arg_future_incompat_report(self) -> Self

fn arg_quiet(self) -> Self

fn arg_timings(self) -> Self