Trait cargo::util::command_prelude::CommandExt

pub trait CommandExt: Sized {

    // Required method
    fn _arg(self, arg: Arg) -> Self;

    // Provided methods
    fn arg_package_spec(
        self,
        package: &'static str,
        all: &'static str,
        exclude: &'static str,
    ) -> Self { ... }
    fn arg_package_spec_no_all(
        self,
        package: &'static str,
        all: &'static str,
        exclude: &'static str,
    ) -> Self { ... }
    fn arg_package_spec_simple(self, package: &'static str) -> Self { ... }
    fn arg_package(self, package: &'static str) -> Self { ... }
    fn arg_parallel(self) -> Self { ... }
    fn arg_jobs(self) -> Self { ... }
    fn arg_unsupported_keep_going(self) -> Self { ... }
    fn arg_redundant_default_mode(
        self,
        default_mode: &'static str,
        command: &'static str,
        supported_mode: &'static str,
    ) -> Self { ... }
    fn arg_targets_all(
        self,
        lib: &'static str,
        bin: &'static str,
        bins: &'static str,
        example: &'static str,
        examples: &'static str,
        test: &'static str,
        tests: &'static str,
        bench: &'static str,
        benches: &'static str,
        all: &'static str,
    ) -> Self { ... }
    fn arg_targets_lib_bin_example(
        self,
        lib: &'static str,
        bin: &'static str,
        bins: &'static str,
        example: &'static str,
        examples: &'static str,
    ) -> Self { ... }
    fn arg_targets_bins_examples(
        self,
        bin: &'static str,
        bins: &'static str,
        example: &'static str,
        examples: &'static str,
    ) -> Self { ... }
    fn arg_targets_bin_example(
        self,
        bin: &'static str,
        example: &'static str,
    ) -> Self { ... }
    fn arg_features(self) -> Self { ... }
    fn arg_release(self, release: &'static str) -> Self { ... }
    fn arg_profile(self, profile: &'static str) -> Self { ... }
    fn arg_doc(self, doc: &'static str) -> Self { ... }
    fn arg_target_triple(self, target: &'static str) -> Self { ... }
    fn arg_target_dir(self) -> Self { ... }
    fn arg_manifest_path(self) -> Self { ... }
    fn arg_manifest_path_without_unsupported_path_tip(self) -> Self { ... }
    fn arg_message_format(self) -> Self { ... }
    fn arg_build_plan(self) -> Self { ... }
    fn arg_unit_graph(self) -> Self { ... }
    fn arg_new_opts(self) -> Self { ... }
    fn arg_registry(self, help: &'static str) -> Self { ... }
    fn arg_index(self, help: &'static str) -> Self { ... }
    fn arg_dry_run(self, dry_run: &'static str) -> Self { ... }
    fn arg_ignore_rust_version(self) -> Self { ... }
    fn arg_ignore_rust_version_with_help(self, help: &'static str) -> Self { ... }
    fn arg_future_incompat_report(self) -> Self { ... }
    fn arg_silent_suggestion(self) -> Self { ... }
    fn arg_timings(self) -> Self { ... }
    fn arg_artifact_dir(self) -> Self { ... }

}

Required Methods

fn _arg(self, arg: Arg) -> Self

Provided Methods

fn arg_package_spec( self, package: &'static str, all: &'static str, exclude: &'static str, ) -> Self

Do not use this method, it is only for backwards compatibility. Use arg_package_spec_no_all instead.

fn arg_package_spec_no_all( self, package: &'static str, all: &'static str, exclude: &'static str, ) -> Self

Variant of arg_package_spec that does not include the --all flag (but does include --workspace). Used to avoid confusion with historical uses of --all.

fn arg_package_spec_simple(self, package: &'static str) -> Self

fn arg_package(self, package: &'static str) -> Self

fn arg_parallel(self) -> Self

fn arg_jobs(self) -> Self

fn arg_unsupported_keep_going(self) -> Self

fn arg_redundant_default_mode( self, default_mode: &'static str, command: &'static str, supported_mode: &'static str, ) -> Self

fn arg_targets_all( self, lib: &'static str, bin: &'static str, bins: &'static str, example: &'static str, examples: &'static str, test: &'static str, tests: &'static str, bench: &'static str, benches: &'static str, all: &'static str, ) -> Self

fn arg_targets_lib_bin_example( self, lib: &'static str, bin: &'static str, bins: &'static str, example: &'static str, examples: &'static str, ) -> Self

fn arg_targets_bins_examples( self, bin: &'static str, bins: &'static str, example: &'static str, examples: &'static str, ) -> Self

fn arg_targets_bin_example( self, bin: &'static str, example: &'static str, ) -> Self

fn arg_features(self) -> Self

fn arg_release(self, release: &'static str) -> Self

fn arg_profile(self, profile: &'static str) -> Self

fn arg_doc(self, doc: &'static str) -> Self

fn arg_target_triple(self, target: &'static str) -> Self

fn arg_target_dir(self) -> Self

fn arg_manifest_path(self) -> Self

fn arg_manifest_path_without_unsupported_path_tip(self) -> Self

fn arg_message_format(self) -> Self

fn arg_build_plan(self) -> Self

fn arg_unit_graph(self) -> Self

fn arg_new_opts(self) -> Self

fn arg_registry(self, help: &'static str) -> Self

fn arg_index(self, help: &'static str) -> Self

fn arg_dry_run(self, dry_run: &'static str) -> Self

fn arg_ignore_rust_version(self) -> Self

fn arg_ignore_rust_version_with_help(self, help: &'static str) -> Self

fn arg_future_incompat_report(self) -> Self

fn arg_silent_suggestion(self) -> Self

Adds a suggestion for the --silent or -s flags to use the --quiet flag instead. This is to help with people familiar with other tools that use -s.

Every command should call this, unless it has its own -s short flag.

fn arg_timings(self) -> Self

fn arg_artifact_dir(self) -> Self

Object Safety

This trait is not object safe.