Module cargo::util::config

Structs

CargoBuildConfig

CargoHttpConfig

CargoNetConfig

Config
Configuration information for cargo. This is not specific to a build, it is information relating to cargo itself.

ConfigError
Internal error for serde errors.

ConfigRelativePath
Use with the get API to fetch a string that will be converted to a PathBuf. Relative paths are converted to absolute paths based on the location of the config file.

PackageCacheLock

SslVersionConfigRange

StringList
A type to deserialize a list of strings from a toml file.

Value
A type which can be deserialized as a configuration value which records where it was deserialized from.

Enums

ConfigValue

Definition

SslVersionConfig
Configuration for ssl-version in http section There are two ways to configure:

Functions

clippy_driver
returns path to clippy-driver binary

homedir

save_credentials

Type Definitions

OptValue