Struct cargo::util::config::Config

pub struct Config { /* fields omitted */ }

Configuration information for cargo. This is not specific to a build, it is information relating to cargo itself.

This struct implements Default: all fields can be inferred.

Implementations

impl Config

pub fn new(shell: Shell, cwd: PathBuf, homedir: PathBuf) -> Config

pub fn default() -> CargoResult<Config>

pub fn home(&self) -> &Filesystem

Gets the user's Cargo home directory (OS-dependent).

pub fn git_path(&self) -> Filesystem

Gets the Cargo Git directory (<cargo_home>/git).

pub fn registry_index_path(&self) -> Filesystem

Gets the Cargo registry index directory (<cargo_home>/registry/index).

pub fn registry_cache_path(&self) -> Filesystem

Gets the Cargo registry cache directory (<cargo_home>/registry/path).

pub fn registry_source_path(&self) -> Filesystem

Gets the Cargo registry source directory (<cargo_home>/registry/src).

pub fn default_registry(&self) -> CargoResult<Option<String>>

Gets the default Cargo registry.

pub fn shell(&self) -> RefMut<'_, Shell>

Gets a reference to the shell, e.g., for writing error messages.

pub fn rustdoc(&self) -> CargoResult<&Path>

Gets the path to the rustdoc executable.

pub fn load_global_rustc(
&self,
ws: Option<&Workspace<'_>>
) -> CargoResult<Rustc>

Gets the path to the rustc executable.

pub fn cargo_exe(&self) -> CargoResult<&Path>

Gets the path to the cargo executable.

pub fn profiles(&self) -> CargoResult<&ConfigProfiles>

pub fn updated_sources(&self) -> RefMut<'_, HashSet<SourceId>>

pub fn values(&self) -> CargoResult<&HashMap<String, ConfigValue>>

pub fn values_mut(&mut self) -> CargoResult<&mut HashMap<String, ConfigValue>>

pub fn set_values(
&self,
values: HashMap<String, ConfigValue>
) -> CargoResult<()>

pub fn reload_rooted_at<P: AsRef<Path>>(&mut self, path: P) -> CargoResult<()>

pub fn cwd(&self) -> &Path

pub fn target_dir(&self) -> CargoResult<Option<Filesystem>>

pub fn set_env(&mut self, env: HashMap<String, String>)

pub fn get_string(&self, key: &str) -> CargoResult<OptValue<String>>

pub fn get_path(&self, key: &str) -> CargoResult<OptValue<PathBuf>>

pub fn get_path_and_args(
&self,
key: &str
) -> CargoResult<OptValue<(PathBuf, Vec<String>)>>

pub fn get_list(
&self,
key: &str
) -> CargoResult<OptValue<Vec<(String, PathBuf)>>>

pub fn get_table(&self, key: &str) -> CargoResult<OptValue<HashMap<String, CV>>>

pub fn configure(
&mut self,
verbose: u32,
quiet: Option<bool>,
color: &Option<String>,
frozen: bool,
locked: bool,
offline: bool,
target_dir: &Option<PathBuf>,
unstable_flags: &[String]
) -> CargoResult<()>

pub fn cli_unstable(&self) -> &CliUnstable

pub fn extra_verbose(&self) -> bool

pub fn network_allowed(&self) -> bool

pub fn offline(&self) -> bool

pub fn frozen(&self) -> bool

pub fn lock_update_allowed(&self) -> bool

pub fn load_values(&self) -> CargoResult<HashMap<String, ConfigValue>>

Loads configuration from the filesystem.

pub fn get_registry_index(&self, registry: &str) -> CargoResult<Url>

Gets the index for a registry.

pub fn get_default_registry_index(&self) -> CargoResult<Option<Url>>

Gets the index for the default registry.

pub fn get_tool(&self, tool: &str) -> CargoResult<PathBuf>

Looks for a path for tool in an environment variable or config path, defaulting to tool as a path.

pub fn jobserver_from_env(&self) -> Option<&Client>

pub fn http(&self) -> CargoResult<&RefCell<Easy>>

pub fn http_config(&self) -> CargoResult<&CargoHttpConfig>

pub fn net_config(&self) -> CargoResult<&CargoNetConfig>

pub fn build_config(&self) -> CargoResult<&CargoBuildConfig>

pub fn crates_io_source_id<F>(&self, f: F) -> CargoResult<SourceId> where
F: FnMut() -> CargoResult<SourceId>,

pub fn creation_time(&self) -> Instant

pub fn get<'de, T: Deserialize<'de>>(&self, key: &str) -> CargoResult<T>

pub fn assert_package_cache_locked<'a>(&self, f: &'a Filesystem) -> &'a Path

pub fn acquire_package_cache_lock(&self) -> CargoResult<PackageCacheLock<'_>>

Acquires an exclusive lock on the global "package cache"

This lock is global per-process and can be acquired recursively. An RAII structure is returned to release the lock, and if this process abnormally terminates the lock is also released.

pub fn release_package_cache_lock(&self)