Struct cargo::util::config::Config

pub struct Config {
    // some fields omitted
}

Methods

impl Config

fn new(shell: MultiShell, cwd: PathBuf, homedir: PathBuf) -> Config

fn default() -> CargoResult<Config>

fn home(&self) -> &Filesystem

fn git_path(&self) -> Filesystem

fn lock_git(&self) -> CargoResult<&FileLock>

All git sources are protected by a single file lock, which is stored in the Config struct. Ideally, we should use a lock per repository, but this leads to deadlocks when several instances of Cargo acquire locks in different order (see #2987).

Holding the lock only during checkout is not enough. For example, two instances of Cargo may checkout different commits from the same branch to the same directory. So the lock is acquired when the first git source is updated and released when the Config struct is destroyed.

fn registry_index_path(&self) -> Filesystem

fn registry_cache_path(&self) -> Filesystem

fn registry_source_path(&self) -> Filesystem

fn shell(&self) -> RefMut<MultiShell>

fn rustdoc(&self) -> CargoResult<&Path>

fn rustc(&self) -> CargoResult<&Rustc>

fn values(&self) -> CargoResult<&HashMap<String, ConfigValue>>

fn cwd(&self) -> &Path

fn target_dir(&self) -> CargoResult<Option<Filesystem>>

fn get_string(&self, key: &str) -> CargoResult<Option<Value<String>>>

fn get_bool(&self, key: &str) -> CargoResult<Option<Value<bool>>>

fn get_path(&self, key: &str) -> CargoResult<Option<Value<PathBuf>>>

fn get_list(&self, key: &str) -> CargoResult<Option<Value<Vec<(String, PathBuf)>>>>

fn get_table(&self, key: &str) -> CargoResult<Option<Value<HashMap<String, CV>>>>

fn get_i64(&self, key: &str) -> CargoResult<Option<Value<i64>>>

fn net_retry(&self) -> CargoResult<i64>

fn expected<T>(&self, ty: &str, key: &str, val: CV) -> CargoResult<T>

fn configure(&self, verbose: u32, quiet: Option<bool>, color: &Option<String>, frozen: bool, locked: bool) -> CargoResult<()>

fn extra_verbose(&self) -> bool

fn network_allowed(&self) -> bool

fn lock_update_allowed(&self) -> bool