Module cargo::util::errors

Structs

CargoTestError
Error when testcases fail

CliError
The CLI error is the error type used at Cargo's CLI-layer.

HttpNot200

InternalError
An unexpected, internal error.

ManifestCauses
An iterator over the ManifestError chain of causes.

ManifestError
Error wrapper related to a particular manifest and providing it's path.

ProcessError

VerboseError
An error wrapper for errors that should only be displayed with --verbose.

Enums

Test

Traits

CargoResultExt

Functions

internal

is_simple_exit_code

process_error
Creates a new process error.

Type Definitions

CargoResult

CliResult