Module cargo::util::errors

Structs

AlreadyPrintedError

An error that does not need to be printed because it does not add any new information to what has already been printed.

CliError

The CLI error is the error type used at Cargo’s CLI-layer.

HttpNotSuccessful

InternalError

An unexpected, internal error.

ManifestCauses

An iterator over the ManifestError chain of causes.

ManifestError

Error wrapper related to a particular manifest and providing it’s path.

VerboseError

An error wrapper for errors that should only be displayed with --verbose.

Constants

DEBUG_HEADERS

These are headers that are included in error messages to help with diagnosing issues.

Functions

internal

Type Aliases

CargoResult

CliResult