cargo::util
Module lints

Structs

Lint

LintGroup

Enums

LintLevel

LintLevelReason

Constants

IM_A_TEAPOT 🔒

This lint is only to be used for testing purposes

LINTS

LINT_GROUPS 🔒

TEST_DUMMY_UNSTABLE 🔒

This lint group is only to be used for testing purposes

UNKNOWN_LINTS 🔒

Functions

analyze_cargo_lints_table

check_im_a_teapot

find_lint_or_group 🔒

get_span

level_priority 🔒

output_unknown_lints 🔒

rel_cwd_manifest_path

Gets the relative path to a manifest from the current working directory, or the absolute path of the manifest if a relative path cannot be constructed

verify_feature_enabled 🔒