Struct cargo::util::LazyCell

pub struct LazyCell<T> { /* fields omitted */ }

Methods

impl<T> LazyCell<T>

pub fn new() -> LazyCell<T>

Creates a new empty lazy cell.

pub fn fill(&self, t: T) -> Result<(), T>

Put a value into this cell.

This function will fail if the cell has already been filled.

pub fn borrow(&self) -> Option<&T>

Borrows the contents of this lazy cell for the duration of the cell itself.

This function will return Some if the cell has been previously initialized, and None if it has not yet been initialized.

pub fn borrow_mut(&mut self) -> Option<&mut T>

Same as borrow, but the mutable version

pub fn into_inner(self) -> Option<T>

Consumes this LazyCell, returning the underlying value.

pub fn get_or_try_init<Error, F>(&self, init: F) -> Result<&T, Error> where
F: FnOnce() -> Result<T, Error>,

Borrows the contents of this lazy cell, initializing it if necessary.