Struct cargo::util::Rustc

pub struct Rustc {
    pub path: PathBuf,
    pub wrapper: Option<PathBuf>,
    pub verbose_version: String,
    pub host: String,
}

Information on the rustc executable

Fields
path: PathBuf

The location of the exe

wrapper: Option<PathBuf>

An optional program that will be passed the path of the rust exe as its first argument, and rustc args following this.

verbose_version: String

Verbose version information (the output of rustc -vV)

host: String

The host triple (arch-platform-OS), this comes from verbose_version.

Methods

impl Rustc

pub fn new(path: PathBuf, wrapper: Option<PathBuf>) -> CargoResult<Rustc>

Run the compiler at path to learn various pieces of information about it, with an optional wrapper.

If successful this function returns a description of the compiler along with a list of its capabilities.

pub fn process(&self) -> ProcessBuilder

Get a process builder set up to use the found rustc version, with a wrapper if Some