Struct cargo::util::Rustc

pub struct Rustc {
    pub verbose_version: String,
    pub host: String,
    pub cap_lints: bool,
}

Fields

verbose_version

host

cap_lints

Methods

impl Rustc

fn new<P: AsRef<Path>>(path: P) -> CargoResult<Rustc>

Run the compiler at path to learn varioues pieces of information about it.

If successful this function returns a description of the compiler along with a list of its capabilities.

fn blank() -> Rustc