Module cargo::util::toml

Modules

embedded ğŸ”’

targets ğŸ”’

This module implements Cargo conventions for directory layout:

Macros

package_field_getter ğŸ”’

Defines simple getter methods for inheritable fields.

Structs

InheritableFields

A group of fields that are inheritable by members of the workspace

ManifestContext ğŸ”’

Constants

DEFAULT_README_FILES ğŸ”’

Traits

ResolveToPath

Functions

default_readme_from_package_root ğŸ”’

Checks if a file with any of the default README file names exists in the package root. If so, returns a String representing that name.

dep_to_dependency ğŸ”’

dependency_inherit_with ğŸ”’

deprecated_underscore ğŸ”’

Warn about paths that have been deprecated and may conflict.

deprecated_ws_default_features ğŸ”’

deserialize_toml ğŸ”’

detailed_dep_to_dependency ğŸ”’

emit_diagnostic ğŸ”’

field_inherit_with ğŸ”’

gather_dependencies ğŸ”’

inheritable_from_path ğŸ”’

inner_dependency_inherit_with ğŸ”’

is_embedded

See also bin/cargo/commands/run.rss is_manifest_command

lints_inherit_with ğŸ”’

lints_to_rustflags ğŸ”’

load_inheritable_fields ğŸ”’

normalize_path_sep ğŸ”’

normalize_path_string_sep ğŸ”’

parse_document ğŸ”’

patch ğŸ”’

prepare_for_publish

prepare_target_for_publish ğŸ”’

prepare_targets_for_publish ğŸ”’

prepare_toml_for_publish ğŸ”’

Prepares the manifest for publishing.

read_manifest

Loads a Cargo.toml from a file on disk.

read_toml_string ğŸ”’

replace ğŸ”’

resolve_dependencies ğŸ”’

resolve_package_readme ğŸ”’

Returns the name of the README file for a manifest::TomlPackage.

resolve_package_toml ğŸ”’

resolve_toml ğŸ”’

See Manifest::resolved_toml for more details

stringify ğŸ”’

to_dependency ğŸ”’

to_dependency_source_id ğŸ”’

to_real_manifest ğŸ”’

to_virtual_manifest ğŸ”’

to_workspace_config ğŸ”’

to_workspace_root_config ğŸ”’

unique_build_targets ğŸ”’

Checks a list of build targets, and ensures the target names are unique within a vector. If not, the name of the offending build target is returned.

unused_dep_keys ğŸ”’

validate_dependencies ğŸ”’

validate_profile

Checks stytax validity and unstable feature gate for a given profile.

validate_profile_layer ğŸ”’

Validates a profile.

validate_profile_override ğŸ”’

Validation that is specific to an override.

validate_profiles ğŸ”’

Checks syntax validity and unstable feature gate for each profile.

verify_lints ğŸ”’

warn_for_cargo_lint_feature ğŸ”’

warn_on_unused ğŸ”’