Function cargo::util::toml::unique_build_targets

fn unique_build_targets(
    targets: &[Target],
    package_root: &Path,
) -> Result<(), HashMap<PathBuf, Vec<Target>>>

Checks a list of build targets, and ensures the target names are unique within a vector. If not, the name of the offending build target is returned.