Module cargo::util::toml::targets

This module implements Cargo conventions for directory layout:

src/lib.rs is a library

src/main.rs is a binary

src/bin/*.rs are binaries

examples/*.rs are examples

tests/*.rs are integration tests

benches/*.rs are benchmarks

It is a bit tricky because we need match explicit information from Cargo.toml with implicit info in directory layout.

Constants

DEFAULT_BENCH_DIR_NAME ğŸ”’

DEFAULT_BIN_DIR_NAME ğŸ”’

DEFAULT_EXAMPLE_DIR_NAME ğŸ”’

DEFAULT_TEST_DIR_NAME ğŸ”’

Functions

clean_benches ğŸ”’

clean_bins ğŸ”’

clean_examples ğŸ”’

clean_lib ğŸ”’

clean_targets ğŸ”’

clean_targets_with_legacy_path ğŸ”’

clean_tests ğŸ”’

configure ğŸ”’

infer_any ğŸ”’

infer_file ğŸ”’

infer_from_directory ğŸ”’

infer_subdirectory ğŸ”’

inferred_bins ğŸ”’

inferred_lib ğŸ”’

inferred_to_toml_targets ğŸ”’

is_not_dotfile ğŸ”’

target_path ğŸ”’

target_path_not_found_error_message ğŸ”’

Build an error message for a target path that cannot be determined either by auto-discovery or specifying.

targets

toml_targets_and_inferred ğŸ”’

validate_target_name ğŸ”’

validate_unique_names ğŸ”’

Will check a list of toml targets, and make sure the target names are unique within a vector.