cargo::util::toml_mut::dependency
Struct Dependency

#[non_exhaustive]
pub struct Dependency {
    pub name: String,
    pub optional: Option<bool>,
    pub public: Option<bool>,
    pub features: Option<IndexSet<String>>,
    pub default_features: Option<bool>,
    pub inherited_features: Option<IndexSet<String>>,
    pub source: Option<Source>,
    pub registry: Option<String>,
    pub rename: Option<String>,
}

A dependency handled by Cargo.

None means the field will be blank in TOML.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

name: String

The name of the dependency (as it is set in its Cargo.toml and known to crates.io).

optional: Option<bool>

Whether the dependency is opted-in with a feature flag.

public: Option<bool>

Whether the dependency is marked as public.

features: Option<IndexSet<String>>

List of features to add (or None to keep features unchanged).

default_features: Option<bool>

Whether default features are enabled.

inherited_features: Option<IndexSet<String>>

List of features inherited from a workspace dependency.

source: Option<Source>

Where the dependency comes from.

registry: Option<String>

Non-default registry.

rename: Option<String>

If the dependency is renamed, this is the new name for the dependency as a string. None if it is not renamed.

Implementations

impl Dependency

pub fn new(name: &str) -> Self

Create a new dependency with a name.

pub fn set_source(self, source: impl Into<Source>) -> Self

Set dependency to a given version.

pub fn clear_version(self) -> Self

Remove the existing version requirement.

pub fn set_optional(self, opt: bool) -> Self

Set whether the dependency is optional.

pub fn set_features(self, features: IndexSet<String>) -> Self

Set features as an array of string (does some basic parsing).

pub fn extend_features(self, features: impl IntoIterator<Item = String>) -> Self

Set features as an array of string (does some basic parsing).

pub fn set_default_features(self, default_features: bool) -> Self

Set the value of default-features for the dependency.

pub fn set_rename(self, rename: &str) -> Self

Set the alias for the dependency.

pub fn set_registry(self, registry: impl Into<String>) -> Self

Set the value of registry for the dependency.

pub fn set_inherited_features(self, features: IndexSet<String>) -> Self

Set features as an array of string (does some basic parsing).

pub fn source(&self) -> Option<&Source>

Get the dependency source.

pub fn version(&self) -> Option<&str>

Get version of dependency.

pub fn registry(&self) -> Option<&str>

Get registry of the dependency.

pub fn rename(&self) -> Option<&str>

Get the alias for the dependency (if any).

pub fn default_features(&self) -> Option<bool>

Whether default features are activated.

pub fn optional(&self) -> Option<bool>

Get whether the dep is optional.

pub fn public(&self) -> Option<bool>

Get whether the dep is public.

pub fn source_id( &self, gctx: &GlobalContext, ) -> CargoResult<MaybeWorkspace<SourceId>>

Get the SourceID for this dependency.

pub fn query( &self, gctx: &GlobalContext, ) -> CargoResult<MaybeWorkspace<Dependency>>

Query to find this dependency.

impl Dependency

pub fn from_toml( gctx: &GlobalContext, workspace_root: &Path, crate_root: &Path, unstable_features: &Features, key: &str, item: &Item, ) -> CargoResult<Self>

Create a dependency from a TOML table entry.

pub fn toml_key(&self) -> &str

Get the dependency name as defined in the manifest, that is, either the alias (rename field if Some), or the official package name (name field).

pub fn to_toml<'a>( &self, gctx: &GlobalContext, workspace_root: &Path, crate_root: &Path, unstable_features: &Features, ) -> CargoResult<Item>

Convert dependency to TOML.

Returns a tuple with the dependency’s name and either the version as a String or the path/git repository as an InlineTable. (If the dependency is set as optional or default-features is set to false, an InlineTable is returned in any case.)

Panic

Panics if the path is relative

pub fn update_toml<'k, 'a>( &self, gctx: &GlobalContext, workspace_root: &Path, crate_root: &Path, unstable_features: &Features, key: &mut KeyMut<'k>, item: &mut Item, ) -> CargoResult<()>

Modify existing entry to match this dependency.